% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_aggregate_data.R
\name{aggregate_data}
\alias{aggregate_data}
\alias{mStat_aggregate_data}
\title{Aggregate Data Function}
\usage{
mStat_aggregate_data(data.obj, subject.var, strata.var = NULL)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{subject.var}{A character specifying the name of the subject variable.}

\item{strata.var}{A character specifying the name of the strata variable. If NULL, no stratification is performed.}
}
\value{
A new list object with aggregated data.
}
\description{
This function aggregates data by a specified subject and optional strata.
}
\examples{

# Prepare data for the function
data(peerj32.obj)

# Call the function
aggregated_data <- mStat_aggregate_data(
  data.obj = peerj32.obj,
  subject.var = "subject",
  strata.var = NULL
)

}
