% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linda2_optimized.R
\name{apply_tree_smoothing_fast}
\alias{apply_tree_smoothing_fast}
\title{Apply tree smoothing to test statistics (optimized version)}
\usage{
apply_tree_smoothing_fast(stat.baseline, tree.info)
}
\description{
Apply tree smoothing to test statistics (optimized version)
}
\keyword{internal}
