% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_import_biom_as_data_obj.R
\name{build_mStat_tax_table}
\alias{build_mStat_tax_table}
\title{Build mStat tax table}
\usage{
build_mStat_tax_table(taxlist)
}
\arguments{
\item{taxlist}{A list where each element is a named vector representing the taxonomy of a feature.}
}
\value{
A matrix representing the taxonomic table.
}
\description{
This function takes a list of taxonomy information and creates a taxonomic table compatible with mStat.
}
\examples{
# Assume `taxlist` is a list of taxonomy information
# tax_table <- build_mStat_tax_table(taxlist)
}
