% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecam.obj.R
\docType{data}
\name{ecam.obj}
\alias{ecam.obj}
\title{Data from the Early Childhood Antibiotics and the Microbiome (ECAM) study}
\format{
An object of class \code{list} of length 3.
}
\source{
Data source: https://github.com/FrederickHuangLin/ANCOM/tree/master/data
}
\usage{
ecam.obj
}
\description{
The ecam.obj dataset is derived from a subset of the Early Childhood
Antibiotics and the Microbiome (ECAM) study, which tracked the microbiome
composition and development of 43 infants in the United States from birth
to 2 years of age, identifying associations between the microbiome and
antibiotic exposure, delivery mode, and diet.
}
\details{
ecam.obj is a MicrobiomeStat Data Object.
}
\references{
Bokulich, Nicholas A., et al. "Antibiotics, birth mode, and diet shape
microbiome maturation during early life." Science translational medicine
8.343 (2016): 343ra82-343ra82.
}
\keyword{datasets}
