% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_beta_change_boxplot_pair.R
\name{generate_beta_change_boxplot_pair}
\alias{generate_beta_change_boxplot_pair}
\title{Generate box plot of beta diversity change}
\usage{
generate_beta_change_boxplot_pair(
  data.obj = NULL,
  dist.obj = NULL,
  subject.var,
  time.var,
  group.var = NULL,
  strata.var = NULL,
  change.base,
  dist.name = c("BC", "Jaccard"),
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{dist.obj}{A dissimilarity object created with the vegdist() function or NULL. If NULL, the function will automatically compute the beta diversity based on the input \code{data.obj}.}

\item{subject.var}{A string specifying the name of the subject variable}

\item{time.var}{A string specifying the name of the time variable}

\item{group.var}{A string specifying the name of the group variable or NULL (default)}

\item{strata.var}{A string specifying the name of the strata variable or NULL (default)}

\item{change.base}{A string or numeric value specifying the first time point used for computing the beta diversity change}

\item{dist.name}{A character vector specifying which beta diversity indices to calculate. Supported indices are "BC" (Bray-Curtis), "Jaccard", "UniFrac" (unweighted UniFrac), "GUniFrac" (generalized UniFrac), "WUniFrac" (weighted UniFrac), and "JS" (Jensen-Shannon divergence). If a name is provided but the corresponding object does not exist within dist.obj, it will be computed internally. If the specific index is not supported, an error message will be returned.}

\item{base.size}{(Optional) Base font size for the plot (default is 16).}

\item{theme.choice}{(Optional) Name of the theme for the plot. Default is "prism". Other options include "plain", "classic", and any other themes compatible with ggplot2.}

\item{custom.theme}{(Optional) A custom ggplot2 theme.}

\item{palette}{(Optional) A palette function or character vector with the colors for the plot.}

\item{pdf}{A logical value indicating whether to save the plot as a PDF file. Default is TRUE}

\item{file.ann}{A character string specifying a custom annotation for the PDF file name or NULL (default)}

\item{pdf.wid}{(Optional) The width of the PDF file if `pdf` is set to `TRUE` (default is 11).}

\item{pdf.hei}{(Optional) The height of the PDF file if `pdf` is set to `TRUE` (default is 8.5).}

\item{...}{Additional parameters passed on to ggsave()}
}
\value{
A ggplot object with the box plot of beta diversity change.

A list of ggplot objects with the box plot of beta diversity change for each diversity index specified.
}
\description{
This function generates a box plot that visualizes the change in beta diversity
between two time points for different groups and strata. It takes a \code{dist.obj},
and can also handle a matrix with coordinates for different groups and strata.
Optionally, a user can specify whether to save the plot as a PDF file.
}
\examples{
# Load required libraries and example data
library(tidyverse)
library(vegan)
data(peerj32.obj)

# Generate the boxplot pair
generate_beta_change_boxplot_pair(
  data.obj = peerj32.obj,
  dist.obj = NULL,
  subject.var = "subject",
  time.var = "time",
  group.var = "group",
  strata.var = NULL,
  change.base = "1",
  dist.name = c('BC'),
  base.size = 20,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)

}
\seealso{
\code{\link[MicrobiomeStat]{mStat_calculate_beta_diversity}} for creating the distance object. This package was created by Caffery Yang. For more information, you can contact them at cafferychen7850@gmail.com.
}
