% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_beta_ordination_long.R
\name{generate_beta_ordination_long}
\alias{generate_beta_ordination_long}
\title{Generate Beta Ordination Plots}
\usage{
generate_beta_ordination_long(
  data.obj = NULL,
  dist.obj = NULL,
  pc.obj = NULL,
  subject.var,
  time.var,
  t0.level,
  ts.levels,
  group.var = NULL,
  strata.var = NULL,
  adj.vars = NULL,
  dist.name = c("BC", "Jaccard"),
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{a list with the data to be used for the plots. If NULL, the distance object will be used instead.}

\item{dist.obj}{a list with distance matrices. If NULL, it will be calculated from the data object.}

\item{pc.obj}{a list with principal coordinates. If NULL, it will be calculated from the distance object.}

\item{subject.var}{a character string specifying the subject variable.}

\item{time.var}{a character string specifying the time variable.}

\item{t0.level}{a character string specifying the baseline time level.}

\item{ts.levels}{a vector of character strings specifying the time series levels.}

\item{group.var}{a character string specifying the grouping variable. If NULL, only color will be used in the plots.}

\item{strata.var}{a character string specifying the stratification variable. If NULL, no stratification will be done.}

\item{adj.vars}{A character vector containing the names of the columns in data.obj$meta.dat to include as covariates in the PERMANOVA analysis. If no covariates are needed, use NULL (default).}

\item{dist.name}{a character vector specifying the distance measures to use. Defaults to c('BC', 'Jaccard').}

\item{base.size}{a numeric value specifying the base size for the plot text.}

\item{theme.choice}{a character string specifying the theme to use. Defaults to "prism".}

\item{custom.theme}{a ggplot2 theme object to use instead of the default theme.}

\item{palette}{a character vector specifying the colors to use for the plots. If NULL, a default palette will be used.}

\item{pdf}{a logical value indicating whether to save the plots as PDF files. Defaults to TRUE.}

\item{file.ann}{a character string specifying an annotation to add to the file names of the saved plots.}

\item{pdf.wid}{a numeric value specifying the width of the saved PDF files.}

\item{pdf.hei}{a numeric value specifying the height of the saved PDF files.}

\item{...}{further arguments to be passed to the underlying functions.}
}
\value{
A list of ggplot2 objects representing the beta ordination plots.
}
\description{
This function generates beta ordination plots using the specified distance measure.
The plots can be stratified by a given variable and the color and shape aesthetics
can be mapped to groups and time variables, respectively. The function also supports
saving the plots as PDF files.
}
\examples{
\dontrun{
data(subset_T2D.obj)
generate_beta_ordination_long(
  data.obj = subset_T2D.obj,
  dist.obj = NULL,
  pc.obj = NULL,
  subject.var = "subject_id",
  time.var = "visit_number",
  t0.level = sort(unique(subset_T2D.obj$meta.dat$visit_number))[1],
  ts.levels = sort(unique(subset_T2D.obj$meta.dat$visit_number))[2:6],
  group.var = "subject_gender",
  strata.var = "subject_race",
  adj.vars = "sample_body_site",
  dist.name = c("BC","Jaccard"),
  base.size = 16,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)
}
}
\seealso{
\code{\link[MicrobiomeStat]{mStat_calculate_beta_diversity}} for creating the distance object, \code{\link[MicrobiomeStat]{mStat_calculate_PC}} for computing the principal coordinates, and \code{\link[ggplot2]{geom_point}}, \code{\link[ggplot2]{geom_boxplot}} for the underlying plot functions used, and \code{\link[MicrobiomeStat]{mStat_convert_DGEList_to_data_obj}}, \code{\link[MicrobiomeStat]{mStat_convert_DESeqDataSet_to_data_obj}}, \code{\link[MicrobiomeStat]{mStat_convert_phyloseq_to_data_obj}}, \code{\link[MicrobiomeStat]{mStat_convert_SummarizedExperiment_to_data_obj}}, \code{\link[MicrobiomeStat]{mStat_import_qiime2_as_data_obj}}, \code{\link[MicrobiomeStat]{mStat_import_mothur_as_data_obj}}, \code{\link[MicrobiomeStat]{mStat_import_dada2_as_data_obj}}, \code{\link[MicrobiomeStat]{mStat_import_biom_as_data_obj}} for data conversion.
}
\author{
Caffery Yang \email{cafferychen7850@gmail.com}
}
