% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_beta_ordination_pair.R
\name{generate_beta_ordination_pair}
\alias{generate_beta_ordination_pair}
\title{Generate beta ordination plots for paired samples}
\usage{
generate_beta_ordination_pair(
  data.obj = NULL,
  dist.obj = NULL,
  pc.obj = NULL,
  subject.var,
  time.var,
  group.var = NULL,
  strata.var = NULL,
  adj.vars = NULL,
  dist.name = c("BC", "Jaccard"),
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{dist.obj}{Distance matrix between samples, usually calculated using
\code{\link[MicrobiomeStat]{mStat_calculate_beta_diversity}} function.
If NULL, beta diversity will be automatically computed from \code{data.obj}
using \code{mStat_calculate_beta_diversity}.}

\item{pc.obj}{A list containing the results of dimension reduction/Principal Component Analysis.
This should be the output from functions like \code{\link[MicrobiomeStat]{mStat_calculate_PC}}, containing the PC coordinates and other metadata.
If NULL (default), dimension reduction will be automatically performed using metric multidimensional scaling (MDS) via \code{\link[MicrobiomeStat]{mStat_calculate_PC}}.
The pc.obj list structure should contain:
\itemize{
 \item{$points:}{A matrix with samples as rows and PCs as columns containing the coordinates.}
 \item{$eig:}{Eigenvalues for each PC dimension.}
 \item{$vectors:}{Loadings vectors for features onto each PC.}
 \item{Other metadata like $method, $dist.name, etc.}
}
See \code{\link[MicrobiomeStat]{mStat_calculate_PC}} function for details on output format.}

\item{subject.var}{The variable in the metadata table that represents the subject.}

\item{time.var}{The variable in the metadata table that represents the time.}

\item{group.var}{(Optional) The variable in the metadata table that represents the grouping factor.}

\item{strata.var}{(Optional) The variable in the metadata table that represents the stratification factor.}

\item{adj.vars}{A character vector containing the names of the columns in data.obj$meta.dat to include as covariates in the PERMANOVA analysis. If no covariates are needed, use NULL (default).}

\item{dist.name}{A character vector specifying which beta diversity indices to calculate. Supported indices are "BC" (Bray-Curtis), "Jaccard", "UniFrac" (unweighted UniFrac), "GUniFrac" (generalized UniFrac), "WUniFrac" (weighted UniFrac), and "JS" (Jensen-Shannon divergence). If a name is provided but the corresponding object does not exist within dist.obj, it will be computed internally. If the specific index is not supported, an error message will be returned. Default is c('BC', 'Jaccard').}

\item{base.size}{(Optional) Base font size for the plot (default is 16).}

\item{theme.choice}{Plot theme choice. Can be one of:
  - "prism": ggprism::theme_prism()
  - "classic": theme_classic()
  - "gray": theme_gray()
  - "bw": theme_bw()
Default is "bw".}

\item{custom.theme}{A custom ggplot theme provided as a ggplot2 theme object. This allows users to override the default theme and provide their own theme for plotting. To use a custom theme, first create a theme object with ggplot2::theme(), then pass it to this argument. For example:

```r
my_theme <- ggplot2::theme(
  axis.title = ggplot2::element_text(size=16, color="red"),
  legend.position = "none"
)
```

Then pass `my_theme` to `custom.theme`. Default is NULL, which will use the default theme based on `theme.choice`.}

\item{palette}{(Optional) A palette function or character vector with the colors for the plot.}

\item{pdf}{(Optional) A boolean indicating whether to save the output as a PDF file (default is TRUE).}

\item{file.ann}{(Optional) A string for annotating the output file name.}

\item{pdf.wid}{(Optional) The width of the PDF file if `pdf` is set to `TRUE` (default is 11).}

\item{pdf.hei}{(Optional) The height of the PDF file if `pdf` is set to `TRUE` (default is 8.5).}

\item{...}{(Optional) Additional arguments to pass to the plotting function.}
}
\value{
A PCoA plot displaying the beta diversity ordination, stratified by the specified grouping and/or strata variables (if provided). The plot will be saved as a PDF if `pdf` is set to `TRUE`.
}
\description{
This function, part of the MicrobiomeStat package, generates Principle Coordinate Analysis (PCoA) plots using the provided distance object. It is specifically designed for the analysis of microbiome data. This function is tailored for paired samples, such as those from longitudinal studies or experiments with multiple time points.
}
\details{
The function is flexible and allows for various modifications, including the choice of distance measure and stratification factor, providing a comprehensive tool for microbiome beta diversity exploration. It integrates well with other MicrobiomeStat functions and takes their output as input.
}
\examples{
\dontrun{

# Load necessary libraries and data
library(vegan)
data(peerj32.obj)

# Perform beta ordination pair analysis using `generate_beta_ordination_pair`
generate_beta_ordination_pair(
  data.obj      = peerj32.obj,
  dist.obj      = NULL,
  pc.obj        = NULL,
  subject.var   = "subject",
  time.var      = "time",
  group.var     = "group",
  strata.var    = NULL,
  adj.vars      = NULL,
  dist.name     = c("BC"),
  base.size     = 16,
  theme.choice  = "bw",
  custom.theme  = NULL,
  palette       = NULL,
  pdf           = TRUE,
  file.ann      = NULL,
  pdf.wid       = 11,
  pdf.hei       = 8.5
)

}
}
\seealso{
\code{\link[MicrobiomeStat]{mStat_calculate_beta_diversity}} for creating the distance object, \code{\link[MicrobiomeStat]{mStat_calculate_PC}} for computing the principal coordinates, and \code{\link[ggplot2]{geom_point}}, \code{\link[ggplot2]{geom_boxplot}} for the underlying plot functions used, and \code{\link[MicrobiomeStat]{mStat_convert_DGEList_to_data_obj}}, \code{\link[MicrobiomeStat]{mStat_convert_DESeqDataSet_to_data_obj}}, \code{\link[MicrobiomeStat]{mStat_convert_phyloseq_to_data_obj}}, \code{\link[MicrobiomeStat]{mStat_convert_SummarizedExperiment_to_data_obj}}, \code{\link[MicrobiomeStat]{mStat_import_qiime2_as_data_obj}}, \code{\link[MicrobiomeStat]{mStat_import_mothur_as_data_obj}}, \code{\link[MicrobiomeStat]{mStat_import_dada2_as_data_obj}}, \code{\link[MicrobiomeStat]{mStat_import_biom_as_data_obj}} for data conversion.
}
