% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_beta_pc_boxplot_long.R
\name{generate_beta_pc_boxplot_long}
\alias{generate_beta_pc_boxplot_long}
\title{Generate beta ordination plots for paired samples}
\usage{
generate_beta_pc_boxplot_long(
  data.obj = NULL,
  dist.obj = NULL,
  pc.obj = NULL,
  pc.ind = c(1, 2),
  subject.var,
  time.var,
  t0.level = NULL,
  ts.levels = NULL,
  group.var = NULL,
  strata.var = NULL,
  dist.name = c("BC"),
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{dist.obj}{A distance object generated from distance matrices.}

\item{pc.ind}{(Optional) A numeric vector indicating which PCoA axes to plot (default is c(1, 2)).}

\item{subject.var}{The variable in the meta_tab table that represents the subject.}

\item{time.var}{The variable in the metadata table that represents the time.}

\item{group.var}{The variable in the metadata table that represents the grouping factor.}

\item{strata.var}{(Optional) The variable in the metadata table that represents the stratification factor.}

\item{dist.name}{A character vector specifying which beta diversity indices to calculate. Supported indices are "BC" (Bray-Curtis), "Jaccard", "UniFrac" (unweighted UniFrac), "GUniFrac" (generalized UniFrac), "WUniFrac" (weighted UniFrac), and "JS" (Jensen-Shannon divergence). If a name is provided but the corresponding object does not exist within dist.obj, it will be computed internally. If the specific index is not supported, an error message will be returned.}

\item{pdf}{(Optional) A boolean indicating whether to save the output as a PDF file (default is TRUE).}

\item{file.ann}{(Optional) A string for annotating the output file name.}

\item{...}{(Optional) Additional arguments to pass to the plotting function.}

\item{pc.mat}{(Optional) A matrix containing the principal coordinates. If not provided, the function will calculate PCoA based on the given distance object.}
}
\value{
A PCoA plot displaying the beta diversity ordination, stratified by the specified grouping and/or strata variables (if provided). The plot will be saved as a PDF if `pdf` is set to `TRUE`.
}
\description{
This function generates Principle Coordinate Analysis (PCoA) plots using the provided distance object,
optionally stratified by group and/or strata variables. The function is designed for paired samples,
such as those from longitudinal studies or experiments with multiple time points.
}
\examples{
# Load required libraries and data
library(microbiome)
library(vegan)
data(peerj32)
peerj32.obj <- list()
peerj32.phy <- peerj32$phyloseq
peerj32.obj <- mStat_convert_phyloseq_to_data_obj(peerj32.phy)
# Calculate Bray-Curtis distance
dist.obj <- mStat_calculate_beta_diversity(peerj32.obj,"BC")
# Add metadata information
attr(dist.obj[["BC"]], "labels") <- ecam.obj$meta_tab
  generate_beta_pc_boxplot_long(
  data.obj = ecam.obj,
  dist.obj = NULL,
  pc.obj = NULL,
  subject.var = "studyid",
  time.var = "month",
  t0.level = "0",
  ts.levels = as.character(sort(as.numeric(unique(ecam.obj$meta.dat$month))))[2],
  group.var = "diet",
  strata.var = "delivery",
  dist.name = c('BC'),
  base.size = 20,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = c(ggsci::pal_npg()(9),ggsci::pal_jama()(7),ggsci::pal_lancet()(9)),
  pdf = TRUE,
  file.ann = "test",
  pdf.wid = 11,
  pdf.hei = 8.5
)
  generate_beta_pc_boxplot_long(
  data.obj = peerj32.obj,
  dist.obj = NULL,
  pc.obj = NULL,
  subject.var = "subject",
  time.var = "time",
  t0.level = "1",
  ts.levels = "2",
  group.var = "group",
  strata.var = NULL,
  dist.name = c('BC'),
  base.size = 20,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)
}
