% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_beta_pc_change_boxplot_pair.R
\name{generate_beta_pc_change_boxplot_pair}
\alias{generate_beta_pc_change_boxplot_pair}
\title{Generate boxplot pair of changes in principal coordinates for beta diversity}
\usage{
generate_beta_pc_change_boxplot_pair(
  data.obj = NULL,
  dist.obj = NULL,
  pc.obj = NULL,
  pc.ind = c(1, 2),
  subject.var,
  time.var,
  group.var = NULL,
  strata.var = NULL,
  change.base = NULL,
  change.func = "difference",
  dist.name = c("BC", "Jaccard"),
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{dist.obj}{A list containing distance matrices (Bray-Curtis, Generalized UniFrac, etc.).}

\item{pc.ind}{A numeric vector of the principal coordinate indices to be plotted. Default is c(1, 2).}

\item{subject.var}{A character string specifying the subject variable in the metadata.}

\item{time.var}{A character string specifying the time variable in the metadata.}

\item{group.var}{A character string specifying the group variable in the metadata.}

\item{strata.var}{A character string specifying the strata variable in the metadata. Default is NULL.}

\item{change.base}{A numeric value specifying the baseline time point for calculating changes.}

\item{dist.name}{A character vector specifying which beta diversity indices to calculate. Supported indices are "BC" (Bray-Curtis), "Jaccard", "UniFrac" (unweighted UniFrac), "GUniFrac" (generalized UniFrac), "WUniFrac" (weighted UniFrac), and "JS" (Jensen-Shannon divergence). If a name is provided but the corresponding object does not exist within dist.obj, it will be computed internally. If the specific index is not supported, an error message will be returned.}

\item{pdf}{A logical value indicating whether to save the plot as a PDF file. Default is TRUE.}

\item{file.ann}{An optional character string to be used as a file annotation when saving the PDF. Default is NULL.}

\item{...}{Additional arguments passed to ggplot2 functions.}

\item{pc.mat}{An optional matrix of principal coordinates. Default is NULL, in which case the function will calculate PCoA.}
}
\value{
A ggplot2 boxplot of the changes in principal coordinates for beta diversity between two time points.
}
\description{
This function creates a boxplot pair of the changes in principal coordinates for beta diversity
between two time points. It takes a list of distance matrices and principal coordinate (PC)
indices as input, along with metadata variables, and outputs a ggplot2 boxplot.
}
\examples{
# Load required libraries and example data
library(microbiome)
library(tidyverse)
library(vegan)
library(ggh4x)
data(peerj32)
peerj32_obj <- list()
peerj32_phy <- peerj32$phyloseq
peerj32.obj <- mStat_convert_phyloseq_to_data_obj(peerj32.phy)
dist.obj <- mStat_calculate_beta_diversity(peerj32.obj, dist.name = c('BC'))
# Add metadata information
attr(dist.obj[["BC"]], "labels") <- peerj32_obj$meta.dat

# Generate the boxplot pair
generate_beta_pc_change_boxplot_pair(
  data.obj = peerj32.obj,
  dist.obj = NULL,
  pc.obj = NULL,
  pc.ind = c(1, 2),
  subject.var = "subject",
  time.var = "time",
  group.var = "group",
  strata.var = NULL,
  change.base = "1",
  change.func = "difference",
  dist.name = c('BC'),
  base.size = 20,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)

}
