% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_beta_pc_trend_test_long.R
\name{generate_beta_pc_trend_test_long}
\alias{generate_beta_pc_trend_test_long}
\title{Trend Test on Principal Coordinates of Beta Diversity Metrics Over Time}
\usage{
generate_beta_pc_trend_test_long(
  data.obj = NULL,
  dist.obj = NULL,
  pc.obj = NULL,
  pc.ind = c(1, 2),
  subject.var,
  time.var,
  t0.level = NULL,
  ts.levels = NULL,
  group.var = NULL,
  adj.vars = NULL,
  dist.name = c("BC"),
  ...
)
}
\arguments{
\item{data.obj}{A list object containing all data.}

\item{dist.obj}{List object containing distance matrices. If NULL, beta diversity is calculated from data.obj.}

\item{pc.obj}{List object containing Principal Coordinates matrices. If NULL, PCoA is performed by default.}

\item{pc.ind}{Numeric vector indicating which Principal Coordinates to include in the trend test.}

\item{subject.var}{Character string indicating the variable for subject identifiers.}

\item{time.var}{Character string indicating the variable for time points. Must be coded as numeric.}

\item{t0.level}{Character string indicating the baseline time point.}

\item{ts.levels}{Vector of character strings indicating the time points to include in the trend test.}

\item{group.var}{Character string indicating the variable for group identifiers.}

\item{adj.vars}{Character string indicating the variables for adjustments in the distance calculations.}

\item{dist.name}{Vector of character strings indicating the distance measures to include in the trend test, such as "BC" for Bray-Curtis.}

\item{...}{Additional arguments to be passed to the mixed-effects model function.}
}
\value{
A list of results for each distance measure and selected Principal Coordinate, including coefficients from the mixed-effects models.
}
\description{
This function performs a trend test on the selected Principal Coordinates (PCs) of beta diversity distances over time for different groups.
Various methods of dimension reduction, such as Principal Coordinates Analysis (PCoA), non-metric multidimensional scaling (NMDS), t-SNE, or UMAP, can be used.
The default method for dimension reduction is PCoA.
The function relies on a numeric time variable and can make adjustments for distance calculations.
}
\examples{
\dontrun{
library(vegan)

  data(ecam.obj)
  generate_beta_pc_trend_test_long(
  data.obj = ecam.obj,
  dist.obj = NULL,
  pc.obj = NULL,
  pc.ind = c(1, 2),
  subject.var = "studyid",
  time.var = "month",
  t0.level = NULL,
  ts.levels = NULL,
  group.var = "diet",
  adj.vars = "delivery",
  dist.name = c('BC'),
)
}
}
