% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_beta_test_long.R
\name{generate_beta_test_long}
\alias{generate_beta_test_long}
\title{Perform longitudinal beta diversity tests using PERMANOVA}
\usage{
generate_beta_test_long(
  data.obj,
  dist.obj = NULL,
  time.var,
  t0.level,
  ts.levels,
  subject.var,
  group.var,
  adj.vars = NULL,
  dist.name = c("BC", "Jaccard", "UniFrac", "GUniFrac", "WUniFrac", "JS")
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, containing microbiome data and metadata.}

\item{dist.obj}{A distance object, if NULL it will be calculated based on data.obj and dist.name. This distance object should be a dissimilarity matrix or an object that can be coerced to a dissimilarity matrix.}

\item{time.var}{The name of the column in data.obj$meta.dat that contains the time variable for subsetting data, if applicable. This should be a character string representing the column name.}

\item{t0.level}{A character or numeric value indicating the baseline level of the time variable for subsetting data.}

\item{ts.levels}{A character or numeric vector indicating the subsequent time points of interest for subsetting data.}

\item{subject.var}{The name of the column in data.obj$meta.dat that contains the subject variable for the PERMANOVA tests. This should be a character string representing the column name.}

\item{group.var}{The name of the column in data.obj$meta.dat that contains the grouping variable for the PERMANOVA tests. This should be a character string representing the column name.}

\item{adj.vars}{A character vector containing the names of the columns in data.obj$meta.dat to include as covariates in the PERMANOVA analysis. If no covariates are needed, use NULL (default).}

\item{dist.name}{A character vector specifying which beta diversity indices to calculate and test (e.g., 'BC', 'Jaccard', 'UniFrac', 'GUniFrac', 'WUniFrac', 'JS'). Default is c('BC', 'Jaccard', 'UniFrac', 'GUniFrac', 'WUniFrac', 'JS').}
}
\value{
A list containing the PERMANOVA results for each beta diversity index. The list includes two elements: "p.tab" - a table of p-values for the PERMANOVA tests across all indices, and "aov.tab" - a table containing detailed PERMANOVA results for each index. The p.tab and aov.tab tables include columns for the terms in the PERMANOVA model, the degrees of freedom, sums of squares, mean squares, F statistics, R-squared values, and p-values.
}
\description{
This function, part of the MicrobiomeStat package, calculates various beta diversity indices for a microbiome dataset. It then performs longitudinal PERMANOVA tests on these indices to assess the effect of a grouping variable and any additional covariates on the beta diversity. The function returns the PERMANOVA results for each of the specified beta diversity indices.
}
\examples{
data(subset_T2D.obj)

# Perform pairwise beta diversity tests using PERMANOVA
beta_test_pair_results <- generate_beta_test_long(
  data.obj = subset_T2D.obj,
  dist.obj = NULL,
  time.var = "visit_number",
  t0.level = sort(unique(subset_T2D.obj$meta.dat$visit_number))[1],
  ts.levels = sort(unique(subset_T2D.obj$meta.dat$visit_number))[2:6],
  subject.var = "subject_id",
  group.var = "subject_race",
  adj.vars = "subject_gender",
  dist.name = c('BC', 'Jaccard')
)


}
\seealso{
\code{\link[GUniFrac]{PermanovaG2}}
\code{\link[microbiome]{peerj32}}
\code{\link[vegan]{adonis2}}
}
