% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_beta_test_pair.R
\name{generate_beta_test_pair}
\alias{generate_beta_test_pair}
\title{Perform pairwise beta diversity tests using PERMANOVA}
\usage{
generate_beta_test_pair(
  data.obj,
  dist.obj = NULL,
  time.var,
  subject.var,
  group.var,
  adj.vars = NULL,
  dist.name = c("BC", "Jaccard", "UniFrac", "GUniFrac", "WUniFrac", "JS")
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{dist.obj}{Distance matrix between samples, usually calculated using
\code{\link[MicrobiomeStat]{mStat_calculate_beta_diversity}} function.
If NULL, beta diversity will be automatically computed from \code{data.obj}
using \code{mStat_calculate_beta_diversity}.}

\item{time.var}{Character string specifying the column name in metadata containing a time variable.
Used for subsetting data before beta diversity calculation if needed. Optional.}

\item{subject.var}{Character string specifying the column name in metadata containing unique subject
IDs. Used as a stratifying variable in PERMANOVA tests. Required.}

\item{group.var}{Character string specifying the column name in metadata containing grouping
categories. Used as the main predictor in PERMANOVA tests. Required.}

\item{adj.vars}{Character vector specifying column names in metadata containing covariates
to include in PERMANOVA model. Optional, can be NULL if no adjustment needed.}

\item{dist.name}{A character vector specifying which beta diversity indices to calculate. Supported indices are "BC" (Bray-Curtis), "Jaccard", "UniFrac" (unweighted UniFrac), "GUniFrac" (generalized UniFrac), "WUniFrac" (weighted UniFrac), and "JS" (Jensen-Shannon divergence). If a name is provided but the corresponding object does not exist within dist.obj, it will be computed internally. If the specific index is not supported, an error message will be returned. Default is c('BC', 'Jaccard').}
}
\value{
A list containing the PERMANOVA results for each beta diversity index. The list includes two elements: "p.tab" - a table of p-values for the PERMANOVA tests dplyr::across all indices, and "aov.tab" - a table containing detailed PERMANOVA results for each index. The p.tab and aov.tab tables include columns for the terms in the PERMANOVA model, the degrees of freedom, sums of squares, mean squares, F statistics, R-squared values, and p-values.
}
\description{
This function calculates various beta diversity indices for a microbiome dataset and performs pairwise PERMANOVA tests on them to assess the effect of a grouping variable and any additional covariates on the beta diversity. It returns the PERMANOVA results for each of the specified beta diversity indices.
}
\examples{
\dontrun{
library(vegan)
data(peerj32.obj)
dist.obj <- mStat_calculate_beta_diversity(peerj32.obj, dist.name = c('BC', 'Jaccard'))

# Perform pairwise beta diversity tests using PERMANOVA
generate_beta_test_pair(
  data.obj = peerj32.obj,
  dist.obj = NULL,
  time.var = "time",
  subject.var = "subject",
  group.var = "group",
  adj.vars = c("sex"),
  dist.name = c('BC', 'Jaccard')
)
}

}
\seealso{
\code{\link[GUniFrac]{PermanovaG2}}
\code{\link[vegan]{adonis2}}
}
