% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_beta_test_single.R
\name{generate_beta_test_single}
\alias{generate_beta_test_single}
\title{Perform beta diversity tests using PERMANOVA}
\usage{
generate_beta_test_single(
  data.obj,
  dist.obj = NULL,
  time.var = NULL,
  t.level = NULL,
  group.var,
  adj.vars = NULL,
  dist.name = c("BC", "Jaccard", "UniFrac", "GUniFrac", "WUniFrac", "JS")
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, containing a 'meta.dat' component with metadata and a 'feature.tab' component with taxa in rows and samples in columns.}

\item{dist.obj}{A distance object, usually a dissimilarity matrix computed based on the microbiome data in 'data.obj'. If NULL, it will be calculated using 'dist.name'.}

\item{time.var}{The name of the column in the 'meta.dat' component of 'data.obj' that contains the time variable for subsetting data. Should be NULL if no time variable is needed.}

\item{t.level}{The specific level of the time variable to use when subsetting data. Should be NULL if no time variable is used.}

\item{group.var}{The name of the column in the 'meta.dat' component of 'data.obj' that contains the grouping variable for the PERMANOVA tests.}

\item{adj.vars}{A character vector with the names of columns in the 'meta.dat' component of 'data.obj' that are used as covariates in the PERMANOVA tests. Should be NULL if no covariates are used.}

\item{dist.name}{A character vector indicating the types of beta diversity indices to calculate and test. Possible values are: 'BC', 'Jaccard', 'UniFrac', 'GUniFrac', 'WUniFrac', 'JS'.}
}
\value{
A list containing the PERMANOVA results for each beta diversity index and an omnibus p-value. The list includes two elements: "p.tab" - a table of p-values for the PERMANOVA tests dplyr::across all indices, and "aov.tab" - a table containing detailed PERMANOVA results for each index. The p.tab and aov.tab tables include columns for the terms in the PERMANOVA model, the degrees of freedom, sums of squares, mean squares, F statistics, R-squared values, and p-values.
}
\description{
This function calculates various beta diversity indices for a microbiome dataset and performs PERMANOVA tests on them to assess the effect of a grouping variable and any additional covariates on the beta diversity. It returns the PERMANOVA results for each of the specified beta diversity indices.
}
\examples{
\dontrun{
library(vegan)
library(GUniFrac)
library(ape)
library(philentropy)
library(MicrobiomeStat)
data(peerj32.obj)

dist.obj <- mStat_calculate_beta_diversity(peerj32.obj, dist.name = c('BC', 'Jaccard'))

# Perform beta diversity tests using PERMANOVA
generate_beta_test_single(
  data.obj = peerj32.obj,
  dist.obj = NULL,
  time.var = "time",
  t.level = "1",
  group.var = "group",
  adj.vars = c("sex"),
  dist.name = c('BC', 'Jaccard')
)
}

}
