% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_beta_trend_test_long.R
\name{generate_beta_trend_test_long}
\alias{generate_beta_trend_test_long}
\title{Generate Beta Diversity Trend Test for Longitudinal Data}
\usage{
generate_beta_trend_test_long(
  data.obj,
  dist.obj = NULL,
  subject.var,
  time.var,
  group.var = NULL,
  adj.vars = NULL,
  dist.name = c("BC"),
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{dist.obj}{Distance matrix between samples, usually calculated using
\code{\link[MicrobiomeStat]{mStat_calculate_beta_diversity}} function.
If NULL, beta diversity will be automatically computed from \code{data.obj}
using \code{mStat_calculate_beta_diversity}.}

\item{subject.var}{The variable in the metadata table that represents the subject.}

\item{time.var}{The variable in the metadata table that represents the time. Ensure that it is coded as numeric.}

\item{group.var}{(Optional) The variable in the metadata table that represents the grouping factor.}

\item{adj.vars}{(Optional) Variables in the metadata table used for adjustment.}

\item{dist.name}{A character vector specifying which beta diversity indices to calculate. Default is "BC" (Bray-Curtis). Supported indices include "BC", "Jaccard", "UniFrac", "GUniFrac", "WUniFrac", and "JS".}

\item{...}{(Optional) Additional arguments to pass to internal functions.}
}
\value{
A list containing the result of the trend test for each specified beta diversity index. The result includes a tibble with the coefficients extracted from the mixed-effects model fitted for each distance.
}
\description{
The function `generate_beta_trend_test_long` computes a statistical trend test for beta diversity in longitudinal data. It considers subject, time, optional grouping and adjusting variables, and provides the trend in beta diversity over time for different groups.
}
\details{
The function starts by validating the input data, followed by processing the time variable and calculating the beta diversity if necessary. Adjustments are made based on the provided adjusting variables, and the mixed-effects model is fitted to the long-format data. The coefficients of the model are extracted and returned for each beta diversity index specified.
}
\note{
A warning message will be displayed to ensure that the time variable is coded as numeric. Non-numeric coding may lead to issues in the trend test computation.
}
\examples{
\dontrun{
data(ecam.obj)
generate_beta_trend_test_long(
  data.obj = ecam.obj,
  dist.obj = NULL,
  subject.var = "studyid",
  time.var = "month",
  group.var = "diet",
  adj.vars = c("antiexposedall","delivery"),
  dist.name = c("BC", "Jaccard")
)
}
}
