% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_areaplot_long.R
\name{generate_taxa_areaplot_long}
\alias{generate_taxa_areaplot_long}
\title{Generate taxa area plots over time}
\usage{
generate_taxa_areaplot_long(
  data.obj,
  subject.var,
  time.var,
  group.var = NULL,
  strata.var = NULL,
  feature.level = "original",
  feature.dat.type = c("count", "proportion", "other"),
  feature.number = 20,
  t0.level = NULL,
  ts.levels = NULL,
  base.size = 10,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A data object containing the taxa count data.}

\item{subject.var}{Name of the subject variable.}

\item{time.var}{Name of the time variable.}

\item{group.var}{Optional, name of the group variable. Default is NULL.}

\item{strata.var}{Optional, name of the stratification variable. Default is NULL.}

\item{feature.level}{The taxonomic level to plot. Default is "original".}

\item{feature.dat.type}{The type of features data; can be "count", "proportion", or "other". Default is "count".}

\item{feature.number}{The number of features to plot. Default is 20.}

\item{t0.level}{The initial time level. Default is NULL.}

\item{ts.levels}{The time series levels. Default is NULL.}

\item{base.size}{The base size for the ggplot2 theme. Default is 10.}

\item{theme.choice}{The theme choice for the ggplot2 theme. Default is "bw".}

\item{custom.theme}{Optional, a custom ggplot2 theme. Default is NULL.}

\item{palette}{Optional, a palette to use for the plot. Default is NULL.}

\item{pdf}{Logical indicating if the plot should be saved as a PDF. Default is TRUE.}

\item{file.ann}{Optional, a file annotation. Default is NULL.}

\item{pdf.wid}{Width of the output PDF. Default is 11.}

\item{pdf.hei}{Height of the output PDF. Default is 8.5.}

\item{...}{Additional arguments to pass to the function.}
}
\value{
A list of ggplot objects, each representing a taxa area plot for the specified feature level.
}
\description{
This function generates taxa area plots for a given data object. The plots will show the relative abundance of
different taxa over time. Raw count data will be automatically normalized using rarefaction and total sum scaling (TSS).
The function also supports the generation of plots for grouped data and stratified data.
}
\examples{
\dontrun{
library(ggh4x)
library(vegan)
data(ecam.obj)
generate_taxa_areaplot_long(
  data.obj = ecam.obj,
  subject.var = "studyid",
  time.var = "month",
  group.var = "diet",
  strata.var = "antiexposedall",
  feature.level = c("Family","Genus","Order","Family"),
  feature.dat.type = "proportion",
  feature.number = 20,
  t0.level = unique(ecam.obj$meta.dat$month)[1],
  ts.levels = unique(ecam.obj$meta.dat$month)[-1],
  base.size = 10,
  theme.choice = "bw",
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL
)
}
}
