% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_barplot_single.R
\name{generate_taxa_barplot_single}
\alias{generate_taxa_barplot_single}
\title{Generate stacked barplot pairs for taxonomic composition data}
\usage{
generate_taxa_barplot_single(
  data.obj,
  subject.var,
  time.var = NULL,
  t.level = NULL,
  group.var = NULL,
  strata.var = NULL,
  feature.level = "original",
  feature.dat.type = c("count", "proportion", "other"),
  feature.number = 20,
  base.size = 10,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.
and 'meta_tab' (metadata table).}

\item{subject.var}{A character vector specifying the subject variable.}

\item{time.var}{A character vector specifying the time variable.}

\item{group.var}{Optional character vector specifying the group variable.}

\item{strata.var}{Optional character vector specifying the strata variable.}

\item{feature.level}{A character vector specifying the taxonomic level to generate the plot for.}

\item{pdf}{A logical value indicating whether to save the plot as a PDF file (TRUE) or to display it (FALSE).}

\item{file.ann}{Optional file annotation to append to the output PDF file.}

\item{...}{Additional arguments to be passed to the plot function.}

\item{other.abund.cutoff}{A numeric value to specify the cutoff for merging taxa with relative abundances
below this threshold into 'Other'.}
}
\value{
A ggplot stacked barplot with relative abundances of taxa stratified by subject and time (and optionally,
group and strata), displaying individual taxa above the specified other.abund.cutoff, and merging those below the
cutoff into 'Other'. If pdf == TRUE, the plot is saved to a PDF file in the working directory.
}
\description{
This function generates stacked barplots of relative abundances for the given microbial taxonomic composition
data, stratified by specified subject and time variables, and optionally by group and strata variables.
}
\examples{
library(microbiome)
library(tidyverse)
library(ggh4x)
library(vegan)
data(peerj32)
peerj32.obj <- list()
peerj32.phy <- peerj32$phyloseq
peerj32.obj <- mStat_convert_phyloseq_to_data_obj( = "Raw")

plot_list_all <- generate_taxa_barplot_single(
  data.obj = peerj32.obj,
  subject.var = "subject",
  time.var = NULL,
  t.level = NULL,
  group.var = "group",
  strata.var = NULL,
  feature.level = "Family",
  feature.dat.type = "count",
  feature.number = 30,
  base.size = 10,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = c(ggsci::pal_nejm()(8),ggsci::pal_npg()(9),ggsci::pal_jama()(7),ggsci::pal_lancet()(9)),
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)
}
