% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_boxplot_long.R
\name{generate_taxa_boxplot_long}
\alias{generate_taxa_boxplot_long}
\title{Generate boxplot pair of individual taxa abundance}
\usage{
generate_taxa_boxplot_long(
  data.obj,
  subject.var,
  time.var,
  t0.level = NULL,
  ts.levels = NULL,
  group.var = NULL,
  strata.var = NULL,
  feature.level = NULL,
  feature.dat.type = c("count", "proportion", "other"),
  features.plot = NULL,
  top.k.plot = NULL,
  top.k.func = NULL,
  Transform = c("identity", "sqrt", "log"),
  prev.filter = 0.05,
  abund.filter = 0.01,
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{subject.var}{A character string specifying the subject variable in the metadata.}

\item{time.var}{A character string specifying the time variable in the metadata.}

\item{group.var}{A character string specifying the group variable in the metadata. Default is NULL.}

\item{strata.var}{A character string specifying the strata variable in the metadata. Default is NULL.}

\item{feature.level}{A character string specifying the taxonomic level for the boxplot. Default is 'Phylum'.}

\item{prev.filter}{A numeric value specifying the prevalence threshold for filtering taxa.}

\item{abund.filter}{A numeric value specifying the abundance threshold for filtering taxa.}

\item{pdf}{A logical value indicating whether to save the plot as a PDF file. Default is TRUE.}

\item{file.ann}{An optional character string to be used as a file annotation when saving the PDF. Default is NULL.}

\item{...}{Additional arguments passed to ggplot2 functions.}
}
\value{
A ggplot2 boxplot of individual taxa abundance at the specified taxonomic level.
}
\description{
This function creates a boxplot pair of individual taxa abundance at a specified
taxonomic level between two or more time points. It takes a data object containing
OTU, taxonomy, and metadata tables as input, along with metadata variables, and
outputs a ggplot2 boxplot.
}
\examples{
# Load required libraries and example data
library(microbiome)
library(tidyverse)
data(peerj32)
peerj32.obj <- list()
peerj32.phy <- peerj32$phyloseq
peerj32.obj <- mStat_convert_phyloseq_to_data_obj(peerj32.phy)

# Generate the boxplot pair
generate_taxa_boxplot_long(
  data.obj = ecam.obj,
  subject.var = "studyid",
  time.var = "month",
  t0.level = NULL,
  ts.levels = NULL,
  group.var = "diet",
  strata.var = NULL,
  feature.level = c("Phylum"),
  features.plot = sample(unique(ecam.obj$feature.ann[,"Phylum"]),1),
  feature.dat.type = "proportion",
  Transform = "log",
  prev.filter = 0,
  abund.filter = 0,
  base.size = 12,
  theme.choice = "classic",
  custom.theme = NULL,
  palette = c(ggsci::pal_npg()(9),ggsci::pal_jama()(7),ggsci::pal_lancet()(9)),
  pdf = TRUE,
  file.ann = "test",
  pdf.wid = 11,
  pdf.hei = 8.5
)

generate_taxa_boxplot_long(
  data.obj = peerj32.obj,
  subject.var = "subject",
  time.var = "time",
  t0.level = "1",
  ts.levels = "2",
  group.var = "group",
  strata.var = "sex",
  feature.level = c("Family"),
  feature.dat.type = "count",
  features.plot = NULL,
  top.k.plot = NULL,
  top.k.func = NULL,
  Transform = "log",
  prev.filter = 0,
  abund.filter = 0,
  base.size = 16,
  theme.choice = "classic",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)
}
