% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_change_heatmap_long.R
\name{generate_taxa_change_heatmap_long}
\alias{generate_taxa_change_heatmap_long}
\title{Generate Taxonomic Change Heatmap Long}
\usage{
generate_taxa_change_heatmap_long(
  data.obj,
  subject.var,
  time.var,
  t0.level = NULL,
  ts.levels = NULL,
  group.var = NULL,
  strata.var = NULL,
  feature.level,
  feature.dat.type = c("count", "proportion", "other"),
  features.plot = NULL,
  top.k.plot = NULL,
  top.k.func = NULL,
  change.func = "relative difference",
  prev.filter = 1e-08,
  abund.filter = 0.01,
  base.size = 10,
  palette = NULL,
  cluster.rows = NULL,
  cluster.cols = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{subject.var}{The name of the subject variable in the samples}

\item{time.var}{The name of the time variable in the samples}

\item{group.var}{The name of the grouping variable in the samples}

\item{strata.var}{The name of the strata variable in the samples}

\item{feature.level}{The taxonomic level to aggregate, can be “Phylum”, “Family” or “Genus”}

\item{prev.filter}{The prevalence filter to apply}

\item{abund.filter}{The abundance filter to apply}

\item{pdf}{If TRUE, save the plot as a PDF file (default: TRUE)}

\item{file.ann}{The file name annotation (default: NULL)}

\item{...}{Additional arguments to be passed to pheatmap (default: NULL)}
}
\value{
An object of class pheatmap, the generated heatmap plot
}
\description{
This function performs hierarchical clustering on microbiome data based on grouping
variables and strata variables in sample metadata and generates stacked heatmaps
using the “pheatmap” package. It can also save the resulting heatmap as a PDF file.
}
\examples{
library("HMP2Data")
library(tidyverse)
library(pheatmap)
momspi.phy <- momspi16S()
momspi.obj <- mStat_convert_phyloseq_to_data_obj( = "Raw")

plot_list <- generate_taxa_change_heatmap_long(
  data.obj = ecam.obj,
  subject.var = "studyid",
  time.var = "month_num",
  t0.level = NULL,
  ts.levels = NULL,
  group.var = "antiexposedall",
  strata.var = "diet",
  feature.level = c("Family"),
  feature.dat.type = "proportion",
  features.plot = NULL,
  top.k.plot = 10,
  top.k.func = "sd",
  change.func = "lfc",
  palette = NULL,
  prev.filter = 0.01,
  abund.filter = 0.01,
  pdf = TRUE,
  file.ann = "test"
)
momspi.obj <- mStat_convert_phyloseq_to_data_obj(momspi.phy)
plot_list <- generate_taxa_change_heatmap_long(
  data.obj = momspi.obj,
  subject.var = "subject_id",
  time.var = "visit_number",
  t0.level = " 1",
  ts.levels = sort(unique(momspi.obj$meta.dat$visit_number))[-1],
  group.var = "sample_body_site",
  strata.var = NULL,
  feature.level = "Phylum",
  feature.dat.type = "count",
  prev.filter = 0,
  abund.filter = 0.01,
  pdf = TRUE,
  file.ann = "test"
)
print(plot_list)


}
\seealso{
\code{\link{pheatmap}}
}
