% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_change_heatmap_pair.R
\name{generate_taxa_change_heatmap_pair}
\alias{generate_taxa_change_heatmap_pair}
\title{Generate Taxa Change Heatmap Pair}
\usage{
generate_taxa_change_heatmap_pair(
  data.obj,
  subject.var,
  time.var,
  group.var = NULL,
  strata.var = NULL,
  change.base = NULL,
  change.func = "relative difference",
  feature.level = NULL,
  feature.dat.type = c("count", "proportion", "other"),
  features.plot = NULL,
  top.k.plot = NULL,
  top.k.func = NULL,
  prev.filter = 0.1,
  abund.filter = 0.1,
  base.size = 10,
  palette = NULL,
  cluster.rows = NULL,
  cluster.cols = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{subject.var}{A character string defining subject variable in the `meta_tab`}

\item{time.var}{A character string defining time variable in the `meta_tab`}

\item{group.var}{A character string defining group variable in the `meta_tab` used for sorting and facetting}

\item{strata.var}{A character string defining strata variable in the `meta_tab` used for sorting and facetting}

\item{change.base}{A numeric value setting base for the change (usually 1)}

\item{change.func}{A character string specifying the function to apply on the change, default is "log"}

\item{feature.level}{A character string defining the taxonomic level to analyze, e.g., "Phylum", "Family", or "Genus"}

\item{prev.filter}{A numeric value defining the prevalence threshold to filter taxa, between 0 and 1}

\item{abund.filter}{A numeric value defining the abundance threshold to filter taxa}

\item{pdf}{A logical value. If TRUE (default), saves the heatmap as a PDF file. If FALSE, the heatmap will be displayed interactively without creating a PDF}

\item{file.ann}{(Optional) A character string specifying a file annotation to include in the generated PDF file's name}

\item{...}{Additional parameters to be passed}

\item{zero.handle}{A character string specifying how to handle zeros, default is "pseudo"}
}
\value{
If the `pdf` parameter is set to TRUE, the function will save a PDF file and return the pheatmap plot. If `pdf` is set to FALSE, the function will return the pheatmap plot without creating a PDF file.
}
\description{
This function generates a heatmap showing the pairwise changes in relative abundances of taxa between different time points.
The data used in this visualization will be first filtered based on prevalence and abundance thresholds.
The plot can either be displayed interactively or saved as a PDF file.
}
\examples{
# Load required libraries and example data
library(microbiome)
library(tidyverse)
library(pheatmap)
library(circlize)
data(peerj32)
peerj32.obj <- list()
peerj32.phy <- peerj32$phyloseq
peerj32.obj <- mStat_convert_phyloseq_to_data_obj(peerj32.phy)
plot_list <- generate_taxa_change_heatmap_pair(
  data.obj = peerj32.obj,
  subject.var = "subject",
  time.var = "time",
  group.var = "group",
  strata.var = NULL,
  change.base = "1",
  change.func = "relative difference",
  feature.level = c("Family"),
  feature.dat.type = "count",
  features.plot = NULL,
  top.k.plot = NULL,
  top.k.func = NULL,
  prev.filter = 0.1,
  abund.filter = 0.1,
  base.size = 10,
  palette = NULL,
  cluster.rows = NULL,
  cluster.cols = FALSE,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)

}
\seealso{
\code{\link{pheatmap}}
}
