% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_dotplot_pair.R
\name{generate_taxa_dotplot_pair}
\alias{generate_taxa_dotplot_pair}
\title{Generate Taxa Stack Dotplot Pair}
\usage{
generate_taxa_dotplot_pair(
  data.obj,
  subject.var,
  time.var,
  group.var = NULL,
  strata.var = NULL,
  feature.level = NULL,
  feature.dat.type = c("count", "proportion", "other"),
  features.plot = NULL,
  top.k.plot = NULL,
  top.k.func = NULL,
  prev.filter = 0.1,
  abund.filter = 0.05,
  base.size = 16,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list).}

\item{subject.var}{A character string defining subject variable in meta_tab}

\item{time.var}{A character string defining time variable in meta_tab}

\item{group.var}{A character string defining group variable in meta_tab used for sorting and facetting}

\item{strata.var}{(Optional) A character string defining strata variable in meta_tab used for sorting and facetting}

\item{feature.level}{A character string defining the taxonomic level to analyze ('Phylum', 'Family', or 'Genus')}

\item{feature.dat.type}{A character string specifying the type of the data in feature.dat. Options are "count", "proportion", or "other".}

\item{features.plot}{A character vector specifying the taxa to be plotted. If NULL (default), the top k taxa by mean abundance will be plotted.}

\item{top.k.plot}{A numeric value specifying the number of top taxa to be plotted if features.plot is NULL. If NULL (default), all taxa will be plotted.}

\item{top.k.func}{A function to compute the top k taxa if features.plot is NULL. If NULL (default), the mean function will be used.}

\item{prev.filter}{A numeric value defining the prevalence threshold to filter taxa, between 0 and 1}

\item{abund.filter}{A numeric value defining the abundance threshold to filter taxa}

\item{base.size}{Base font size for the generated plots.}

\item{theme.choice}{The choice of theme for the plot. Default is "bw".}

\item{custom.theme}{Custom ggplot2 theme (optional).}

\item{palette}{Color palette used for the plots.}

\item{pdf}{If TRUE, save the plot as a PDF file (default: TRUE)}

\item{file.ann}{(Optional) A character string specifying a file annotation to include in the generated PDF file's name}

\item{pdf.wid}{Width of the PDF plots.}

\item{pdf.hei}{Height of the PDF plots.}

\item{...}{Additional parameters to be passed}
}
\value{
If the `pdf` parameter is set to TRUE, the function will save a PDF file and return the final ggplot object. If `pdf` is set to FALSE, the function will return the final ggplot object without creating a PDF file.
}
\description{
This function generates a stacked dotplot of specified taxa level with paired samples. The data used in this
visualization will be first filtered based on prevalence and abundance thresholds. The plot can either be displayed
interactively or saved as a PDF file.
}
\examples{
# Load required libraries
data(peerj32.obj)

# Call the function
dotplot_result <- generate_taxa_dotplot_pair(
  data.obj = peerj32.obj,
  subject.var = "subject",
  time.var = "time",
  group.var = "group",
  strata.var = "sex",
  feature.level = c("Family"),
  feature.dat.type = "count",
  features.plot = NULL,
  top.k.plot = NULL,
  top.k.func = NULL,
  prev.filter = 0.01,
  abund.filter = 0.01,
  base.size = 16,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)

# View the result
print(dotplot_result)
}
