% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_indiv_boxplot_long.R
\name{generate_taxa_indiv_boxplot_long}
\alias{generate_taxa_indiv_boxplot_long}
\title{Generate Boxplot of Individual Taxa Abundance Over Time}
\usage{
generate_taxa_indiv_boxplot_long(
  data.obj,
  subject.var,
  time.var,
  t0.level = NULL,
  ts.levels = NULL,
  group.var = NULL,
  strata.var = NULL,
  feature.level = NULL,
  features.plot = NULL,
  feature.dat.type = c("count", "proportion", "other"),
  top.k.plot = NULL,
  top.k.func = NULL,
  transform = c("identity", "sqrt", "log"),
  prev.filter = 0.01,
  abund.filter = 0.01,
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list).}

\item{subject.var}{A character string specifying the subject variable in the metadata.}

\item{time.var}{A character string specifying the time variable in the metadata.}

\item{t0.level}{The base level for time points in longitudinal data.}

\item{ts.levels}{The levels for time points in longitudinal data.}

\item{group.var}{Optional grouping variable in metadata.}

\item{strata.var}{Optional stratification variable in metadata.}

\item{feature.level}{Taxonomic level(s) for boxplots.}

\item{features.plot}{A character vector specifying the taxa to be plotted. If NULL (default), the top k taxa by mean abundance will be plotted.}

\item{feature.dat.type}{A character string specifying the type of the data in feature.dat. Options are "count", "proportion", or "other".}

\item{top.k.plot}{A numeric value specifying the number of top taxa to be plotted if features.plot is NULL. If NULL (default), all taxa will be plotted.}

\item{top.k.func}{A function to compute the top k taxa if features.plot is NULL. If NULL (default), the mean function will be used.}

\item{transform}{transformation to apply before plotting. Default is "log10".}

\item{prev.filter}{Prevalence threshold for filtering taxa. Default 0.05.}

\item{abund.filter}{Abundance threshold for filtering taxa. Default 0.01.}

\item{base.size}{Base font size for the generated plots.}

\item{theme.choice}{Plot theme choice (default: "bw").}

\item{custom.theme}{Custom ggplot2 theme (optional).}

\item{palette}{Color palette used for the plots.}

\item{pdf}{Logical, if TRUE save plot as a multi-page PDF file. Default is TRUE.}

\item{file.ann}{Optional string for file annotation to add to PDF name.}

\item{pdf.wid}{Width of the PDF plots.}

\item{pdf.hei}{Height of the PDF plots.}

\item{...}{Additional arguments passed to ggplot2 functions.}
}
\value{
A ggplot object showing the abundance distribution of taxa over time.
}
\description{
This function creates a boxplot showing the abundance distribution of individual taxa at a specified taxonomic level over time from longitudinal data. It takes a MicrobiomeStat data object as input.
}
\examples{
\dontrun{
# Generate the boxplot pair
data(ecam.obj)
generate_taxa_indiv_boxplot_long(
  data.obj = ecam.obj,
  subject.var = "studyid",
  time.var = "month",
  t0.level = NULL,
  ts.levels = NULL,
  group.var = NULL,
  strata.var = NULL,
  feature.level = c("Phylum"),
  feature.dat.type = "proportion",
  transform = "log",
  prev.filter = 0.01,
  abund.filter = 0.01,
  base.size = 20,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = c("#1f77b4", "#ff7f0e", "#2ca02c", "#d62728", "#9467bd",
  "#8c564b", "#e377c2", "#7f7f7f", "#bcbd22", "#17becf", "#aec7e8",
  "#ffbb78", "#98df8a", "#ff9896", "#c5b0d5", "#c49c94", "#f7b6d2",
  "#c7c7c7", "#dbdb8d", "#9edae5", "#f0f0f0", "#3182bd"),
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)

data(peerj32.obj)

generate_taxa_indiv_boxplot_long(
  data.obj = peerj32.obj,
  subject.var = "subject",
  time.var = "time",
  t0.level = "1",
  ts.levels = "2",
  group.var = "group",
  strata.var = NULL,
  feature.level = c("Family"),
  features.plot = NULL,
  feature.dat.type = "other",
  top.k.plot = NULL,
  top.k.func = NULL,
  transform = "log",
  prev.filter = 0.01,
  abund.filter = 0.01,
  base.size = 20,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)
}
}
