% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_indiv_boxplot_single.R
\name{generate_taxa_indiv_boxplot_single}
\alias{generate_taxa_indiv_boxplot_single}
\title{Generate boxplot of individual taxa abundance distribution}
\usage{
generate_taxa_indiv_boxplot_single(
  data.obj,
  subject.var,
  time.var,
  t.level,
  group.var = NULL,
  strata.var = NULL,
  feature.level = NULL,
  features.plot = NULL,
  feature.dat.type = c("count", "proportion", "other"),
  top.k.plot = NULL,
  top.k.func = NULL,
  Transform = c("identity", "sqrt", "log"),
  prev.filter = 0.05,
  abund.filter = 0.01,
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{MicrobiomeStat data object containing OTU, taxonomy, and metadata.}

\item{subject.var}{Character string specifying subject variable in metadata.}

\item{time.var}{Character string specifying time variable in metadata.}

\item{group.var}{Character string specifying optional group variable in metadata.}

\item{strata.var}{Character string specifying optional strata variable in metadata.}

\item{feature.level}{Character vector specifying taxonomic level(s) for boxplots. Default 'Phylum'.}

\item{features.plot}{Character vector of specific taxa names to plot. Default NULL plots all.}

\item{top.k.plot}{Integer specifying number of top taxa to plot if features.plot is NULL. Default NULL plots all.}

\item{top.k.func}{Function specifying metric for determining top k taxa if features.plot is NULL. Default NULL uses row means.}

\item{Transform}{Transformation to apply to abundance data before plotting. Default is "log10".}

\item{prev.filter}{Prevalence threshold for filtering taxa before plotting.}

\item{abund.filter}{Abundance threshold for filtering taxa before plotting.}

\item{pdf}{Logical, if TRUE save plot as a multi-page PDF file. Default is TRUE.}

\item{file.ann}{Optional character string to add to PDF file name.}

\item{...}{Additional arguments passed to ggplot2 functions.}
}
\value{
A ggplot object showing the abundance distribution of taxa.
}
\description{
This function creates a boxplot showing the abundance distribution of individual taxa at a specified
taxonomic level at one or more time points from cross-sectional data. It takes a MicrobiomeStat data object as input.
}
\examples{
# Load required libraries and example data
library(microbiome)
library(tidyverse)
data(peerj32)
peerj32.obj <- list()
peerj32.phy <- peerj32$phyloseq
peerj32.obj <- mStat_convert_phyloseq_to_data_obj(peerj32.phy)

# Generate the boxplot pair
generate_taxa_indiv_boxplot_single(
  data.obj = ecam.obj,
  subject.var = "studyid",
  time.var = "month",
  t0.level = NULL,
  ts.levels = NULL,
  group.var = NULL,
  strata.var = NULL,
  feature.level = c("Phylum"),
  feature.dat.type = "proportion",
  Transform = "log",
  prev.filter = 0.1,
  abund.filter = 0.1,
  base.size = 20,
  theme.choice = "classic",
  custom.theme = NULL,
  palette = c(ggsci::pal_npg()(9),ggsci::pal_jama()(7),ggsci::pal_lancet()(9)),
  pdf = TRUE,
  file.ann = "test",
  pdf.wid = 11,
  pdf.hei = 8.5
)

generate_taxa_indiv_boxplot_single(
  data.obj = peerj32.obj,
  subject.var = "subject",
  time.var = "time",
  t.level = "1",
  group.var = "group",
  strata.var = "sex",
  feature.level = c("Family"),
  features.plot = NULL,
  feature.dat.type = "count",
  top.k.plot = NULL,
  top.k.func = NULL,
  Transform = "log",
  prev.filter = 0,
  abund.filter = 0,
  base.size = 20,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)
}
