% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_change_boxplot_pair.R,
%   R/generate_taxa_indiv_change_boxplot_pair.R
\name{generate_taxa_indiv_change_boxplot_pair}
\alias{generate_taxa_indiv_change_boxplot_pair}
\alias{generate_taxa_change_boxplot_pair}
\title{Generate taxa-level individual change boxplot pairs}
\usage{
generate_taxa_change_boxplot_pair(
  data.obj,
  subject.var,
  time.var,
  group.var = NULL,
  strata.var = NULL,
  change.base = NULL,
  change.func = "relative difference",
  feature.level = NULL,
  feature.dat.type = c("count", "proportion", "other"),
  features.plot = NULL,
  top.k.plot = NULL,
  top.k.func = NULL,
  prev.filter,
  abund.filter,
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)

generate_taxa_indiv_change_boxplot_pair(
  data.obj,
  subject.var,
  time.var,
  group.var = NULL,
  strata.var = NULL,
  change.base = NULL,
  change.func = "relative difference",
  feature.level = NULL,
  features.plot = NULL,
  feature.dat.type = c("count", "proportion", "other"),
  top.k.plot = NULL,
  top.k.func = NULL,
  prev.filter = 0.001,
  abund.filter = 0.001,
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list).}

\item{subject.var}{A character string defining the subject variable in the metadata.}

\item{time.var}{A character string defining the time variable in the metadata.}

\item{group.var}{A character string defining the group variable in the metadata used for sorting and facetting.}

\item{strata.var}{(Optional) A character string defining strata variable in the metadata used for sorting and facetting.}

\item{change.base}{A numeric value specifying the base time point for calculating the change in abundance.}

\item{change.func}{A character string specifying the function to apply for calculating the change in abundance. Default is 'log'.}

\item{feature.level}{A character vector specifying the taxa level(s) to include in the analysis. Default is c('Phylum', 'Family', 'Genus').}

\item{feature.dat.type}{A character string specifying the type of the data in feature.dat. Options are "count", "proportion", or "other".}

\item{features.plot}{A character vector specifying the taxa to be plotted. If NULL (default), the top k taxa by mean abundance will be plotted.}

\item{top.k.plot}{A numeric value specifying the number of top taxa to be plotted if features.plot is NULL. If NULL (default), all taxa will be plotted.}

\item{top.k.func}{A function to compute the top k taxa if features.plot is NULL. If NULL (default), the mean function will be used.}

\item{prev.filter}{A numeric value defining the prevalence threshold to filter taxa, between 0 and 1.}

\item{abund.filter}{A numeric value defining the abundance threshold to filter taxa.}

\item{base.size}{Base font size for the generated plots.}

\item{theme.choice}{Plot theme choice (default: "bw").}

\item{custom.theme}{Custom ggplot2 theme (optional).}

\item{palette}{Color palette used for the plots.}

\item{pdf}{A logical value. If TRUE (default), saves the plot as a PDF file. If FALSE, the plot will be displayed interactively without creating a PDF.}

\item{file.ann}{(Optional) A character string specifying a file annotation to include in the generated PDF file's name.}

\item{pdf.wid}{Width of the PDF plots.}

\item{pdf.hei}{Height of the PDF plots.}

\item{...}{Additional parameters to be passed.}

\item{zero.handle}{A character string specifying the method to handle zero values. Default is 'pseudo'.}
}
\value{
A ggplot object representing the taxa-level individual change boxplot pair.

A ggplot object representing the taxa-level individual change boxplot pair.
}
\description{
This function generates a boxplot pair that shows the change in abundance of taxa at different levels (Phylum, Family, or Genus) between two time points.
It allows users to compare the changes in abundance within and between different groups and strata.

This function generates boxplot pairs showing the change in abundance of taxa at different taxonomic levels between two time points.
It allows users to compare the changes in abundance within and between different groups (specified by the `group.var` parameter).
If a stratification variable (`strata.var`) is provided, the function will also stratify the analysis by this variable.
}
\examples{
# Load required libraries and data
library(microbiome)
library(vegan)
library(tidyverse)
library(ggh4x)
data(peerj32)
peerj32.obj <- list()
peerj32.phy <- peerj32$phyloseq
peerj32.obj <- mStat_convert_phyloseq_to_data_obj(peerj32.phy)

# Generate the boxplot pair
generate_taxa_change_boxplot_pair(
  data.obj = peerj32.obj,
  subject.var = "subject",
  time.var = "time",
  group.var = "group",
  strata.var = NULL,
  change.base = "1",
  change.func = "lfc",
  feature.level = c("Family"),
  feature.dat.type = "count",
  features.plot = NULL,
  top.k.plot = 6,
  top.k.func = "sd",
  prev.filter = 0.1,
  abund.filter = 0.01,
  base.size = 20,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = "test",
  pdf.wid = 11,
  pdf.hei = 8.5
)
# Load required libraries and data
library(microbiome)
library(vegan)
library(tidyverse)
library(ggh4x)
data(peerj32)
peerj32.obj <- list()
peerj32.phy <- peerj32$phyloseq
peerj32.obj <- mStat_convert_phyloseq_to_data_obj(peerj32.phy)

# Generate the boxplot pair
generate_taxa_indiv_change_boxplot_pair(
  data.obj = peerj32.obj,
  subject.var = "subject",
  time.var = "time",
  group.var = "group",
  strata.var = NULL,
  change.base = "1",
  change.func = "lfc",
  feature.level = c("Family"),
  features.plot = NULL,
  feature.dat.type = "count",
  top.k.plot = NULL,
  top.k.func = NULL,
  prev.filter = 0.1,
  abund.filter = 0.01,
  base.size = 20,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)
}
