% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_spaghettiplot_long.R
\name{generate_taxa_spaghettiplot_long}
\alias{generate_taxa_spaghettiplot_long}
\title{Generate longitudinal line plots of taxonomic composition}
\usage{
generate_taxa_spaghettiplot_long(
  data.obj,
  subject.var,
  time.var,
  t0.level = NULL,
  ts.levels = NULL,
  group.var = NULL,
  strata.var = NULL,
  feature.level = NULL,
  feature.dat.type = c("count", "proportion", "other"),
  features.plot = NULL,
  top.k.plot = NULL,
  top.k.func = NULL,
  prev.filter = 0.01,
  abund.filter = 0.01,
  base.size = 16,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{subject.var}{A character string defining subject variable in meta_tab}

\item{time.var}{A character string defining time variable in meta_tab}

\item{group.var}{A character string defining group variable in meta_tab used for sorting and facetting}

\item{strata.var}{(Optional) A character string defining strata variable in meta_tab used for sorting and facetting}

\item{prev.filter}{A numeric value defining the prevalence threshold to filter taxa, between 0 and 1}

\item{abund.filter}{A numeric value defining the abundance threshold to filter taxa}

\item{pdf}{A logical value. If TRUE (default), saves the dotplot as a PDF file. If FALSE, the dotplot will be displayed interactively without creating a PDF}

\item{file.ann}{(Optional) A character string specifying a file annotation to include in the generated PDF file's name}

\item{...}{Additional parameters to be passed}

\item{taxa.level}{A character string defining the taxonomic level to analyze ('Phylum', 'Family', or 'Genus')}
}
\value{
If the `pdf` parameter is set to TRUE, the function will save a PDF file and return the final ggplot object. If `pdf` is set to FALSE, the function will return the final ggplot object without creating a PDF file.
}
\description{
Generate longitudinal line plots of taxonomic composition
}
\examples{
library(ggh4x)
plot_list_all <- generate_taxa_spaghettiplot_long(
  data.obj = ecam.obj,
  subject.var = "studyid",
  time.var = "month",
  t0.level = NULL,
  ts.levels = NULL,
  group.var = "diet",
  strata.var = "antiexposedall",
  feature.level = c("Phylum"),
  features.plot = NULL,
  feature.dat.type = "proportion",
  top.k.plot = 5,
  top.k.func = "mean",
  prev.filter = 0,
  abund.filter = 0,
  base.size = 16,
  theme.choice = "classic",
  palette = NULL,
  pdf = TRUE,
  file.ann = "test"
)

}
