% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_trend_volcano_long.R
\name{generate_taxa_trend_volcano_long}
\alias{generate_taxa_trend_volcano_long}
\title{Generate Volcano Plots for Longitudinal Taxa Trend Test}
\usage{
generate_taxa_trend_volcano_long(
  data.obj,
  group.var = NULL,
  time.var = NULL,
  test.list,
  feature.sig.level = 0.1,
  feature.mt.method = "fdr",
  palette = c("#F9F871", "#F4A261", "#FF6347"),
  pdf = FALSE,
  pdf.wid = 7,
  pdf.hei = 5
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list).}

\item{group.var}{The grouping variable tested, found in metadata.}

\item{time.var}{The time variable used in the analysis.}

\item{test.list}{The list of test results returned by generate_taxa_trend_test_long.}

\item{feature.sig.level}{The significance level cutoff for highlighting taxa.}

\item{feature.mt.method}{Multiple testing correction method, "fdr" or "none".}

\item{palette}{An optional parameter specifying the color palette to be used for the plot.
It can be either a character string specifying the name of a predefined
palette or a vector of color codes in a format accepted by ggplot2
(e.g., hexadecimal color codes). Available predefined palettes include
'npg', 'aaas', 'nejm', 'lancet', 'jama', 'jco', and 'ucscgb', inspired
by various scientific publications and the `ggsci` package. If `palette`
is not provided or an unrecognized palette name is given, a default color
palette will be used. Ensure the number of colors in the palette is at
least as large as the number of groups being plotted.}

\item{pdf}{Boolean; whether to save the plot as a PDF file.}

\item{pdf.wid}{Numeric; width of the saved PDF file.}

\item{pdf.hei}{Numeric; height of the saved PDF file.}
}
\value{
A list of ggplot objects of volcano plots for each taxonomic level.
}
\description{
Generate Volcano Plots for Longitudinal Taxa Trend Test
}
\details{
The function generates volcano plots for each taxonomic level based on the test results. It visualizes the longitudinal trends of taxa abundance over time and highlights statistically significant changes.
}
\examples{
\dontrun{
data("subset_T2D.obj")
test.list <- generate_taxa_trend_test_long(
  data.obj = subset_T2D.obj,
  subject.var = "subject_id",
  time.var = "visit_number",
  group.var = "subject_race",
  adj.vars = "sample_body_site",
  prev.filter = 0.1,
  abund.filter = 0.001,
  feature.level = c("Genus","Family"),
  feature.dat.type = c("count")
)

plot.list <- generate_taxa_trend_volcano_long(
  data.obj = subset_T2D.obj,
  group.var = "subject_race",
  time.var = "visit_number_num",
  test.list = test.list,
  feature.sig.level = 0.1,
  feature.mt.method = "none")
}

}
