% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_subset_dist.R
\name{is.na.null}
\alias{is.na.null}
\title{Check if an Object is NA or NULL}
\usage{
\method{is.na}{null}(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
A logical value indicating whether the input is `NA` or `NULL`.
}
\description{
This function checks if the given object is `NA` or `NULL`.
}
\details{
The function first checks if the object `x` is `NULL`. If not, it checks if the first element of `x` is `NA`.
If either condition is true, it returns `TRUE`; otherwise, it returns `FALSE`.
}
\examples{
\dontrun{
# Check if a variable is NA or NULL
is.na.null(NA)  # TRUE
is.na.null(NULL)  # TRUE
is.na.null(1)  # FALSE
}

}
