% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_change_test_pair.R
\name{is_categorical}
\alias{is_categorical}
\title{Determine if a variable is categorical}
\usage{
is_categorical(x)
}
\arguments{
\item{x}{The variable to be tested for being categorical.}
}
\value{
A logical value indicating whether the input variable is categorical (TRUE) or not (FALSE).
}
\description{
This function checks if a given variable is categorical. A variable is considered categorical if it is a factor, a character, or a logical type. Additionally, numeric variables with a small number of unique values (less than a specified threshold) are also considered categorical.
}
\examples{
is_categorical(factor(c("A", "B", "C")))
is_categorical(c("A", "B", "C"))
is_categorical(c(TRUE, FALSE))
is_categorical(c(1, 2, 3, 4))
is_categorical(c(1.2, 3.5, 4.6, 7.8))

}
