% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_alpha_test_pair.R
\name{is_rarefied}
\alias{is_rarefied}
\title{Check if the data has been rarefied}
\usage{
is_rarefied(data.obj)
}
\arguments{
\item{data.obj}{A list object that includes feature.tab (an OTU table with
taxa as rows and samples as columns) and meta.dat (a metadata table with
samples as rows and variables as columns).}
}
\value{
A boolean value indicating whether the data is rarefied. It returns
  TRUE if the data is rarefied, FALSE otherwise.
}
\description{
This function checks if the data has been rarefied by inspecting if the sum
of each column (which represents each sample in the OTU table) is equal.
}
\examples{
# Assuming peerj32.obj is a data object with OTU and metadata tables
data(peerj32.obj)
is_rarefied(peerj32.obj)

}
