% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_aggregate_by_taxonomy2.R
\name{mStat_aggregate_by_taxonomy2}
\alias{mStat_aggregate_by_taxonomy2}
\title{Aggregate OTU Table by Taxonomy Level}
\usage{
mStat_aggregate_by_taxonomy2(feature.tab, feature.ann, feature.level = NULL)
}
\arguments{
\item{feature.tab}{OTU table as matrix or data.frame}

\item{feature.ann}{Taxonomy table as matrix or data.frame}

\item{feature.level}{Taxonomy level to aggregate by as character vector.}
}
\value{
Aggregated OTU table matrix.
}
\description{
This function aggregates an OTU table by a specified taxonomy level.
It checks for consistency between OTU and taxonomy tables before aggregation.
}
\details{
The function first checks that the row names of OTU and taxonomy tables match.
If any mismatches are found, a message is displayed but the function continues.
The OTU and taxonomy tables are joined by row names.
The OTU table is then aggregated by summing abundance values for each sample at the specified taxonomy level.
Rows with all zero values after aggregation are removed.
}
\examples{
\dontrun{
# Load required libraries
library(vegan)
data(peerj32.obj)

# Specify the taxonomy level
feature.level <- c("Family")

# Aggregate data object by taxonomy level
mStat_aggregate_by_taxonomy2(peerj32.obj$feature.tab, peerj32.obj$feature.ann, feature.level)
}
}
