% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_calculate_adjusted_distance.R
\name{mStat_calculate_adjusted_distance}
\alias{mStat_calculate_adjusted_distance}
\title{Calculate Adjusted Distances for Specified Variables}
\usage{
mStat_calculate_adjusted_distance(data.obj, dist.obj, adj.vars, dist.name)
}
\arguments{
\item{data.obj}{A MicrobiomeStat data object. This central component contains:
\itemize{
  \item \strong{feature.tab}: Matrix representation linking research objects and samples.
  \item \strong{meta.dat}: Data frame with rows representing samples and columns as annotations.
  \item \strong{feature.ann}: Matrix annotations with classification data.
  \item \strong{phylogenetic tree}: (Optional) Tree depicting evolutionary relationships.
  \item \strong{feature.agg.list}: (Optional) Aggregated results from feature.tab and feature.ann.
}}

\item{dist.obj}{Distance matrix between samples, usually calculated using
\code{\link[MicrobiomeStat]{mStat_calculate_beta_diversity}} function.
If NULL, beta diversity will be automatically computed from \code{data.obj}
using \code{mStat_calculate_beta_diversity}.}

\item{adj.vars}{Character vector listing variable names intended for adjustment.}

\item{dist.name}{Character vector specifying names of distance matrices present in `dist.obj`.}
}
\value{
A list of adjusted distance matrices. Each matrix is named according to the `dist.name` parameter.
}
\description{
Computes adjusted distance matrices for a set of variables using a dynamic model matrix.
}
\details{
Calculate Adjusted Distances Based on Specified Adjustment Variables

This function provides a way to calculate distances adjusted for specified variables in a dataset.
It incorporates the results of multidimensional scaling (MDS) and adjusts distances based on a linear model.

The function uses cmdscale for multidimensional scaling and then adjusts the resultant distances based on a linear model using the variables provided in `adj.vars`.
}
\examples{
data("subset_T2D.obj")
subset_T2D.dist.obj <- mStat_calculate_beta_diversity(subset_T2D.obj, c("BC","Jaccard"))
adj.dist.obj <- mStat_calculate_adjusted_distance(
data.obj = subset_T2D.obj,
dist.obj = subset_T2D.dist.obj,
adj.vars = c("subject_gender", "subject_race"),
dist.name = c("BC"))

data("peerj32.obj")
peerj32.dist.obj <- mStat_calculate_beta_diversity(peerj32.obj, c("BC"))
adj.dist.obj <- mStat_calculate_adjusted_distance(
data.obj = peerj32.obj,
dist.obj = peerj32.dist.obj,
adj.vars = c("sex"),
dist.name = c("BC"))

}
