% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_combine_data.R
\name{mStat_combine_data}
\alias{mStat_combine_data}
\title{mStat_combine_data}
\usage{
mStat_combine_data(data.obj1, data.obj2)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}
}
\value{
A list with the combined 'feature.tab', 'meta.dat' and 'feature.ann'.
}
\description{
This function is designed specifically for combining two MicrobiomeStat data objects,
assuming that they are both in 'Raw' format. Each MicrobiomeStat data object should be a list that contains
'norm.status' (indicating the status of normalization, should be 'Raw'), 'feature.tab' (a matrix of features),
'meta.dat' (a data frame of metadata), and 'feature.ann' (a matrix of feature annotations).
}
\examples{
# combined_data <- mStat_combine_data(data_obj1, data_obj2)

}
