% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_convert_DESeqDataSet_to_data_obj.R
\name{mStat_convert_DESeqDataSet_to_data_obj}
\alias{mStat_convert_DESeqDataSet_to_data_obj}
\title{Convert a DESeqDataSet Object to a MicrobiomeStat Data Object}
\usage{
mStat_convert_DESeqDataSet_to_data_obj(dds.obj)
}
\arguments{
\item{dds.obj}{A DESeqDataSet object to be converted.}
}
\value{
A MicrobiomeStat data object (a list) containing the following elements:
\itemize{
  \item feature.tab: A matrix containing the counts data with rows as features and columns as samples.
  \item meta.dat: A data frame containing the samples data.
  \item feature.ann: A matrix containing feature annotations.
}
}
\description{
This function is a part of the MicrobiomeStat package. It converts a DESeqDataSet object into a list (MicrobiomeStat data object) containing the counts, samples data and feature annotations.
The returned list represents a microbiome dataset with features (OTUs or taxa), sample metadata and feature annotations.
}
\details{
The function first checks if each component (counts, samples, feature annotations) of the DESeqDataSet object is not null. If a component is not null, it is converted to the appropriate format and added to the output list. The counts data and feature annotations are converted to a matrix, while the samples data is converted to a data frame. Note that only the rows (features) in the counts data that have a sum > 0 are retained.
}
\examples{
\dontrun{
  # Load necessary packages
  # library(airway)
  # library(DESeq2)

  # Load dataset
  # data("airway")
  # dds <- DESeqDataSet(airway, design = ~ cell + dex)

  # Convert to MicrobiomeStat data object
  # data.obj <- mStat_convert_DESeqDataSet_to_data_obj(dds)
}

}
\author{
Caffery Yang
}
