% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_convert_phyloseq_to_data_obj.R
\name{mStat_convert_phyloseq_to_data_obj}
\alias{mStat_convert_phyloseq_to_data_obj}
\title{Convert a Phyloseq Object to a MicrobiomeStat's Data Object}
\usage{
mStat_convert_phyloseq_to_data_obj(phylo.obj)
}
\arguments{
\item{phylo.obj}{A phyloseq object to be converted. This should contain an OTU (operational taxonomic unit) table, sample data, taxonomy table, and a phylogenetic tree.}
}
\value{
A MicrobiomeStat data object (a list) containing the following elements:
\itemize{
  \item feature.tab: A matrix of the feature table. Rows with a sum of zero are removed, so only the features present in the samples are included.
  \item meta.dat: A data frame of the sample data. This contains the metadata for each of the samples.
  \item feature.ann: A matrix of the feature annotation table (taxonomy table). Only the rows that exist in the feature table are included.
  \item tree: A phylogenetic tree. The tree is rooted by midpointing if it is not already rooted. Tips not present in the feature table are dropped.
}
}
\description{
This function is a part of the MicrobiomeStat package. It takes a phyloseq object, extracts relevant information and repackages it in the MicrobiomeStat's data object format. This enables easy use of phyloseq data with MicrobiomeStat's analysis functions.
}
\details{
This function checks each component (feature table, sample data, taxonomy table, and phylogenetic tree) of the phyloseq object for null values. If a component is not null, it is converted to the appropriate format and added to the MicrobiomeStat data object. The feature and taxonomy tables are converted to matrices, while the sample data is converted to a data frame.

The function automatically handles the feature table orientation by checking the \code{taxa_are_rows} slot of the phyloseq object. If \code{taxa_are_rows} is FALSE (i.e., samples are rows and taxa are columns in the phyloseq object), the feature table will be automatically transposed to ensure that the resulting feature table has taxa as rows and samples as columns, which is the required format for MicrobiomeStat functions.

The phylogenetic tree is checked if it is rooted, and if not, it is rooted by midpointing. Tips not present in the feature table are dropped from the tree. This ensures the output data object is consistent and ready for further microbiome statistical analysis.
}
\examples{
\dontrun{
  #library(microbiome)
  # Load phangorn for midpoint rooting if tree is present
  #library(phangorn)
  #data(peerj32)
  #peerj32.phy <- peerj32$phyloseq
  #data.obj <- mStat_convert_phyloseq_to_data_obj(peerj32.phy)
}

}
\references{
McMurdie PJ, Holmes S. phyloseq: An R Package for Reproducible Interactive Analysis and Graphics of Microbiome Census Data. PLoS ONE. 2013;8(4):e61217.
}
\author{
Jun Chen
}
