% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mStat_convert_SummarizedExperiment_to_data_obj.R
\name{mStat_convert_SummarizedExperiment_to_data_obj}
\alias{mStat_convert_SummarizedExperiment_to_data_obj}
\title{Convert SummarizedExperiment Object to a MicrobiomeStat Data Object}
\usage{
mStat_convert_SummarizedExperiment_to_data_obj(se.obj)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object to be converted.}
}
\value{
A MicrobiomeStat data object (a list) containing the following elements:
\itemize{
  \item feature.tab: A matrix containing the assay data with rows as features and columns as samples.
  \item feature.ann: A matrix containing the rowData (feature annotations). Only features present in the assay data are included.
  \item meta.dat: A data frame containing the colData (metadata).
}
}
\description{
This function converts a SummarizedExperiment object into a list (MicrobiomeStat data object) containing the assays data, rowData, and colData.
The returned list represents a microbiome dataset with features (OTUs or taxa), metadata, and feature annotations.
}
\details{
The function first checks if each component (assays, rowData, colData) of the SummarizedExperiment object is not null. If a component is not null, it is converted to the appropriate format and added to the output list. The assays data is converted to a matrix, while the rowData and colData are converted to data frames. Note that only the rows (features) in the assay data that have a sum > 0 are retained.
}
\examples{
\dontrun{
  # Assuming 'se' is your SummarizedExperiment object
  # data_obj <- mStat_convert_SummarizedExperiment_to_data_obj(se)

  # If you have airway data available as a SummarizedExperiment object
  # you can convert it to a MicrobiomeStat data object using:
  # library(airway)
  # data(airway)
  # airway_obj <- mStat_convert_SummarizedExperiment_to_data_obj(airway)
}

}
\author{
Caffery Yang
}
