% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_filter.R
\name{mStat_filter}
\alias{mStat_filter}
\title{Filter a Microbiome Data Frame by Prevalence and Abundance}
\usage{
mStat_filter(x, prev.filter, abund.filter)
}
\arguments{
\item{x}{A matrix or data frame containing taxa (in rows) by sample (in columns)
microbial abundance data.}

\item{prev.filter}{Numeric, the minimum prevalence threshold for a taxon to be
retained. Prevalence is calculated as the proportion of samples where the
taxon is present.}

\item{abund.filter}{Numeric, the minimum average abundance threshold for a taxon
to be retained.}
}
\value{
A matrix or data frame with filtered taxa based on the specified thresholds.
}
\description{
This function filters taxa in a microbiome data frame based on a minimum
prevalence and average abundance threshold.
}
\details{
The function first reshapes the input data to long format and then groups by taxa
to compute the average abundance and prevalence. After this, it filters out taxa
based on the provided prevalence and abundance thresholds.
}
\examples{
\dontrun{
data_matrix <- matrix(c(0, 3, 4, 0, 2, 7, 8, 9, 10), ncol=3)
colnames(data_matrix) <- c("sample1", "sample2", "sample3")
rownames(data_matrix) <- c("taxa1", "taxa2", "taxa3")

filtered_data <- mStat_filter(data_matrix, 0.5, 5)
print(filtered_data)
}

}
