% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtreeExtra_wrapper.R
\name{mStat_geom_fruit}
\alias{mStat_geom_fruit}
\title{Wrapper for ggtreeExtra::geom_fruit to handle NSE correctly}
\usage{
mStat_geom_fruit(data, geom = NULL, mapping = NULL, ...)
}
\arguments{
\item{data}{Data to use for the layer}

\item{geom}{Name of the geom function to use (e.g., geom_tile)}

\item{mapping}{Aesthetic mapping}

\item{...}{Additional arguments to pass to ggtreeExtra::geom_fruit}
}
\value{
A layer to be added to a ggtree plot
}
\description{
Wrapper for ggtreeExtra::geom_fruit to handle NSE correctly
}
\keyword{internal}
