% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_get_palette.R
\name{mStat_get_palette}
\alias{mStat_get_palette}
\title{Retrieve Custom Color Palettes for Data Visualization}
\usage{
mStat_get_palette(palette = NULL)
}
\arguments{
\item{palette}{A character string specifying the name of a predefined palette from `ggsci` or
a vector of color codes. If `NULL`, a default palette is used.
Available predefined palettes include 'npg', 'aaas', 'nejm', 'lancet', 'jama', 'jco', and 'ucscgb',
all derived from the `ggsci` package.}
}
\value{
A vector of color codes. If a predefined palette name is provided, the corresponding
        palette from the `ggsci` package is returned. If a vector of color codes is provided, it is returned as-is.
        If the input is NULL or an unrecognized string, a default palette is returned.
}
\description{
This function provides a selection of color palettes inspired by various scientific publications,
particularly the `ggsci` package, which offers a wide range of color palettes used in scientific journals and publications.
It allows users to choose a predefined palette from `ggsci` or specify their own set of colors.
}
\examples{
# Default palette
mStat_get_palette()

# Predefined palette from `ggsci`: 'nejm'
mStat_get_palette("nejm")

# Other predefined palettes
# 'npg' palette
mStat_get_palette("npg")

# 'aaas' palette
mStat_get_palette("aaas")

# 'lancet' palette
mStat_get_palette("lancet")

# 'jama' palette
mStat_get_palette("jama")

# 'jco' palette
mStat_get_palette("jco")

# 'ucscgb' palette
mStat_get_palette("ucscgb")

# Custom color vector
mStat_get_palette(c("#123456", "#654321"))

# Unrecognized input returns default palette
mStat_get_palette("unknown")

# Using a longer custom color vector
mStat_get_palette(c("#FF5733", "#33FF57", "#3357FF", "#FF33FF"))

# Using an empty string (should return default palette)
mStat_get_palette("")

# Using a single color code (treated as custom vector)
mStat_get_palette("#FF5733")
}
