% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_import_dada2_as_data_obj.R
\name{mStat_import_dada2_as_data_obj}
\alias{mStat_import_dada2_as_data_obj}
\title{Convert DADA2 Data into MicrobiomeStat's Data Object}
\usage{
mStat_import_dada2_as_data_obj(
  seq_tab,
  tax_tab = NULL,
  sam_tab = NULL,
  phy_tree = NULL
)
}
\arguments{
\item{seq_tab}{A sequence table from DADA2, where rows represent samples, and columns represent sequences (features). This table is crucial for subsequent MicrobiomeStat analyses.}

\item{tax_tab}{(Optional) A taxonomy assignment table derived from DADA2. Each row corresponds to a feature (sequence), and each column corresponds to a different taxonomic rank. This table enriches the feature information.}

\item{sam_tab}{(Optional) A sample metadata table. Each row corresponds to a sample, and each column provides different metadata attributes. This table enriches the sample information.}

\item{phy_tree}{(Optional) A phylogenetic tree of class 'phylo'. This tree represents the phylogenetic relationships between features (sequences) and can be incorporated into certain analyses.}
}
\value{
A MicrobiomeStat data object, containing the following elements:
  \itemize{
    \item feature.tab: A matrix of the feature (sequence) table.
    \item meta.dat: A data frame of the sample metadata.
    \item feature.ann: A matrix of the feature (sequence) taxonomy annotations.
    \item tree: A phylogenetic tree, if provided.
  }
This object is readily accepted by MicrobiomeStat's downstream analysis functions.
}
\description{
The `mStat_import_dada2_as_data_obj` function acts as a bridge between DADA2 output and MicrobiomeStat, effectively enabling smooth data transition between the two. It accepts essential DADA2 outputs and converts them into a MicrobiomeStat compatible format for subsequent microbiome data analysis.
}
\details{
This function facilitates seamless integration of DADA2 outputs into the MicrobiomeStat analysis pipeline. It accepts DADA2-derived sequence table, taxonomy table, sample metadata, and optionally a phylogenetic tree, and structures them into a MicrobiomeStat-friendly format. Subsequently, the resulting object can be directly fed into various MicrobiomeStat's statistical analysis functions.
}
\examples{
\dontrun{
  seq_tab <- readRDS(
    system.file(
        "extdata", "dada2_seqtab.rds",
        package = "microbiomeMarker"
    )
  )
  tax_tab <- readRDS(
    system.file(
        "extdata", "dada2_taxtab.rds",
        package = "microbiomeMarker"
    )
  )
  sam_tab <- read.table(
    system.file(
        "extdata", "dada2_samdata.txt",
        package = "microbiomeMarker"
    ),
    sep = "\t",
    header = TRUE,
    row.names = 1
  )
  data_obj <- mStat_import_dada2_as_data_obj(seq_tab = seq_tab,
  tax_tab = tax_tab, sam_tab = sam_tab)

}

}
\seealso{
MicrobiomeStat documentation for further details on how to use the returned data object.
}
\author{
Caffery Yang
}
