% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_import_qiime2_as_data_obj.R
\name{mStat_import_qiime2_as_data_obj}
\alias{mStat_import_qiime2_as_data_obj}
\title{Convert Qiime2 Data into MicrobiomeStat's Data Object}
\usage{
mStat_import_qiime2_as_data_obj(
  otu_qza,
  taxa_qza = NULL,
  sam_tab = NULL,
  refseq_qza = NULL,
  tree_qza = NULL
)
}
\arguments{
\item{otu_qza}{A feature table with observations (OTUs, ASVs, etc.) in .qza format from Qiime2.}

\item{taxa_qza}{(Optional) A taxonomy assignment table in .qza format from Qiime2.}

\item{sam_tab}{(Optional) A sample metadata table.}

\item{refseq_qza}{(Optional) A representative sequences table in .qza format from Qiime2.}

\item{tree_qza}{(Optional) A phylogenetic tree in .qza format from Qiime2.}
}
\value{
A MicrobiomeStat data object that contains the feature table, metadata, feature annotations, phylogenetic tree, and normalization status. This object is ready to be used in MicrobiomeStat's downstream analysis functions.
}
\description{
This function serves as a bridge between Qiime2 and MicrobiomeStat, allowing seamless integration of Qiime2 data with the powerful MicrobiomeStat analysis toolkit. It accepts various Qiime2 microbiome study-related files and converts them into a MicrobiomeStat data object.
}
\details{
MicrobiomeStat is a comprehensive and user-friendly tool designed for robust microbiome data analysis. The resulting data object from this function can be directly fed into a series of MicrobiomeStat's statistical tools for further analysis.
}
\examples{
\dontrun{
# library(Biostrings)
data_obj <- mStat_import_qiime2_as_data_obj(otu_qza = "path_to_otu.qza",
                                             taxa_qza = "path_to_taxa.qza",
                                             sam_tab = "path_to_sample_metadata",
                                             refseq_qza = "path_to_refseq.qza",
                                             tree_qza = "path_to_tree.qza",
                                             norm.status = "Raw")
}
otuqza_file <- system.file("extdata", "table.qza", package = "microbiomeMarker")
taxaqza_file <- system.file("extdata", "taxonomy.qza", package = "microbiomeMarker")
sample_file <- system.file("extdata", "sample-metadata.tsv", package = "microbiomeMarker")
treeqza_file <- system.file("extdata", "tree.qza", package = "microbiomeMarker")
data.obj <- mStat_import_qiime2_as_data_obj(
    otu_qza = otuqza_file, taxa_qza = taxaqza_file,
    sam_tab = sample_file, tree_qza = treeqza_file,
)

}
\seealso{
MicrobiomeStat documentation for further details on how to use the returned data object.
}
\author{
Caffery Yang
}
