% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_related.R
\name{mStat_plot_to_grob}
\alias{mStat_plot_to_grob}
\title{Convert supported plot objects to grobs}
\usage{
mStat_plot_to_grob(plot, ...)
}
\arguments{
\item{plot}{Plot-like object.}

\item{...}{Additional arguments forwarded to `as.grob()`.}
}
\value{
A grob object suitable for `grid::grid.draw()`.
}
\description{
Convert supported plot objects to grobs
}
\keyword{internal}
