% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_subset_dist.R
\name{mStat_subset_dist}
\alias{mStat_subset_dist}
\title{Subset Distance Matrix}
\usage{
mStat_subset_dist(dist.obj, samIDs)
}
\arguments{
\item{dist.obj}{A list of distance matrices.}

\item{samIDs}{A vector of sample IDs to subset by. This can be a logical vector, a numeric vector, or a character vector of sample IDs.}
}
\value{
A list of subsetted distance matrices.
}
\description{
This function subsets a list of distance matrices by a specified set of sample IDs.
}
\details{
The function first checks if samIDs is logical or numeric, and if so, converts it to a character vector of sample IDs.
Then, it subsets each distance matrix in dist.obj by the sample IDs.
If a matrix is `NA` or `NULL`, it remains unchanged.
}
\examples{
\dontrun{
# Load required libraries
library(vegan)
data(peerj32.obj)

# Calculate beta diversity
dist.obj <- mStat_calculate_beta_diversity(peerj32.obj, dist.name = c('BC', 'Jaccard'))

# Select sample IDs
sample_ids <- sample(colnames(peerj32.obj$feature.tab), 10)

# Subset distance object by sample IDs
subsetted_dist.obj <- mStat_subset_dist(dist.obj, sample_ids)
}

}
