% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_subset_PC.R
\name{mStat_subset_PC}
\alias{mStat_subset_PC}
\title{Subset Principal Coordinates Analysis (PCoA) Results}
\usage{
mStat_subset_PC(pc.obj, samIDs)
}
\arguments{
\item{pc.obj}{A list of PCoA results, usually calculated using the
\code{\link[MicrobiomeStat]{mStat_calculate_PC}} function.}

\item{samIDs}{A vector of sample IDs to subset by. This can be a logical vector, a numeric vector,
or a character vector of sample IDs.}
}
\value{
A list of subsetted PCoA results.
}
\description{
This function subsets the results of a Principal Coordinates Analysis (PCoA) by specified sample IDs.
}
\details{
The function first checks if samIDs is logical or numeric, and if so, converts it to a character vector of sample IDs.
Then, it subsets each PCoA result in pc.obj by the sample IDs.
If a result is `NA` or `NULL`, it remains unchanged.
}
\examples{
\dontrun{
# Load required libraries
library(vegan)
data(peerj32.obj)

# Calculate PCoA
dist.obj <- mStat_calculate_beta_diversity(peerj32.obj, dist.name = c('BC'))
pc.obj <- mStat_calculate_PC(dist.obj, method = c('mds'), k = 2, dist.name = c('BC'))

# Select sample IDs
sample_ids <- sample(colnames(peerj32.obj$feature.tab), 10)

# Subset PCoA object by sample IDs
subsetted_pc.obj <- mStat_subset_PC(pc.obj, sample_ids)
}

}
