% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_generate_report_pair_sections.R
\name{pair_report_sections}
\alias{pair_report_sections}
\title{Report Section Generators for Pair Analysis}
\description{
Internal functions that generate R Markdown template sections
for the pair (two time points) report.
}
\keyword{internal}
