% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peerj32.obj.R
\docType{data}
\name{peerj32.obj}
\alias{peerj32.obj}
\title{Probiotics Intervention Data from the PeerJ32 Study}
\format{
A MicrobiomeStat Data Object containing the following data matrices as described in detail by Lahti et al. (2013):
}
\source{
Data source: http://dx.doi.org/10.7717/peerj.32
}
\usage{
data(peerj32.obj)
}
\description{
The peerj32.obj dataset contains high-throughput profiling data from 389 human blood serum lipids and 130 intestinal genus-level bacteria from 44 samples (22 subjects from 2 time points; before and after probiotic/placebo intervention). The data set is designed to investigate associations between intestinal bacteria and host lipid metabolism.
}
\references{
Lahti et al. (2013) PeerJ 1:e32 http://dx.doi.org/10.7717/peerj.32
}
\author{
Leo Lahti <microbiome-admin@googlegroups.com>
}
\keyword{datasets}
