% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_utils.R
\name{resolve_params}
\alias{resolve_params}
\title{Resolve All Parameters for Target Function}
\usage{
resolve_params(args, target_func, design_info)
}
\arguments{
\item{args}{List of user-provided arguments}

\item{target_func}{Name of target function}

\item{design_info}{Study design information from detect_study_design()}
}
\value{
List of resolved parameters ready for do.call()
}
\description{
Main parameter resolution function that prepares all parameters for the target function.
}
\keyword{internal}
