% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linda2_optimized.R
\name{should_use_optimized}
\alias{should_use_optimized}
\title{Check if optimized smoothing should be used}
\usage{
should_use_optimized(m, tree.smooth)
}
\description{
Check if optimized smoothing should be used
}
\keyword{internal}
