% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_pairs.obj.R
\docType{data}
\name{subset_pairs.obj}
\alias{subset_pairs.obj}
\title{Paired Sample 16S Sequencing Data from the MiPair Study}
\format{
A MicrobiomeStat Data Object containing the following data matrices as described in detail by Jang H, Koh H, Gu W, Kang B. (2022):
}
\source{
Data source: https://github.com/YJ7599/MiPairGit
}
\usage{
data(subset_pairs.obj)
}
\description{
The subset_pairs.obj dataset contains paired sample 16S sequencing data. This data set is designed for design-based comparative analysis with paired microbiome data.
}
\references{
Jang H, Koh H, Gu W, Kang B. (2022) Integrative web cloud computing and analytics using MiPair for design-based comparative analysis with paired microbiome data. Scientific Reports 12(20465).
}
\author{
Jang H, Koh H, Gu W, Kang B
}
\keyword{datasets}
