% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_T2D.obj.R
\docType{data}
\name{subset_T2D.obj}
\alias{subset_T2D.obj}
\title{Subset T2D Microbiome Data from the HMP2 Study}
\format{
A MicrobiomeStat Data Object with the following components:
}
\usage{
data(subset_T2D.obj)
}
\description{
The subset_T2D.obj dataset contains 16S rRNA data for 79 patients with type 2 diabetes (T2D) over multiple visits, gathered as part of the HMP2 study. The dataset includes a phyloseq object with 12,062 taxa and 2,208 samples.
}
\keyword{datasets}
