% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_alpha_boxplot_long.R
\name{generate_alpha_boxplot_long}
\alias{generate_alpha_boxplot_long}
\title{Generate boxplot for specified alpha diversity index}
\usage{
generate_alpha_boxplot_long(
  data.obj,
  alpha.obj = NULL,
  alpha.name = c("shannon", "simpson", "observed_species", "chao1", "ace", "pielou"),
  subject.var,
  time.var,
  t0.level = NULL,
  ts.levels = NULL,
  group.var = NULL,
  strata.var = NULL,
  base.size = 12,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{alpha.obj}{An optional list containing pre-calculated alpha diversity indices. If NULL (default), alpha diversity indices will be calculated using mStat_calculate_alpha_diversity function from MicrobiomeStat package.}

\item{alpha.name}{The alpha diversity index to be plotted. Supported indices include "shannon", "simpson", "observed_species", "chao1", "ace", and "pielou".}

\item{subject.var}{The variable in the metadata table that represents the subject.}

\item{time.var}{The variable in the metadata table that represents the time.}

\item{t0.level}{The level in the metadata table that represents the initial time point. This parameter is used to determine the starting point for time series data. If NULL, the first unique time point in the data will be used.}

\item{ts.levels}{The levels in the metadata table that represent the subsequent time points.
These parameters are used to determine the time points that follow the initial time point (t0.level) for time series data. If NULL, all unique time points in the data (except for the initial time point) will be used.}

\item{group.var}{An optional variable in the metadata table that represents the grouping factor.}

\item{strata.var}{An optional variable in the metadata table that represents the stratification factor.}

\item{base.size}{The base font size for the plot.}

\item{theme.choice}{A character string indicating the choice of pre-defined ggplot2 theme for the plot. Supported choices include "prism" (default), "classic", "gray", and "bw".}

\item{custom.theme}{An optional custom ggplot2 theme. If provided, this theme will be used instead of the pre-defined themes.}

\item{palette}{An optional color palette for the plot. If not provided, a default color palette will be used. The palette should be a vector of color codes in a format accepted by ggplot2 (e.g., hexadecimal color codes). The number of colors in the palette should be at least as large as the number of groups being plotted.}

\item{pdf}{A boolean indicating whether to save the output as a PDF file.}

\item{file.ann}{A string for annotating the output file name.}

\item{pdf.wid}{The width of the output PDF file. Default is 11.}

\item{pdf.hei}{The height of the output PDF file. Default is 8.5.}

\item{...}{Additional arguments to pass to the plotting function.}
}
\value{
A boxplot displaying the specified alpha diversity index across different groupings and time points, stratified by the specified stratification variable (if provided). The boxplot will be saved as a PDF if `pdf` is set to `TRUE`.
}
\description{
This function generates a boxplot of a specified alpha diversity index across different groupings and time points, with optional stratification. The output can be saved as a PDF.
}
\examples{
\dontrun{
library("HMP2Data")
T2D <- T2D16S()
T2D.obj <- mStat_convert_phyloseq_to_data_obj(T2D.phy)

subset_T2D.obj <- mStat_subset_data(T2D.obj,colnames(T2D.obj$feature.tab
[,colSums(T2D.obj$feature.tab) >= 2000]))
T2D.alpha.obj <- mStat_calculate_alpha_diversity(subset_T2D.obj$feature.tab,"shannon")

generate_alpha_boxplot_long(
  data.obj = subset_T2D.obj,
  alpha.obj = T2D.alpha.obj,
  alpha.name = c("shannon"),
  subject.var = "subject_id",
  time.var = "visit_number",
  t0.level = sort(unique(T2D.obj$meta.dat$visit_number))[1],
  ts.levels = sort(unique(T2D.obj$meta.dat$visit_number))[2:6],
  group.var = "subject_race",
  strata.var = NULL,
  base.size = 20,
  theme.choice = "bw",
  palette = ggsci::pal_npg()(9),
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 20,
  pdf.hei = 8.5)

generate_alpha_boxplot_long(
  data.obj = peerj32.obj,
  alpha.obj = NULL,
  alpha.name = c("simpson"),
  subject.var = "subject",
  time.var = "time",
  t0.level = "1",
  ts.levels = "2",
  group.var = "group",
  strata.var = NULL,
  base.size = 20,
  theme.choice = "bw",
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5)
}
}
