% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_alpha_boxplot_single.R
\name{generate_alpha_boxplot_single}
\alias{generate_alpha_boxplot_single}
\title{Generate boxplot for alpha diversity index at a single time point}
\usage{
generate_alpha_boxplot_single(
  data.obj,
  alpha.obj = NULL,
  alpha.name = c("shannon", "simpson", "observed_species", "chao1", "ace", "pielou"),
  subject.var,
  time.var = NULL,
  t.level = NULL,
  group.var = NULL,
  strata.var = NULL,
  adj.vars = NULL,
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{alpha.obj}{An optional list containing pre-calculated alpha diversity indices. If NULL (default), alpha diversity indices will be calculated using mStat_calculate_alpha_diversity function from MicrobiomeStat package.}

\item{alpha.name}{The alpha diversity index to be plotted. Supported indices include "shannon", "simpson", "observed_species", "chao1", "ace", and "pielou".}

\item{subject.var}{The variable in the metadata table that represents the subject.}

\item{time.var}{The variable in the metadata table that represents the time.}

\item{t.level}{The specific time point to be analyzed.}

\item{group.var}{An optional variable in the metadata table that represents the grouping factor.}

\item{strata.var}{An optional variable in the metadata table that represents the stratification factor.}

\item{adj.vars}{A character vector of variable names to be used for adjustment.}

\item{base.size}{The base font size for the plot. Default is 16.}

\item{theme.choice}{A character string indicating the choice of pre-defined ggplot2 theme for the plot. Supported choices include "prism" (default), "classic", "gray", and "bw".}

\item{custom.theme}{An optional custom ggplot2 theme. If provided, this theme will be used instead of the pre-defined themes.}

\item{palette}{An optional palette to use for the plot. If NULL (default), a pre-defined palette will be used.}

\item{pdf}{A boolean indicating whether to save the output as a PDF file. Default is TRUE.}

\item{file.ann}{A string for annotating the output file name.}

\item{pdf.wid}{The width of the output PDF file. Default is 11.}

\item{pdf.hei}{The height of the output PDF file. Default is 8.5.}

\item{...}{Additional arguments to pass to the plotting function.}
}
\value{
A list of boxplots displaying the specified alpha diversity indices at the specified time point dplyr::across different groupings, stratified by the specified stratification variable (if provided). Each boxplot in the list corresponds to one of the alpha diversity indices specified in `alpha.name`. The boxplots will be saved as PDF files if `pdf` is set to `TRUE`.
}
\description{
This function generates a boxplot of specified alpha diversity indices at a single time point dplyr::across different groupings, with optional stratification. The output can be saved as a PDF.
}
\examples{
\dontrun{
library(vegan)
library(ggh4x)
data(peerj32.obj)

generate_alpha_boxplot_single(
  data.obj = peerj32.obj,
  alpha.obj = NULL,
  alpha.name = c("simpson"),
  subject.var = "subject",
  time.var = "time",
  t.level = "2",
  group.var = "group",
  strata.var = NULL,
  adj.vars = "sex",
  base.size = 16,
  theme.choice = "bw",
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5)

  data("subset_T2D.obj")
generate_alpha_boxplot_single(
  data.obj = subset_T2D.obj,
  alpha.obj = NULL,
  alpha.name = c("shannon"),
  subject.var = "subject_id",
  time.var = "visit_number",
  t.level = "   3",
  group.var = "subject_race",
  strata.var = "subject_gender",
  adj.vars = "sample_body_site",
  base.size = 16,
  theme.choice = "bw",
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 20,
  pdf.hei = 8.5)

}
}
