% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_alpha_change_test_pair.R
\name{generate_alpha_change_test_pair}
\alias{generate_alpha_change_test_pair}
\title{Generate Alpha Diversity Change Test Pair}
\usage{
generate_alpha_change_test_pair(
  data.obj,
  alpha.obj = NULL,
  time.var,
  alpha.name,
  subject.var,
  group.var,
  adj.vars,
  change.base,
  change.func = "lfc"
)
}
\arguments{
\item{data.obj}{A data object containing both the count (OTU table) and the metadata.}

\item{alpha.obj}{A data object containing alpha diversity metrics. If NULL, it will be calculated from the data.obj.}

\item{time.var}{The name of the variable in the data that represents time points.}

\item{alpha.name}{The name of the alpha diversity metric to be tested.}

\item{subject.var}{The name of the variable in the data that represents the subject IDs.}

\item{group.var}{The name of the variable in the data that represents the group IDs.}

\item{adj.vars}{A vector of names of the variables in the data that should be used as covariates in the model.}

\item{change.base}{The name of the baseline time point for calculating changes in alpha diversity. If NULL, the first unique time point in the data will be used.}

\item{change.func}{The function to be used for calculating changes in alpha diversity. If "lfc", the change will be calculated as the log-fold change. If a function, it will be applied directly.}
}
\value{
A list of tables, one for each alpha diversity metric, summarizing the results of the statistical tests.
Each table contains the following columns: Term (the name of the variable in the model), Estimate (the estimated coefficient),
Std.Error (the standard error of the coefficient), Statistic (the t or F statistic), P.Value (the p-value of the test).
}
\description{
This function generates a pair of statistical tests for comparing alpha diversity metrics
at two different time points for a given dataset. The function supports various options
for adjusting the tests and calculating the difference in alpha diversity.
}
\examples{
\dontrun{
library(vegan)
data(peerj32.obj)

generate_alpha_change_test_pair(
data.obj = peerj32.obj,
alpha.obj = NULL,
time.var = "time",
alpha.name = c("shannon", "simpson"),
subject.var = "subject",
group.var = "group",
adj.vars = "sex",
change.base = "1"
)
}
}
