% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_alpha_change_test_pair.R
\name{generate_alpha_change_test_pair}
\alias{generate_alpha_change_test_pair}
\title{Compare alpha diversity between time points}
\usage{
generate_alpha_change_test_pair(
  data.obj,
  alpha.obj = NULL,
  alpha.name = NULL,
  depth = NULL,
  time.var,
  subject.var,
  group.var,
  adj.vars,
  change.base,
  change.func = "lfc"
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{alpha.obj}{An optional list containing pre-calculated alpha diversity indices. If NULL (default), alpha diversity indices will be calculated using mStat_calculate_alpha_diversity function from MicrobiomeStat package.}

\item{alpha.name}{The alpha diversity index to be plotted. Supported indices include "shannon", "simpson", "observed_species", "chao1", "ace", and "pielou". Previously named as `alpha.index`.}

\item{time.var}{Character string specifying the column name in metadata containing
time values for each sample. Required to identify pairs of time
points to calculate changes between.}

\item{subject.var}{Character string specifying the column name in metadata containing
unique subject IDs. Required to pair samples from the same subject
across time points.}

\item{group.var}{Character string specifying the column name in metadata containing
grouping categories. Used as a predictor in the models to test for
differences in changes between groups. Optional, can be NULL.}

\item{adj.vars}{Character vector specifying column names in metadata containing
covariates to adjust for in the linear models. Optional, can be
left NULL if no adjustment is needed.}

\item{change.base}{The name of the baseline time point for calculating changes in alpha diversity. If NULL, the first unique time point in the data will be used.}

\item{change.func}{A method for calculating the change in alpha diversity between two time points.
This can either be the string "lfc" or a custom function.
- If "lfc": The change in alpha diversity is computed as the log-fold change. Specifically, the function calculates the natural logarithm of the ratio of the alpha diversity at the second time point to the first.
- If a function: This function should take two numeric arguments representing the alpha diversity values at two distinct time points. It should return a single numeric value indicating the change. For instance, if you want to compute the absolute difference between the two time points, you could provide a function like `function(a, b) {return a - b}`.
It is crucial that the function is set up to handle the order of the time points correctly, as the first argument will always be the later time point, and the second argument will be the earlier one.}
}
\value{
A list of tables, one for each alpha diversity metric, summarizing the results of the statistical tests.
Each table contains the following columns: Term (the name of the variable in the model), Estimate (the estimated coefficient),
Std.Error (the standard error of the coefficient), Statistic (the t or F statistic), P.Value (the p-value of the test).
}
\description{
This function performs paired tests to compare alpha diversity metrics
between two time points in a microbiome dataset.
}
\details{
For each alpha diversity metric, the difference between the two time points
is calculated for each subject. The difference is used as the response
variable in a linear model, with grouping and adjustment variables as
predictors.

The linear model coefficients are extracted into a results table. If the
grouping variable has multiple levels, ANOVA is performed to test for overall
significance.

Options are provided to customize the alpha diversity difference calculation
(e.g. log-fold change) and adjust for covariates.

In summary, this provides a flexible paired test for analyzing changes in
alpha diversity over time, accounting for groups and covariates.


This function performs the following statistical tests:

- For each alpha diversity metric, a linear model is fitted with the difference in alpha diversity between two time points as the response, and the grouping variable and any adjustment variables as predictors.
The linear model coefficients are extracted and formatted into a results table.

- If the grouping variable has more than 2 levels, ANOVA is performed on the linear model to obtain the overall significance of the grouping variable.
The ANOVA results are also formatted into the results table.

So in summary, the function provides a paired test for the change in alpha diversity between two time points,
with options to adjust for covariates and test for differences between groups. The output contains coefficient tables
summarizing the results of the linear models and ANOVA tests.
}
\examples{
\dontrun{
library(vegan)
data(peerj32.obj)

generate_alpha_change_test_pair(
data.obj = peerj32.obj,
alpha.obj = NULL,
time.var = "time",
alpha.name = c("shannon"),
subject.var = "subject",
group.var = "group",
adj.vars = "sex",
change.base = "1"
)
}
}
