% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_alpha_change_test_pair.R
\name{generate_alpha_change_test_pair}
\alias{generate_alpha_change_test_pair}
\title{Alpha Diversity Association Test}
\usage{
generate_alpha_change_test_pair(
  data.obj,
  alpha.obj = NULL,
  time.var,
  alpha.name,
  subject.var,
  group.var,
  adj.vars,
  change.base,
  change.func = "lfc"
)
}
\arguments{
\item{data.obj}{A list object that includes feature.tab (an OTU table with taxa as rows and samples as columns) and meta.dat (a metadata table with samples as rows and variables as columns).}

\item{alpha.obj}{A pre-calculated alpha diversity object. If NULL, the function will calculate alpha diversity based on data.obj.}

\item{time.var}{A string representing the time variable's name in the metadata. The default is NULL.}

\item{alpha.name}{A character vector with the names of alpha diversity indices to compute. Options include: "shannon", "simpson", "observed_species", "chao1", "ace", and "pielou".}

\item{group.var}{A string representing the group variable's name in the metadata.}

\item{adj.vars}{A character vector with the names of adjustment variables in the metadata.}

\item{change.base}{A logical value indicating whether to change the base level of the group variable. Default is FALSE.}

\item{change.func}{A function or a string indicating the method to calculate change in diversity indices. Default is "lfc" (log fold change).}
}
\value{
A list containing the association tests for each alpha diversity index.
}
\description{
This function implements an association test for multiple alpha diversity measures. It accepts a data.obj as input and returns a list of tests generated for each alpha diversity index.
}
\examples{
generate_alpha_change_test_pair(
data.obj = peerj32.obj,
alpha.obj = NULL,
time.var = "time",
alpha.name = c("shannon", "simpson"),
subject.var = "subject",
group.var = "group",
adj.vars = "sex",
change.base = "1"
)

}
