% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_alpha_spaghettiplot_long.R
\name{generate_alpha_spaghettiplot_long}
\alias{generate_alpha_spaghettiplot_long}
\title{Generate an alpha diversity line plot for longitudinal data}
\usage{
generate_alpha_spaghettiplot_long(
  data.obj,
  alpha.obj = NULL,
  alpha.name = c("shannon", "simpson", "observed_species", "chao1", "ace", "pielou"),
  subject.var,
  time.var,
  t0.level,
  ts.levels,
  group.var = NULL,
  strata.var = NULL,
  base.size = 16,
  palette = NULL,
  theme.choice = "bw",
  custom.theme = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{alpha.name}{The alpha diversity index to be calculated (e.g., "Shannon").}

\item{subject.var}{The name of the subject variable.}

\item{time.var}{The name of the time variable.}

\item{group.var}{The name of the group variable.}

\item{strata.var}{The name of the strata variable (default is NULL).}

\item{pdf}{Logical, whether to save the plot as a PDF (default is TRUE).}

\item{file.ann}{The annotation to be added to the PDF file name (default is NULL).}

\item{...}{Additional arguments passed to ggplot().}
}
\value{
A ggplot object of the alpha diversity line plot.
}
\description{
This function creates a ggplot object of alpha diversity (e.g., Shannon index) line plot for longitudinal data,
showing individual subject trajectories and the mean trajectory for each group.
}
\examples{
\dontrun{
library("HMP2Data")
T2D <- T2D16S()

T2D.obj <- mStat_convert_phyloseq_to_data_obj(T2D)

subset_T2D.obj <- mStat_subset_data(T2D.obj,colnames(T2D.obj$feature.tab
[,colSums(T2D.obj$feature.tab) >= 2000]))
T2D.alpha.obj <- mStat_calculate_alpha_diversity(subset_T2D.obj$feature.tab,"shannon")

generate_alpha_spaghettiplot_long(
  data.obj = subset_T2D.obj,
  alpha.obj = T2D.alpha.obj,
  alpha.name = c("shannon"),
  subject.var = "subject_id",
  time.var = "visit_number",
  t0.level = sort(unique(T2D.obj$meta.dat$visit_number))[1],
  ts.levels = sort(unique(T2D.obj$meta.dat$visit_number))[-1],
  group.var = "subject_race",
  strata.var = NULL,
  theme.choice = "bw",
  palette = ggsci::pal_npg()(9),
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)
print(plot)
}
}
