% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_alpha_test_long.R
\name{generate_alpha_test_long}
\alias{generate_alpha_test_long}
\title{Longitudinal Alpha Diversity Association Test in MicrobiomeStat}
\usage{
generate_alpha_test_long(
  data.obj,
  alpha.obj = NULL,
  time.var,
  t0.level,
  ts.levels,
  alpha.name,
  subject.var,
  group.var,
  adj.vars
)
}
\arguments{
\item{data.obj}{A list object that includes feature.tab (an OTU table with
taxa as rows and samples as columns) and meta.dat (a metadata table with
samples as rows and variables as columns).}

\item{alpha.obj}{A pre-calculated alpha diversity object. If NULL, the
function will calculate alpha diversity based on data.obj.}

\item{time.var}{A string representing the time variable's name in the
metadata. The default is NULL.}

\item{t0.level}{The level in time.var considered as the reference level in
the longitudinal study. Typically the first time point.}

\item{ts.levels}{The levels in time.var that are to be considered in
the analysis, excluding the reference level.}

\item{alpha.name}{A character vector with the names of alpha diversity
indices to compute. Options include: "shannon", "simpson",
"observed_species", "chao1", "ace", and "pielou".}

\item{group.var}{A string representing the group variable's name in the
metadata.}

\item{adj.vars}{A character vector with the names of adjustment variables in
the metadata.}
}
\value{
A list containing the association tests for each alpha diversity
index.
}
\description{
This function, part of the MicrobiomeStat package, implements a method for
performing an association test for multiple alpha diversity measures in
longitudinal data. The test is based on a mixed-effects model, which is more
robust and efficient than previous implementations. The function is especially
suitable for microbiome studies involving repeated measurements over time.
}
\details{
The mixed-effects model includes a temporal component, accounting for the
correlation among repeated measurements on the same subject over time.
The function also accepts additional adjustment variables for greater
flexibility in modeling.

The output is a list of coefficient tables, one for each alpha diversity index.
Each table includes the term, estimate, standard error, t value, and p-value
for each fixed effect in the model.
}
\examples{
library("HMP2Data")
T2D <- T2D16S()
T2D.obj <- mStat_convert_phyloseq_to_data_obj(T2D.phy)

subset_T2D.obj <- mStat_subset_data(T2D.obj,colnames(T2D.obj$feature.tab
[,colSums(T2D.obj$feature.tab) >= 2000]))
alpha_test_results <- generate_alpha_test_long(
data.obj = subset_T2D.obj,
alpha.obj = NULL,
time.var = "visit_number",
t0.level = sort(unique(T2D.obj$meta.dat$visit_number))[1],
ts.levels = sort(unique(T2D.obj$meta.dat$visit_number))[2:6],
alpha.name = c("shannon", "simpson"),
subject.var = "subject_id",
group.var = "subject_race",
adj.vars = "subject_gender"
)

}
