% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_alpha_test_pair.R
\name{generate_alpha_test_pair}
\alias{generate_alpha_test_pair}
\title{Alpha Diversity Association Test}
\usage{
generate_alpha_test_pair(
  data.obj,
  alpha.obj = NULL,
  time.var,
  alpha.name,
  subject.var,
  group.var,
  adj.vars
)
}
\arguments{
\item{data.obj}{A list object that includes feature.tab (an OTU table with
taxa as rows and samples as columns) and meta.dat (a metadata table with
samples as rows and variables as columns).}

\item{alpha.obj}{A pre-calculated alpha diversity object. If NULL, the
function will calculate alpha diversity based on data.obj.}

\item{time.var}{A string representing the time variable's name in the
metadata. The default is NULL.}

\item{alpha.name}{A character vector with the names of alpha diversity
indices to compute. Options include: "shannon", "simpson",
"observed_species", "chao1", "ace", and "pielou".}

\item{subject.var}{A string specifying the subject variable column in the metadata.}

\item{group.var}{A string representing the group variable's name in the
metadata.}

\item{adj.vars}{A character vector with the names of adjustment variables in
the metadata.}
}
\value{
A list containing the association tests for each alpha diversity
index.
}
\description{
This function implements an association test for multiple alpha diversity
measures. The test is based on a mixed-effects model fitted by the `lmer`
function from the `lmerTest` package. The function accepts a data object as
input and returns a list of tests, one for each alpha diversity index.
}
\details{
The mixed-effects model includes the time variable, group variable, and any
additional adjustment variables as fixed effects, and the subject variable as
a random effect.

The output is a list of coefficient tables, one for each alpha diversity
index. Each table includes the term, estimate, standard error, t value, and
p-value for each fixed effect in the model.
}
\examples{
data(peerj32.obj)
alpha_test_results <- generate_alpha_test_pair(
data.obj = peerj32.obj,
alpha.obj = NULL,
time.var = "time",
alpha.name = c("shannon", "simpson"),
subject.var = "subject",
group.var = "group",
adj.vars = "sex"
)

}
