% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_alpha_trend_test_long.R
\name{generate_alpha_trend_test_long}
\alias{generate_alpha_trend_test_long}
\title{Longitudinal Alpha Diversity Volatility Test in MicrobiomeStat}
\usage{
generate_alpha_trend_test_long(
  data.obj,
  alpha.obj = NULL,
  alpha.name,
  time.var,
  subject.var,
  group.var,
  adj.vars = NULL
)
}
\arguments{
\item{data.obj}{A list object that includes feature.tab (an OTU table with
taxa as rows and samples as columns) and meta.dat (a metadata table with
samples as rows and variables as columns).}

\item{alpha.obj}{An object containing precomputed alpha diversity indices,
typically calculated by the function `mStat_calculate_alpha_diversity`.
If NULL (the default), alpha diversity will be calculated from the data.obj using
`mStat_calculate_alpha_diversity`.}

\item{alpha.name}{A string with the name of the alpha diversity index to compute.
Options could include: "shannon", "simpson", "observed_species", "chao1", "ace", and "pielou".}

\item{time.var}{A string representing the time variable's name in the
metadata. The default is NULL.}

\item{subject.var}{A string indicating the variable for subject identifiers.}

\item{group.var}{A string representing the group variable's name in the
metadata.}

\item{adj.vars}{A character vector with the names of adjustment variables in
the metadata, used in fitting the linear model for residuals.}
}
\value{
A summary object containing the results of the linear model testing
the association between alpha diversity volatility and the group variable.
}
\description{
This function, part of the MicrobiomeStat package, calculates the volatility
of alpha diversity measures in longitudinal data and tests the association
between the volatility and a group variable. Volatility is calculated as the mean
of absolute differences between consecutive alpha diversity measures, normalized
by the time difference (mean(abs(alpha_j - alpha_{j-1}) / (t_j - t_{j-1}))).
}
\details{
The function first obtains the residuals by fitting a linear model with alpha
diversity as the response and adjustment variables as predictors. It then calculates
volatility for each subject and tests its association with a group variable using a
linear model. If the group variable specified in `group.var` has more than two levels,
an ANOVA is performed to test the association between alpha diversity volatility
and the group variable.
}
\examples{
\dontrun{
data("subset_T2D.obj")
subset_T2D.obj2 <- subset_T2D.obj
subset_T2D.obj2$meta.dat$visit_number <- as.numeric(subset_T2D.obj2$meta.dat$visit_number)
generate_alpha_trend_test_long(
data.obj = subset_T2D.obj2,
alpha.name = c("shannon","simpson"),
time.var = "visit_number",
subject.var = "subject_id",
group.var = NULL,
adj.vars = "subject_gender"
)
}
}
