% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_beta_change_boxplot_pair.R
\name{generate_beta_change_boxplot_pair}
\alias{generate_beta_change_boxplot_pair}
\title{Compare beta diversity between time points}
\usage{
generate_beta_change_boxplot_pair(
  data.obj = NULL,
  dist.obj = NULL,
  subject.var,
  time.var,
  group.var = NULL,
  strata.var = NULL,
  adj.vars = NULL,
  change.base,
  dist.name = c("BC", "Jaccard"),
  base.size = 16,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{dist.obj}{Distance matrix between samples, usually calculated using
\code{\link[MicrobiomeStat]{mStat_calculate_beta_diversity}} function.
If NULL, beta diversity will be automatically computed from \code{data.obj}
using \code{mStat_calculate_beta_diversity}.}

\item{subject.var}{A string specifying the name of the subject variable}

\item{time.var}{A string specifying the name of the time variable}

\item{group.var}{A string specifying the name of the group variable or NULL (default)}

\item{strata.var}{A string specifying the name of the strata variable or NULL (default)}

\item{adj.vars}{A string specifying the name of the adjustment variable or NULL (default)}

\item{change.base}{A string or numeric value specifying the first time point used for computing the beta diversity change}

\item{dist.name}{A character vector specifying which beta diversity indices to calculate. Supported indices are "BC" (Bray-Curtis), "Jaccard", "UniFrac" (unweighted UniFrac), "GUniFrac" (generalized UniFrac), "WUniFrac" (weighted UniFrac), and "JS" (Jensen-Shannon divergence). If a name is provided but the corresponding object does not exist within dist.obj, it will be computed internally. If the specific index is not supported, an error message will be returned. Default is c('BC', 'Jaccard').}

\item{base.size}{(Optional) Base font size for the plot (default is 16).}

\item{theme.choice}{Character string specifying ggplot2 theme to use.
Options include:
- "prism": ggprism::theme_prism()
- "classic": theme_classic()
- "gray": theme_gray()
- "bw": theme_bw()

Can also take a custom ggplot2 theme object.
Default is "bw".}

\item{custom.theme}{A custom ggplot theme provided as a ggplot2 theme object. This allows users to override the default theme and provide their own theme for plotting. To use a custom theme, first create a theme object with ggplot2::theme(), then pass it to this argument. For example:

```r
my_theme <- ggplot2::theme(
  axis.title = ggplot2::element_text(size=16, color="red"),
  legend.position = "none"
)
```

Then pass `my_theme` to `custom.theme`. Default is NULL, which will use the default theme based on `theme.choice`.}

\item{palette}{(Optional) A palette function or character vector with the colors for the plot.}

\item{pdf}{A logical value indicating whether to save the plot as a PDF file. Default is TRUE}

\item{file.ann}{A character string specifying a custom annotation for the PDF file name or NULL (default)}

\item{pdf.wid}{(Optional) The width of the PDF file if `pdf` is set to `TRUE` (default is 11).}

\item{pdf.hei}{(Optional) The height of the PDF file if `pdf` is set to `TRUE` (default is 8.5).}

\item{...}{Additional parameters passed on to ggsave()}
}
\value{
A named list of ggplot objects visualizing beta diversity change.
The list contains one plot for each distance metric specified in \code{dist.name}.
Each plot shows boxplots of the beta diversity changes, with samples faceted by
the \code{group_var} and \code{strata_var} variables if provided.
}
\description{
This function generates boxplots to visualize changes in beta diversity
between two time points, for different groups and strata.
}
\details{
For each subject, it calculates the beta diversity (distance)
between the sample at baseline time point and the sample at later time point.
This within-subject change in beta diversity over time is used as the outcome.

Boxplots are generated to compare the distribution of beta diversity changes
across groups and strata. Violin plots are also included to show the underlying distribution.

Adjustment for covariates is supported by using adjusted distances.
}
\examples{
\dontrun{
# Load required libraries and example data
library(vegan)
data(peerj32.obj)

dist.obj <- mStat_calculate_beta_diversity(peerj32.obj, dist.name = "BC")
generate_beta_change_boxplot_pair(
  data.obj = peerj32.obj,
  dist.obj = NULL,
  subject.var = "subject",
  time.var = "time",
  group.var = "group",
  strata.var = "sex",
  adj.vars = "sex",
  change.base = "1",
  dist.name = c('BC'),
  base.size = 20,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)
}
}
\seealso{
\code{\link[MicrobiomeStat]{mStat_calculate_beta_diversity}} for creating the distance object.
}
