% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_beta_change_spaghettiplot_long.R
\name{generate_beta_change_spaghettiplot_long}
\alias{generate_beta_change_spaghettiplot_long}
\title{Generate beta diversity change spaghetti plot for longitudinal data}
\usage{
generate_beta_change_spaghettiplot_long(
  data.obj,
  dist.obj = NULL,
  subject.var,
  time.var,
  t0.level = NULL,
  ts.levels = NULL,
  group.var = NULL,
  strata.var = NULL,
  dist.name = c("BC"),
  base.size = 16,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{dist.obj}{(Optional) A distance object generated from distance matrices using 'mStat_calculate_beta_diversity' function on data.obj. If data.obj is not provided, metadata can be retrieved from dist.obj.}

\item{subject.var}{The variable in the metadata table that represents the subject.}

\item{time.var}{The variable in the metadata table that represents the time.}

\item{t0.level}{(Optional) The baseline time point level.}

\item{ts.levels}{(Optional) Time point levels.}

\item{group.var}{(Optional) The variable in the metadata table that represents the grouping factor.}

\item{strata.var}{(Optional) The variable in the metadata table that represents the stratification factor.}

\item{dist.name}{A character vector specifying which beta diversity indices to calculate. Supported indices are "BC" (Bray-Curtis), "Jaccard", "UniFrac" (unweighted UniFrac), "GUniFrac" (generalized UniFrac), "WUniFrac" (weighted UniFrac), and "JS" (Jensen-Shannon divergence). If a name is provided but the corresponding object does not exist within dist.obj, it will be computed internally. If the specific index is not supported, an error message will be returned.}

\item{base.size}{(Optional) Base font size for the plot (default is 16).}

\item{theme.choice}{(Optional) Name of the theme for the plot. Default is "bw". Other options include "plain", "classic", and any other themes compatible with ggplot2.}

\item{custom.theme}{(Optional) A custom ggplot2 theme.}

\item{palette}{(Optional) A palette function or character vector with the colors for the plot.}

\item{pdf}{(Optional) A boolean indicating whether to save the output as a PDF file (default is TRUE).}

\item{file.ann}{(Optional) A string for annotating the output file name.}

\item{pdf.wid}{(Optional) The width of the PDF file if `pdf` is set to `TRUE` (default is 11).}

\item{pdf.hei}{(Optional) The height of the PDF file if `pdf` is set to `TRUE` (default is 8.5).}

\item{...}{(Optional) Additional arguments to pass to the plotting function.}
}
\value{
A spaghetti plot displaying the beta diversity change over time, stratified by the specified grouping and/or strata variables (if provided). The plot will be saved as a PDF if `pdf` is set to `TRUE`.
}
\description{
This function generates a spaghetti plot visualizing the change in beta diversity over time for different groups and strata in longitudinal data. Optionally, a user can specify whether to save the plot as a PDF file.
}
\examples{
library("MicrobiomeStat")
library("tidyverse")
library(vegan)
ecam.phy <- ecam
ecam.obj <- mStat_convert_phyloseq_to_data_obj(ecam.phy)

rownames(ecam.obj$meta.dat) <- ecam.obj$meta.dat$X.SampleID
ecam.obj$meta.dat$subject.id <- substring(ecam.obj$meta.dat$X.SampleID, 1, 10)

dist.obj <- mStat_calculate_beta_diversity(ecam.obj, "BC")
# Add meta_tab information
attr(dist.obj[["BC"]], "labels") <- ecam.obj$meta.dat
# Generate the boxplot pair
generate_beta_change_spaghettiplot_long(
  data.obj = ecam.obj,
  dist.obj = NULL,
  subject.var = "studyid",
  time.var = "month",
  t0.level = NULL,
  ts.levels = NULL,
  group.var = "diet",
  strata.var = NULL,
  dist.name = c("BC"),
  base.size = 20,
  theme.choice = "bw",
  palette = ggsci::pal_npg()(9),
  pdf = TRUE,
  file.ann = "test",
  pdf.wid = 11,
  pdf.hei = 8.5
)
}
