% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_beta_ordination_single.R
\name{generate_beta_ordination_single}
\alias{generate_beta_ordination_single}
\title{Generate beta diversity change spaghetti plot for longitudinal data}
\usage{
generate_beta_ordination_single(
  data.obj = NULL,
  dist.obj = NULL,
  pc.obj = NULL,
  subject.var,
  time.var = NULL,
  t.level = NULL,
  group.var = NULL,
  strata.var = NULL,
  dist.name = c("BC", "Jaccard"),
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{dist.obj}{(Optional) An object storing the pre-calculated distance matrices. If provided, the function will not compute the distances again. This can save a lot of computational time when working with large datasets.}

\item{subject.var}{A character string specifying the name of the variable that indicates the subject in the metadata table.}

\item{time.var}{A character string specifying the name of the variable that indicates the time in the metadata table.}

\item{group.var}{(Optional) A character string specifying the name of the variable that indicates the grouping factor in the metadata table.}

\item{strata.var}{(Optional) A character string specifying the name of the variable that indicates the stratification factor in the metadata table.}

\item{dist.name}{A character vector specifying which beta diversity indices to calculate. Supported indices are "BC" (Bray-Curtis), "Jaccard", "UniFrac" (unweighted UniFrac), "GUniFrac" (generalized UniFrac), "WUniFrac" (weighted UniFrac), and "JS" (Jensen-Shannon divergence). If a name is provided but the corresponding object does not exist within dist.obj, it will be computed internally. If the specific index is not supported, an error message will be returned.}

\item{base.size}{(Optional) A numeric value specifying the base size of the plot elements. The default is 16.}

\item{theme.choice}{(Optional) A character string specifying the name of the ggplot2 theme to be used for the plot. The default is "bw".}

\item{custom.theme}{(Optional) A custom ggplot2 theme. If provided, this theme will be used instead of the theme specified by `theme.choice`.}

\item{palette}{(Optional) A character vector specifying the color palette to be used for the plot. If not provided, a default color palette will be used.}

\item{pdf}{(Optional) A logical value indicating whether to save the plot as a PDF file. The default is TRUE.}

\item{file.ann}{(Optional) A character string to be added as an annotation to the file name of the output PDF file.}

\item{pdf.wid}{(Optional) A numeric value specifying the width of the output PDF file. The default is 11.}

\item{pdf.hei}{(Optional) A numeric value specifying the height of the output PDF file. The default is 8.5.}

\item{...}{(Optional) Additional arguments to pass to the `ggplot()` function.}

\item{t0.level}{(Optional) The baseline time point level. If not provided, the function will infer it from the data.}

\item{ts.levels}{(Optional) A vector specifying the levels of the time points to be considered. If not provided, the function will infer it from the data.}
}
\value{
A PCoA plot displaying the beta diversity ordination, stratified by the specified grouping and/or strata variables (if provided). The plot will be saved as a PDF if `pdf` is set to `TRUE`.
}
\description{
This function generates a spaghetti plot visualizing the change in beta diversity over time. The plot is produced for different groups and strata in the given longitudinal data. The function allows users to customize the appearance of the plot and the data processing parameters. If desired, the generated plot can be saved as a PDF file.
}
\details{
The function is flexible and allows for various modifications, including the choice of distance measure and stratification factor, providing a comprehensive tool for microbiome beta diversity exploration. It integrates well with other MicrobiomeStat functions and takes their output as input.
}
\examples{
# Load required libraries and example data
library(microbiome)
library(tidyverse)
library(vegan)
data(peerj32)
peerj32.obj <- list()
peerj32.phy <- peerj32$phyloseq
peerj32.obj <- mStat_convert_phyloseq_to_data_obj(peerj32.phy)
dist.obj <- mStat_calculate_beta_diversity(peerj32.obj, dist.name = c('BC', 'Jaccard'))
pc.obj <- mStat_calculate_PC(dist.obj, method = c('mds'), k = 2, dist.name = c('BC','Jaccard'))
# Generate the boxplot pair
generate_beta_ordination_single(
  data.obj = peerj32.obj,
  dist.obj = dist.obj,
  pc.obj = pc.obj,
  subject.var = "subject",
  time.var = "time",
  t.level = "2",
  group.var = "group",
  strata.var = "sex",
  dist.name = c("BC"),
  base.size = 20,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = "test",
  pdf.wid = 11,
  pdf.hei = 8.5
)
}
\seealso{
\code{\link[MicrobiomeStat]{mStat_calculate_beta_diversity}} for creating the distance object, \code{\link[MicrobiomeStat]{mStat_calculate_PC}} for computing the principal coordinates, and \code{\link[ggplot2]{geom_point}}, \code{\link[ggplot2]{geom_boxplot}} for the underlying plot functions used, and \code{\link[MicrobiomeStat]{mStat_convert_DGEList_to_data_obj}}, \code{\link[MicrobiomeStat]{mStat_convert_DESeqDataSet_to_data_obj}}, \code{\link[MicrobiomeStat]{mStat_convert_phyloseq_to_data_obj}}, \code{\link[MicrobiomeStat]{mStat_convert_SummarizedExperiment_to_data_obj}}, \code{\link[MicrobiomeStat]{mStat_import_qiime2_as_data_obj}}, \code{\link[MicrobiomeStat]{mStat_import_mothur_as_data_obj}}, \code{\link[MicrobiomeStat]{mStat_import_dada2_as_data_obj}}, \code{\link[MicrobiomeStat]{mStat_import_biom_as_data_obj}} for data conversion.
}
\author{
Caffery Yang \email{cafferychen7850@gmail.com}
}
