% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_beta_ordination_single.R
\name{generate_beta_ordination_single}
\alias{generate_beta_ordination_single}
\title{Generate Beta Diversity Ordination Plot}
\usage{
generate_beta_ordination_single(
  data.obj = NULL,
  dist.obj = NULL,
  pc.obj = NULL,
  subject.var,
  time.var = NULL,
  t.level = NULL,
  group.var = NULL,
  strata.var = NULL,
  dist.name = c("BC", "Jaccard"),
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A data object. Default is NULL.}

\item{dist.obj}{A distance object. Default is NULL.}

\item{pc.obj}{A principal component object. Default is NULL.}

\item{subject.var}{String. Variable to be used as subject.}

\item{time.var}{String. Variable to be used for time. Default is NULL.}

\item{t.level}{Time level to filter the data by. Default is NULL.}

\item{group.var}{String. Variable to be used for grouping. Default is NULL.}

\item{strata.var}{String. Variable to be used for stratification. Default is NULL.}

\item{dist.name}{Character vector. Name of distance(s) to be used. Default is c('BC', 'Jaccard').}

\item{base.size}{Numeric. Base size for plot elements. Default is 16.}

\item{theme.choice}{String. Theme choice for the plot. Default is 'prism'.}

\item{custom.theme}{A custom theme to be used. Default is NULL.}

\item{palette}{A custom color palette to be used. Default is NULL.}

\item{pdf}{Logical. If TRUE, the plots are saved as PDF files. Default is TRUE.}

\item{file.ann}{File annotation. Default is NULL.}

\item{pdf.wid}{Width of the PDF. Default is 11.}

\item{pdf.hei}{Height of the PDF. Default is 8.5.}

\item{...}{Additional arguments to be passed.}
}
\value{
A PCoA plot displaying the beta diversity ordination, stratified by the specified grouping and/or strata variables (if provided). The plot will be saved as a PDF if `pdf` is set to `TRUE`.
}
\description{
This function generates ordination plots (Principal Component Analysis) based on
beta-diversity distances. It also allows for stratification and grouping of samples,
and calculation of distances at a specific time point.
}
\details{
The function is flexible and allows for various modifications, including the choice of distance measure and stratification factor, providing a comprehensive tool for microbiome beta diversity exploration. It integrates well with other MicrobiomeStat functions and takes their output as input.
}
\examples{
# Load required libraries and example data
library(vegan)
data(peerj32.obj)
dist.obj <- mStat_calculate_beta_diversity(peerj32.obj, dist.name = c('BC', 'Jaccard'))
pc.obj <- mStat_calculate_PC(dist.obj, method = c('mds'), k = 2, dist.name = c('BC','Jaccard'))
# Generate the boxplot pair
generate_beta_ordination_single(
  data.obj = peerj32.obj,
  dist.obj = dist.obj,
  pc.obj = pc.obj,
  subject.var = "subject",
  time.var = "time",
  t.level = "2",
  group.var = "group",
  strata.var = "sex",
  dist.name = c("BC"),
  base.size = 20,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = "test",
  pdf.wid = 11,
  pdf.hei = 8.5
)
}
\seealso{
\code{\link[MicrobiomeStat]{mStat_calculate_beta_diversity}} for creating the distance object, \code{\link[MicrobiomeStat]{mStat_calculate_PC}} for computing the principal coordinates, and \code{\link[ggplot2]{geom_point}}, \code{\link[ggplot2]{geom_boxplot}} for the underlying plot functions used, and \code{\link[MicrobiomeStat]{mStat_convert_DGEList_to_data_obj}}, \code{\link[MicrobiomeStat]{mStat_convert_DESeqDataSet_to_data_obj}}, \code{\link[MicrobiomeStat]{mStat_convert_phyloseq_to_data_obj}}, \code{\link[MicrobiomeStat]{mStat_convert_SummarizedExperiment_to_data_obj}}, \code{\link[MicrobiomeStat]{mStat_import_qiime2_as_data_obj}}, \code{\link[MicrobiomeStat]{mStat_import_mothur_as_data_obj}}, \code{\link[MicrobiomeStat]{mStat_import_dada2_as_data_obj}}, \code{\link[MicrobiomeStat]{mStat_import_biom_as_data_obj}} for data conversion.
}
\author{
Caffery Yang \email{cafferychen7850@gmail.com}
}
