% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_beta_pc_boxplot_long.R
\name{generate_beta_pc_boxplot_long}
\alias{generate_beta_pc_boxplot_long}
\title{Create Boxplots for Beta Diversity Metrics}
\usage{
generate_beta_pc_boxplot_long(
  data.obj = NULL,
  dist.obj = NULL,
  pc.obj = NULL,
  pc.ind = c(1, 2),
  subject.var,
  time.var,
  t0.level = NULL,
  ts.levels = NULL,
  group.var = NULL,
  strata.var = NULL,
  adj.vars = NULL,
  dist.name = c("BC"),
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object containing all data.}

\item{dist.obj}{List object containing distance matrices.}

\item{pc.obj}{List object containing Principal Coordinates matrices.}

\item{pc.ind}{Numeric vector indicating which Principal Coordinates to plot.}

\item{subject.var}{Character string indicating the variable for subject identifiers.}

\item{time.var}{Character string indicating the variable for time points.}

\item{t0.level}{Character string indicating the baseline time point.}

\item{ts.levels}{Vector of character strings indicating the time points to include in the plot.}

\item{group.var}{Character string indicating the variable for group identifiers.}

\item{strata.var}{Character string indicating the variable for stratum identifiers.}

\item{adj.vars}{A character vector containing the names of the columns in data.obj$meta.dat to include as covariates in the PERMANOVA analysis. If no covariates are needed, use NULL (default).}

\item{dist.name}{Vector of character strings indicating the distance measures to include in the plot.}

\item{base.size}{Numeric value indicating the base font size for the plot.}

\item{theme.choice}{Character string indicating the ggplot theme to use for the plot.}

\item{custom.theme}{Custom ggplot theme to use.}

\item{palette}{Vector of colors for the plot.}

\item{pdf}{Logical indicating whether to save the plot as a PDF file.}

\item{file.ann}{Optional character string to append to the file name.}

\item{pdf.wid}{Numeric value indicating the width of the PDF file.}

\item{pdf.hei}{Numeric value indicating the height of the PDF file.}

\item{...}{Additional arguments to be passed to the plotting function.}
}
\value{
A list of ggplot objects for each distance measure and Principal Coordinate.
}
\description{
This function generates boxplots of beta diversity metrics for different groups over time.
It allows the use of Principal Coordinates Analysis (PCoA), non-metric multidimensional scaling (NMDS), t-SNE, or UMAP for dimension reduction.
Default method of dimension reduction is PCoA.
}
\examples{
\dontrun{

# Load required libraries and data
library(vegan)

# Example with ecam.obj dataset
data(ecam.obj)
generate_beta_pc_boxplot_long(
  data.obj = ecam.obj,
  dist.obj = NULL,
  pc.obj = NULL,
  subject.var = "studyid",
  time.var = "month",
  t0.level = "0",
  ts.levels = as.character(sort(as.numeric(unique(ecam.obj$meta.dat$month))))[2:4],
  group.var = "diet",
  strata.var = "delivery",
  adj.vars = NULL,
  dist.name = c('BC'),
  base.size = 20,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)

# Example with peerj32.obj dataset
data(peerj32.obj)
generate_beta_pc_boxplot_long(
  data.obj = peerj32.obj,
  dist.obj = NULL,
  pc.obj = NULL,
  subject.var = "subject",
  time.var = "time",
  t0.level = "1",
  ts.levels = "2",
  group.var = "group",
  strata.var = "sex",
  adj.vars = NULL,
  dist.name = c('BC'),
  base.size = 20,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)

}
}
