% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_beta_pc_volatility_test_long.R
\name{generate_beta_pc_volatility_test_long}
\alias{generate_beta_pc_volatility_test_long}
\title{Generate Beta PC Volatility Test in Long Format}
\usage{
generate_beta_pc_volatility_test_long(
  data.obj,
  dist.obj = NULL,
  pc.obj = NULL,
  pc.ind = c(1, 2),
  subject.var,
  time.var,
  group.var = NULL,
  adj.vars = NULL,
  dist.name = c("BC"),
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, containing components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). If dist.obj is provided, data.obj is not required.}

\item{dist.obj}{Distance matrix between samples, usually calculated using
\code{\link[MicrobiomeStat]{mStat_calculate_beta_diversity}} function.
If NULL, beta diversity will be automatically computed from \code{data.obj}
using \code{mStat_calculate_beta_diversity}.}

\item{pc.obj}{Object containing PCoA coordinates from `mStat_calculate_PC()`, default is NULL.}

\item{pc.ind}{A vector indicating the indices of principal coordinates to use (default c(1, 2)).}

\item{subject.var}{The variable in the metadata table that represents the subject.}

\item{time.var}{The variable in the metadata table that represents the time. Ensure that it is coded as numeric.}

\item{group.var}{(Optional) The variable in the metadata table that represents the grouping factor.}

\item{adj.vars}{(Optional) Variables in the metadata table used for adjustment.}

\item{dist.name}{A character vector specifying which beta diversity indices to calculate. Supported indices are "BC" (Bray-Curtis), "Jaccard", "UniFrac" (unweighted UniFrac), "GUniFrac" (generalized UniFrac), "WUniFrac" (weighted UniFrac), and "JS" (Jensen-Shannon divergence). If a name is provided but the corresponding object does not exist within dist.obj, it will be computed internally. If the specific index is not supported, an error message will be returned. Default is c('BC', 'Jaccard').}

\item{...}{(Optional) Additional arguments to pass to internal functions.}
}
\value{
A list of results for each distance measure and selected Principal Coordinate, including coefficients from the mixed-effects models.
}
\description{
This function computes the volatility test for principal coordinates (PC)
of beta diversity. It allows for a variety of distance measures and optional
adjustments. The function relies on a numeric time variable.
}
\examples{
\dontrun{
library(vegan)
data(ecam.obj)
generate_beta_pc_volatility_test_long(
  data.obj = ecam.obj,
  dist.obj = NULL,
  pc.obj = NULL,
  pc.ind = c(1, 2),
  subject.var = "studyid",
  time.var = "month",
  group.var = "diet",
  adj.vars = "delivery",
  dist.name = c('BC')
)
}
}
