% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_beta_test_pair.R
\name{generate_beta_test_pair}
\alias{generate_beta_test_pair}
\title{Perform pairwise beta diversity tests using PERMANOVA}
\usage{
generate_beta_test_pair(
  data.obj,
  dist.obj = NULL,
  time.var,
  subject.var,
  group.var,
  adj.vars = NULL,
  dist.name = c("BC", "Jaccard", "UniFrac", "GUniFrac", "WUniFrac", "JS")
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, containing microbiome data and metadata.}

\item{dist.obj}{A distance object, if NULL it will be calculated based on data.obj and dist.name. This distance object should be a dissimilarity matrix or an object that can be coerced to a dissimilarity matrix.}

\item{time.var}{The name of the column in data.obj$meta.dat that contains the time variable for subsetting data, if applicable. This should be a character string representing the column name.}

\item{subject.var}{The name of the column in data.obj$meta.dat that contains the subject variable for the PERMANOVA tests. This should be a character string representing the column name.}

\item{group.var}{The name of the column in data.obj$meta.dat that contains the grouping variable for the PERMANOVA tests. This should be a character string representing the column name.}

\item{adj.vars}{A character vector containing the names of the columns in data.obj$meta.dat to include as covariates in the PERMANOVA analysis. If no covariates are needed, use NULL (default).}

\item{dist.name}{A character vector specifying which beta diversity indices to calculate and test (e.g., 'BC', 'Jaccard', 'UniFrac', 'GUniFrac', 'WUniFrac', 'JS'). Default is c('BC', 'Jaccard', 'UniFrac', 'GUniFrac', 'WUniFrac', 'JS').}
}
\value{
A list containing the PERMANOVA results for each beta diversity index. The list includes two elements: "p.tab" - a table of p-values for the PERMANOVA tests across all indices, and "aov.tab" - a table containing detailed PERMANOVA results for each index. The p.tab and aov.tab tables include columns for the terms in the PERMANOVA model, the degrees of freedom, sums of squares, mean squares, F statistics, R-squared values, and p-values.
}
\description{
This function calculates various beta diversity indices for a microbiome dataset and performs pairwise PERMANOVA tests on them to assess the effect of a grouping variable and any additional covariates on the beta diversity. It returns the PERMANOVA results for each of the specified beta diversity indices.
}
\examples{
\dontrun{
library(vegan)
library(GUniFrac)
library(microbiome)
library(ape)
library(philentropy)
library(MicrobiomeStat)

# Load example data
data("peerj32")

peerj32.obj <- mStat_convert_phyloseq_to_data_obj(peerj32$phyloseq)

dist.obj <- mStat_calculate_beta_diversity(peerj32.obj, dist.name = c('BC', 'Jaccard'))

# Perform pairwise beta diversity tests using PERMANOVA
beta_test_pair_results <- generate_beta_test_pair(
  data.obj = peerj32.obj,
  dist.obj = NULL,
  time.var = "time",
  subject.var = "subject",
  group.var = "group",
  adj.vars = c("sex"),
  dist.name = c('BC', 'Jaccard')
)
}

}
\seealso{
\code{\link[GUniFrac]{PermanovaG2}}
\code{\link[microbiome]{peerj32}}
\code{\link[vegan]{adonis2}}
}
