% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_beta_volatility_test_long.R
\name{generate_beta_volatility_test_long}
\alias{generate_beta_volatility_test_long}
\title{Generate Beta Diversity Volatility Test for Longitudinal Data}
\usage{
generate_beta_volatility_test_long(
  data.obj,
  dist.obj = NULL,
  subject.var,
  time.var,
  group.var = NULL,
  adj.vars = NULL,
  dist.name = c("BC"),
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, containing components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). If `dist.obj` is provided, `data.obj` is not required.}

\item{dist.obj}{(Optional) A distance object generated from distance matrices using the 'mStat_calculate_beta_diversity' function on data.obj. If not provided, it will be computed internally.}

\item{subject.var}{The variable in the metadata table that represents the subject.}

\item{time.var}{The variable in the metadata table that represents the time. Ensure that it is coded as numeric.}

\item{group.var}{(Optional) The variable in the metadata table that represents the grouping factor.}

\item{adj.vars}{(Optional) Variables in the metadata table used for adjustment.}

\item{dist.name}{A character vector specifying which beta diversity indices to calculate. Default is "BC" (Bray-Curtis).}

\item{...}{(Optional) Additional arguments to pass to internal functions.}
}
\value{
A list containing the result of the volatility test for each specified beta diversity index. The result includes a tibble with the coefficients extracted from the linear model fitted for each distance, and an ANOVA table if the group variable is multi-categorical.
}
\description{
This function computes a volatility test for longitudinal data on beta diversity. It considers subject, time, optional grouping and adjusting variables, and computes the volatility in beta diversity over time for different groups. The volatility test relies on a numeric time variable.
}
\details{
The function starts by validating the input data, processing the time variable, and calculating the beta diversity if necessary. Adjustments are made based on the provided adjusting variables. The volatility of the beta diversity is computed for each subject, and linear models are fitted to test the association between volatility and the specified group variable. The coefficients and ANOVA results are extracted and returned for each beta diversity index specified.
}
\note{
A warning message will be displayed to ensure that the time variable is coded as numeric. Non-numeric coding may lead to issues in the volatility test computation.
}
\examples{
data(ecam.obj)
ecam.obj$meta.dat$month <- as.numeric(ecam.obj$meta.dat$month)
generate_beta_volatility_test_long(
  data.obj = ecam.obj,
  dist.obj = NULL,
  subject.var = "studyid",
  time.var = "month",
  t0.level = NULL,
  ts.levels = NULL,
  group.var = "diet",
  adj.vars = NULL,
  dist.name = c("BC", "Jaccard")
)
}
