% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_association_test_long.R
\name{generate_taxa_association_volcano_long}
\alias{generate_taxa_association_volcano_long}
\title{Generate Volcano Plots for Taxa Association Test}
\usage{
generate_taxa_association_volcano_long(
  data.obj,
  group.var = NULL,
  test.list,
  feature.sig.level = 0.1,
  feature.mt.method = "fdr"
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{group.var}{The grouping variable tested, found in metadata}

\item{test.list}{The list of test results returned by generate_taxa_trend_test_long}

\item{feature.sig.level}{The significance level cutoff for highlighting taxa}

\item{feature.mt.method}{Multiple testing correction method, "fdr" or "none"}
}
\value{
A list of ggplot objects of volcano plots for each taxonomic level
}
\description{
Generate Volcano Plots for Taxa Association Test
}
\examples{
# Generate test results
# test_list <- generate_taxa_association_test_long(...)
# Create volcano plots
# volcano_plots <- generate_taxa_association_volcano_long(
#   data.obj, group.var, test_list,
#   feature.sig.level = 0.05, feature.mt.method = "fdr"
# )

}
