% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_barplot_long.R
\name{generate_taxa_barplot_long}
\alias{generate_taxa_barplot_long}
\title{Generate Taxa Barplots for Longitudinal Data}
\usage{
generate_taxa_barplot_long(
  data.obj,
  subject.var,
  time.var,
  group.var = NULL,
  strata.var = NULL,
  feature.level = "original",
  feature.dat.type = c("count", "proportion", "other"),
  feature.number = 20,
  t0.level = NULL,
  ts.levels = NULL,
  base.size = 10,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object containing the input data.}

\item{subject.var}{A string indicating the variable for subject identifiers.}

\item{time.var}{A string indicating the variable for time points.}

\item{group.var}{A string indicating the variable for group identifiers. Default is NULL.}

\item{strata.var}{A string indicating the variable for stratum identifiers. Default is NULL.}

\item{feature.level}{A string indicating the taxonomic level to plot.}

\item{feature.dat.type}{A string indicating the type of data in the input object.}

\item{feature.number}{A numeric value indicating the number of features to include in the plot.}

\item{t0.level}{A string indicating the reference time point. Default is NULL.}

\item{ts.levels}{A vector indicating the time points to include in the analysis. Default is NULL.}

\item{base.size}{A numeric value indicating the base font size for the plot.}

\item{theme.choice}{A string indicating the ggplot theme to use for the plot.}

\item{custom.theme}{A custom ggplot theme if the user wants to apply it.}

\item{palette}{A character vector specifying the color palette. Default is NULL.}

\item{pdf}{A logical value indicating whether to save the plot as a PDF. Default is TRUE.}

\item{file.ann}{A string for additional annotation to the file name. Default is NULL.}

\item{pdf.wid}{A numeric value specifying the width of the PDF file. Default is 11.}

\item{pdf.hei}{A numeric value specifying the height of the PDF file. Default is 8.5.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
A ggplot object of the taxa barplot.
}
\description{
This function generates barplots to visualize the taxonomic composition of samples for longitudinal data.
It also provides options for grouping and stratifying data.
}
\details{
This function generates a barplot of taxa proportions for longitudinal data.
The barplot can be stratified by a group variable and/or other variables.
It also allows for different taxonomic levels to be used and a specific number of features to be included in the plot.
The function also has options to customize the size, theme, and color palette of the plot, and to save the plot as a PDF.
}
\examples{
\dontrun{
library(ggh4x)
library(vegan)
data(ecam.obj)
plot_list_all <- generate_taxa_barplot_long(
  data.obj = ecam.obj,
  subject.var = "studyid",
  time.var = "month",
  group.var = "delivery",
  strata.var = "diet",
  feature.level = "Family",
  feature.dat.type = "proportion",
  feature.number = 10,
  t0.level = NULL,
  ts.levels = NULL,
  theme.choice = "bw",
  palette = NULL,
  pdf = TRUE,
  file.ann = "test"
)
}
}
