% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_change_boxplot_pair.R
\name{generate_taxa_change_boxplot_pair}
\alias{generate_taxa_change_boxplot_pair}
\title{Generate Individual Taxa Change Boxplot Pair}
\usage{
generate_taxa_change_boxplot_pair(
  data.obj,
  subject.var,
  time.var,
  group.var = NULL,
  strata.var = NULL,
  change.base = NULL,
  change.func = "relative difference",
  feature.level = NULL,
  feature.dat.type = c("count", "proportion", "other"),
  features.plot = NULL,
  top.k.plot = NULL,
  top.k.func = NULL,
  prev.filter,
  abund.filter,
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A data object containing the data to be plotted.}

\item{subject.var}{A character string specifying the subject variable.}

\item{time.var}{A character string specifying the time variable.}

\item{group.var}{A character string specifying the group variable. Default is NULL.}

\item{strata.var}{A character string specifying the strata variable. Default is NULL.}

\item{change.base}{The time point to be used as the base for calculating change in abundance.}

\item{change.func}{The function to be used for calculating change in abundance. Default is "relative difference".}

\item{feature.level}{The feature level at which to plot the data.}

\item{feature.dat.type}{The type of the feature data. Can be "count", "proportion", or "other". Default is "count".}

\item{features.plot}{A character vector specifying the features to be plotted. Default is NULL.}

\item{top.k.plot}{The top k features to be plotted. Default is NULL.}

\item{top.k.func}{The function to be used for selecting the top k features. Default is NULL.}

\item{prev.filter}{The prevalence filter to be applied to the data.}

\item{abund.filter}{The abundance filter to be applied to the data.}

\item{base.size}{The base size for the plot. Default is 16.}

\item{theme.choice}{The theme choice for the plot. Default is "prism".}

\item{custom.theme}{An optional custom theme for the plot. Default is NULL.}

\item{palette}{A character vector specifying the color palette for the plot. Default is NULL.}

\item{pdf}{A logical value indicating whether to save the plot as a PDF. Default is TRUE.}

\item{file.ann}{An optional character string to be appended to the file name of the PDF. Default is NULL.}

\item{pdf.wid}{The width of the PDF. Default is 11.}

\item{pdf.hei}{The height of the PDF. Default is 8.5.}

\item{...}{Additional arguments passed to the underlying functions.}
}
\value{
A list of ggplot objects, each of which is a boxplot visualizing the changes in abundance of individual taxa over time.
}
\description{
This function generates boxplot pairs visualizing the changes in abundance of individual taxa over time.
The boxplots show the change in abundance for each taxon for different groups and strata.
It also allows for a prevalence and abundance filter to be applied to the data, and can optionally save the plots as a PDF.
}
\examples{
# Load required libraries and data
library(vegan)
library(ggh4x)
data(peerj32.obj)

# Generate the boxplot pair
generate_taxa_change_boxplot_pair(
  data.obj = peerj32.obj,
  subject.var = "subject",
  time.var = "time",
  group.var = "group",
  strata.var = NULL,
  change.base = "1",
  change.func = "lfc",
  feature.level = c("Family"),
  feature.dat.type = "count",
  features.plot = NULL,
  top.k.plot = 6,
  top.k.func = "sd",
  prev.filter = 0.1,
  abund.filter = 0.01,
  base.size = 20,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = "test",
  pdf.wid = 11,
  pdf.hei = 8.5
)
}
