% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_change_heatmap_long.R
\name{generate_taxa_change_heatmap_long}
\alias{generate_taxa_change_heatmap_long}
\title{Generate Taxonomic Change Heatmap Long}
\usage{
generate_taxa_change_heatmap_long(
  data.obj,
  subject.var,
  time.var,
  t0.level = NULL,
  ts.levels = NULL,
  group.var = NULL,
  strata.var = NULL,
  feature.level,
  feature.dat.type = c("count", "proportion", "other"),
  features.plot = NULL,
  top.k.plot = NULL,
  top.k.func = NULL,
  change.func = "relative difference",
  prev.filter = 1e-08,
  abund.filter = 0.01,
  base.size = 10,
  palette = NULL,
  cluster.rows = NULL,
  cluster.cols = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list).}

\item{subject.var}{A character string specifying the subject variable in the metadata.}

\item{time.var}{A character string specifying the time variable in the metadata.}

\item{t0.level}{The base level for time points in longitudinal data.}

\item{ts.levels}{The levels for time points in longitudinal data.}

\item{group.var}{A character string specifying the grouping variable in the metadata. Default is NULL.}

\item{strata.var}{(Optional) A character string specifying the stratification variable in the metadata. Default is NULL.}

\item{feature.level}{A character string defining the taxonomic level to analyze ('Phylum', 'Family', or 'Genus').}

\item{feature.dat.type}{A character string specifying the type of the data in feature.dat. Options are "count", "proportion", or "other".}

\item{features.plot}{A character vector specifying the taxa to be plotted. If NULL (default), the top k taxa by mean abundance will be plotted.}

\item{top.k.plot}{A numeric value specifying the number of top taxa to be plotted if features.plot is NULL. If NULL (default), all taxa will be plotted.}

\item{top.k.func}{A function to compute the top k taxa if features.plot is NULL. If NULL (default), the mean function will be used.}

\item{change.func}{A function or character string specifying the method for computing the change. Default is "difference".}

\item{prev.filter}{A numeric value defining the prevalence threshold to filter taxa, between 0 and 1.}

\item{abund.filter}{A numeric value defining the abundance threshold to filter taxa.}

\item{base.size}{Base font size for the generated plots.}

\item{palette}{Color palette used for the plots.}

\item{cluster.rows}{A logical variable indicating if rows should be clustered. Default is TRUE.}

\item{cluster.cols}{A logical variable indicating if columns should be clustered. Default is FALSE.}

\item{pdf}{If TRUE, save the plot as a PDF file (default: TRUE)}

\item{file.ann}{(Optional) A character string specifying a file annotation to include in the generated PDF file's name.}

\item{pdf.wid}{Width of the PDF plots.}

\item{pdf.hei}{Height of the PDF plots.}

\item{...}{Additional arguments to be passed}
}
\value{
If the `pdf` parameter is set to TRUE, the function will save a PDF file and return the pheatmap::pheatmap plot. If `pdf` is set to FALSE, the function will return the pheatmap plot without creating a PDF file.

An object of class pheatmap, the generated heatmap plot
}
\description{
This function performs hierarchical clustering on microbiome data based on grouping
variables and strata variables in sample metadata and generates stacked heatmaps
using the “pheatmap” package. It can also save the resulting heatmap as a PDF file.
}
\examples{
\dontrun{
library(pheatmap)
data(ecam.obj)

generate_taxa_change_heatmap_long(
  data.obj = ecam.obj,
  subject.var = "studyid",
  time.var = "month_num",
  t0.level = NULL,
  ts.levels = NULL,
  group.var = "antiexposedall",
  strata.var = "diet",
  feature.level = c("Family"),
  feature.dat.type = "proportion",
  features.plot = NULL,
  top.k.plot = 10,
  top.k.func = "sd",
  change.func = "lfc",
  palette = NULL,
  prev.filter = 0.01,
  abund.filter = 0.01,
  pdf = TRUE,
  file.ann = NULL
)
}

}
\seealso{
\code{\link{pheatmap}}
}
