% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_change_test_pair.R
\name{generate_taxa_change_test_pair}
\alias{generate_taxa_change_test_pair}
\title{Compute and Analyze Taxa Changes Over Time}
\usage{
generate_taxa_change_test_pair(
  data.obj,
  subject.var,
  time.var = NULL,
  group.var,
  adj.vars,
  change.base,
  change.func,
  feature.level,
  prev.filter = 0,
  abund.filter = 0,
  feature.dat.type = c("count", "proportion", "other"),
  ...
)
}
\arguments{
\item{data.obj}{A list object in MicrobiomeStat format.}

\item{subject.var}{The name of the subject variable column in the metadata.}

\item{time.var}{The name of the time variable column in the metadata (optional).}

\item{group.var}{The name of the grouping variable column for linear modeling in the metadata.}

\item{adj.vars}{Names of additional variables to be used as covariates in the analysis.}

\item{change.base}{The baseline time point for detecting changes in taxa. If NULL, the first unique value from the time.var column will be used (optional).}

\item{change.func}{The function to be used for calculating changes. Options include: "lfc" (log fold change), "relative difference", and a user-defined function (default is "lfc").}

\item{feature.level}{The taxonomic level at which to perform the analysis.}

\item{prev.filter}{A numeric value indicating the minimum prevalence filter for the taxa (default is 0).}

\item{abund.filter}{A numeric value indicating the minimum abundance filter for the taxa (default is 0).}

\item{feature.dat.type}{Type of the feature data. Options include: "count", "proportion", and "other" (default is "count").}

\item{...}{Additional parameters to be passed to the ZicoSeq function.}
}
\value{
A named list where each element corresponds to a feature level and contains a dataframe with the calculated taxa changes, their corresponding p-values, and other statistics from the linear model.
}
\description{
This function from the MicrobiomeStat package computes the taxa changes over time for different groups. The changes are calculated based on a specified function. The function also performs linear modelling on these changes using a specified group variable and covariates, and returns a report summarizing these analyses.
If no time variable is provided, the function will select the first unique value from the metadata as the baseline for comparison.
}
\examples{
\dontrun{
library(vegan)
library(GUniFrac)
library(microbiome)
library(ape)
library(philentropy)
library(MicrobiomeStat)

# Load example data
data("peerj32")
peerj32.obj <- mStat_convert_phyloseq_to_data_obj(peerj32$phyloseq)

generate_taxa_change_test_pair(
  data.obj = peerj32.obj,
  subject.var = "subject",
  time.var = "time",
  group.var = "group",
  adj.vars = c("sex"),
  change.base = "1",
  change.func = "lfc",
  feature.level = "original",
  prev.filter = 0.01,
  abund.filter = 0.01,
  feature.dat.type = "count"
)
}

}
