% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_indiv_change_scatterplot_pair.R
\name{generate_taxa_indiv_change_scatterplot_pair}
\alias{generate_taxa_indiv_change_scatterplot_pair}
\title{Generate Individual Change Scatterplot Pairs for Taxonomic Composition Data}
\usage{
generate_taxa_indiv_change_scatterplot_pair(
  data.obj,
  subject.var,
  time.var,
  group.var = NULL,
  strata.var = NULL,
  change.base = NULL,
  change.func = "difference",
  feature.level = NULL,
  features.plot = NULL,
  feature.dat.type = c("count", "proportion", "other"),
  top.k.plot = NULL,
  top.k.func = NULL,
  prev.filter = 0.1,
  abund.filter = 0.1,
  base.size = 16,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A list object containing the input data.}

\item{subject.var}{A string indicating the variable for subject identifiers.}

\item{time.var}{A string indicating the variable for time points.}

\item{group.var}{Optional string specifying the variable for groups.}

\item{strata.var}{Optional string specifying the variable for strata.}

\item{change.base}{A string indicating the base time point for change computation.}

\item{change.func}{A string or function to compute the change in abundance. If a string, it should be one of "difference", "relative difference", or "lfc" (log fold change). If a function, it should take two arguments representing the abundances at two time points and return a numeric value indicating the change.}

\item{feature.level}{A string indicating the taxonomic level to plot.}

\item{features.plot}{A character vector of features to include in the plot. If NULL, top features will be selected based on `top.k.plot` and `top.k.func`.}

\item{feature.dat.type}{A string indicating the type of data in the input object. Options are "count", "proportion", "other".}

\item{top.k.plot}{An integer indicating the top K features to plot based on the function specified in `top.k.func`.}

\item{top.k.func}{A function to determine the top K features to plot.}

\item{prev.filter}{A numeric value indicating the minimum prevalence for a feature to be included in the plot.}

\item{abund.filter}{A numeric value indicating the minimum abundance for a feature to be included in the plot.}

\item{base.size}{A numeric value specifying the base font size for the plot.}

\item{theme.choice}{A string specifying the theme of the plot. Default is "bw".}

\item{custom.theme}{A custom ggplot theme if the user wants to apply it. Default is NULL.}

\item{palette}{A character vector specifying the color palette. Default is NULL.}

\item{pdf}{A logical value indicating whether to save the plot as a PDF. Default is TRUE.}

\item{file.ann}{A string for additional annotation to the file name. Default is NULL.}

\item{pdf.wid}{A numeric value specifying the width of the PDF file. Default is 11.}

\item{pdf.hei}{A numeric value specifying the height of the PDF file. Default is 8.5.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
A list of ggplot objects, one for each taxonomic level.
}
\description{
This function generates scatterplots to visualize the change in taxonomic composition of samples between two time points in a longitudinal study.
It also provides options for grouping and stratifying data, and selecting the top k features based on a user-defined function.
}
\details{
This function generates a scatterplot of the change in taxa abundances between two time points in a longitudinal study.
The scatterplot can be stratified by a group variable and/or other variables.
It also allows for different taxonomic levels to be used and a specific number of features to be included in the plot.
The function also has options to customize the size, theme, and color palette of the plot, and to save the plot as a PDF.
}
\examples{
\dontrun{
 library(vegan)
 library(tidyverse)
 data(peerj32.obj)

 # Generate the scatterplot pairs
 plot_list <- generate_taxa_indiv_change_scatterplot_pair(
   data.obj = peerj32.obj,
   subject.var = "subject",
   time.var = "time",
   group.var = "cons",
   strata.var = "sex",
   change.base = "1",
   taxa.level = "Phylum",
   top.k.plot = 3,
   prev.filter = 0.1,
   abund.filter = 0.01
 )
}

}
