% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_spaghettiplot_long.R
\name{generate_taxa_spaghettiplot_long}
\alias{generate_taxa_spaghettiplot_long}
\title{Generate longitudinal spaghettiplot plots of taxonomic composition}
\usage{
generate_taxa_spaghettiplot_long(
  data.obj,
  subject.var,
  time.var,
  t0.level = NULL,
  ts.levels = NULL,
  group.var = NULL,
  strata.var = NULL,
  feature.level = NULL,
  feature.dat.type = c("count", "proportion", "other"),
  features.plot = NULL,
  top.k.plot = NULL,
  top.k.func = NULL,
  prev.filter = 0.01,
  abund.filter = 0.01,
  base.size = 16,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)
}
\arguments{
\item{data.obj}{A data object created by mStat_convert_phyloseq_to_data_obj.}

\item{subject.var}{A character string defining the subject variable in meta_tab.}

\item{time.var}{A character string defining the time variable in meta_tab.}

\item{t0.level}{The base level for time points in longitudinal data.}

\item{ts.levels}{The levels for time points in longitudinal data.}

\item{group.var}{A character string defining the group variable in meta_tab used for sorting and facetting.}

\item{strata.var}{(Optional) A character string defining strata variable in meta_tab used for sorting and facetting.}

\item{feature.level}{A character string specifying the taxonomic level for the plot.}

\item{feature.dat.type}{Character. Specifies the type of the data in feature.dat. Options are "count" (default) or "proportion".}

\item{features.plot}{A character vector specifying the taxa to be plotted. If NULL (default), the top k taxa by mean abundance will be plotted.}

\item{top.k.plot}{A numeric value specifying the number of top taxa to be plotted if features.plot is NULL. If NULL (default), all taxa will be plotted.}

\item{top.k.func}{A function to compute the top k taxa if features.plot is NULL. If NULL (default), the mean function will be used.}

\item{prev.filter}{A numeric value defining the prevalence threshold to filter taxa, between 0 and 1.}

\item{abund.filter}{A numeric value defining the abundance threshold to filter taxa.}

\item{base.size}{Base font size for the generated plots.}

\item{theme.choice}{Plot theme choice (default: "bw").}

\item{custom.theme}{Custom ggplot2 theme (optional).}

\item{palette}{Color palette used for the plots.}

\item{pdf}{A logical value. If TRUE (default), saves the plot as a PDF file. If FALSE, the plot will be displayed interactively without creating a PDF.}

\item{file.ann}{(Optional) A character string specifying a file annotation to include in the generated PDF file's name.}

\item{pdf.wid}{Width of the PDF plots.}

\item{pdf.hei}{Height of the PDF plots.}

\item{...}{Additional parameters to be passed.}
}
\value{
A ggplot object containing the longitudinal line plot.
}
\description{
This function generates a spaghettiplot plot for the taxonomic composition of microbiome data over time.
}
\examples{
\dontrun{
data(ecam.obj)
plot_list_all <- generate_taxa_spaghettiplot_long(
  data.obj = ecam.obj,
  subject.var = "studyid",
  time.var = "month",
  t0.level = NULL,
  ts.levels = NULL,
  group.var = "diet",
  strata.var = "antiexposedall",
  feature.level = c("Phylum"),
  features.plot = NULL,
  feature.dat.type = "proportion",
  top.k.plot = 5,
  top.k.func = "mean",
  prev.filter = 0,
  abund.filter = 0,
  base.size = 16,
  theme.choice = "classic",
  palette = NULL,
  pdf = TRUE,
  file.ann = "test"
)
}
}
