% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_test_pair.R
\name{generate_taxa_test_pair}
\alias{generate_taxa_test_pair}
\title{Generate Taxa Test Pair}
\usage{
generate_taxa_test_pair(
  data.obj,
  subject.var,
  time.var = NULL,
  group.var,
  adj.vars,
  feature.level,
  prev.filter = 0,
  abund.filter = 0,
  feature.dat.type = c("count", "proportion"),
  ...
)
}
\arguments{
\item{data.obj}{A list object in MicrobiomeStat format.}

\item{subject.var}{A string that specifies the name of the subject variable column in the metadata.}

\item{time.var}{A string that specifies the name of the time variable column in the metadata. If not provided, it's NULL by default.}

\item{group.var}{A string that specifies the name of the grouping variable column in the metadata for linear modelling.}

\item{adj.vars}{A vector of strings that specify the names of additional variables to be used as covariates in the analysis.}

\item{feature.level}{A string specifying the taxonomic level at which to perform the analysis.}

\item{prev.filter}{A numeric value specifying the minimum prevalence filter for the taxa. It's set to 0 by default.}

\item{abund.filter}{A numeric value specifying the minimum abundance filter for the taxa. It's set to 0 by default.}

\item{feature.dat.type}{A string that specifies the type of the feature data. Options include: "count", "proportion". "count" is the default.}

\item{...}{Additional parameters to be passed to the function.}
}
\value{
A named list where each element corresponds to a feature level. Each list element contains a dataframe with the calculated taxa changes, their corresponding p-values,
and other statistics obtained from the linear model.
}
\description{
This function takes as input a MicrobiomeStat data object and various specifications regarding the variables,
and returns an analysis of changes in taxa over time across different groups. It also performs linear modelling on these changes,
utilizing specified group variables and covariates.
}
\examples{
\dontrun{
library(vegan)
library(GUniFrac)
library(microbiome)
library(ape)
library(philentropy)
library(MicrobiomeStat)
library(tidyverse)

# Load example data
data("peerj32")
peerj32.obj <- mStat_convert_phyloseq_to_data_obj(peerj32$phyloseq)

results <- generate_taxa_test_pair(
  data.obj = peerj32.obj,
  subject.var = "subject",
  time.var = "time",
  group.var = "group",
  adj.vars = c("sex"),
  feature.level = "Family",
  prev.filter = 0,
  abund.filter = 0,
  feature.dat.type = "count"
)

}

}
