% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_test_pair.R
\name{generate_taxa_test_pair}
\alias{generate_taxa_test_pair}
\title{Generate Taxa Test Pair}
\usage{
generate_taxa_test_pair(
  data.obj,
  subject.var,
  time.var = NULL,
  group.var,
  adj.vars,
  feature.level,
  prev.filter = 0,
  abund.filter = 0,
  feature.dat.type = c("count", "proportion"),
  feature.sig.level = 0.1,
  feature.mt.method = "fdr",
  ...
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{subject.var}{A string that specifies the name of the subject variable column in the metadata.}

\item{time.var}{A string that specifies the name of the time variable column in the metadata. If not provided, it's NULL by default.}

\item{group.var}{A string that specifies the name of the grouping variable column in the metadata for linear modelling.}

\item{adj.vars}{A vector of strings that specify the names of additional variables to be used as covariates in the analysis.}

\item{feature.level}{The column name in the feature annotation matrix (feature.ann) of data.obj
to use for summarization and plotting. This can be the taxonomic level like "Phylum", or any other
annotation columns like "Genus" or "OTU_ID". Should be a character vector specifying one or more
column names in feature.ann. Multiple columns can be provided, and data will be plotted separately
for each column. Default is NULL, which defaults to all columns in feature.ann if `features.plot`
is also NULL.}

\item{prev.filter}{Numeric value specifying the minimum prevalence threshold for filtering
taxa before analysis. Taxa with prevalence below this value will be removed.
Prevalence is calculated as the proportion of samples where the taxon is present.
Default 0 removes no taxa by prevalence filtering.}

\item{abund.filter}{Numeric value specifying the minimum abundance threshold for filtering
taxa before analysis. Taxa with mean abundance below this value will be removed.
Abundance refers to counts or proportions depending on \code{feature.dat.type}.
Default 0 removes no taxa by abundance filtering.}

\item{feature.dat.type}{The type of the feature data, which determines how the data is handled in downstream analyses.
Should be one of:
- "count": Raw count data, will be normalized by the function.
- "proportion": Data that has already been normalized to proportions/percentages.
- "other": Custom abundance data that has unknown scaling. No normalization applied.
The choice affects preprocessing steps as well as plot axis labels.
Default is "count", which assumes raw OTU table input.}

\item{feature.sig.level}{A numeric threshold, usually between 0 and 1, for assessing the significance of individual taxa. Default is 0.1.}

\item{feature.mt.method}{A character string specifying the method employed for multiple testing correction (e.g., "fdr" for False Discovery Rate). Default is "fdr".}

\item{...}{Additional parameters to be passed to the linda function.}
}
\value{
A named list containing data frames summarizing taxon test results for each taxonomic level.
}
\description{
This function takes a MicrobiomeStat data object as input, filters taxa based on
prevalence and abundance thresholds, aggregates taxon abundances by sample,
fits linear mixed effects models using the linda method to identify significant
taxon changes across groups over time, accounting for specified covariates.
}
\details{
Each list element corresponds to a taxonomic level specified in `feature.level`.
The data frame contains columns for taxon name, log2 fold change, p-values, adjusted p-values,
mean abundance, mean prevalence, and the output element from `linda` where the taxon was found significant.
}
\examples{
\dontrun{
data(peerj32.obj)
generate_taxa_test_pair(
  data.obj = peerj32.obj,
  subject.var = "subject",
  time.var = "time",
  group.var = "group",
  adj.vars = c("sex"),
  feature.level = c("Genus"),
  prev.filter = 0.1,
  abund.filter = 0.0001,
  feature.dat.type = "count",
  feature.sig.level = 0.1,
  feature.mt.method = "none"
)
}

}
