% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxa_trend_test_long.R
\name{generate_taxa_trend_volcano_long}
\alias{generate_taxa_trend_volcano_long}
\title{Generate volcano plots for longitudinal taxa trend test}
\usage{
generate_taxa_trend_volcano_long(
  data.obj,
  group.var = NULL,
  time.var,
  test.list,
  feature.sig.level = 0.1,
  feature.mt.method = "fdr"
)
}
\arguments{
\item{data.obj}{A MicrobiomeStat data object}

\item{group.var}{The grouping variable tested, found in metadata}

\item{time.var}{The time variable used in the analysis}

\item{test.list}{The list of test results returned by generate_taxa_trend_test_long}

\item{feature.sig.level}{The significance level cutoff for highlighting taxa}

\item{feature.mt.method}{Multiple testing correction method, "fdr" or "none"}
}
\value{
A list of ggplot objects of volcano plots for each taxonomic level
}
\description{
Generate volcano plots for longitudinal taxa trend test
}
\examples{
# data("subset_T2D.obj")
# test_list <- generate_taxa_trend_test_long(data.obj, ...)
# volcano_plots <- generate_taxa_trend_volcano_long(data.obj,
                                                 # group.var,
                                                 # time.var,
                                                 # test_list,
                                                 # feature.sig.level = 0.05,
                                                 # feature.mt.method = "fdr")

}
