% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_alpha_test_pair.R
\name{is_rarefied}
\alias{is_rarefied}
\title{Check if Data has been Rarefied}
\usage{
is_rarefied(data.obj)
}
\arguments{
\item{data.obj}{A MicrobiomeStat data object, which is a list containing
at minimum the following components:
\itemize{
  \item \code{feature.tab}: A matrix of feature abundances (taxa/genes as
    rows, samples as columns)
  \item \code{meta.dat}: A data frame of sample metadata (samples as rows)
}
Optional components include:
\itemize{
  \item \code{feature.ann}: A matrix/data frame of feature annotations
    (e.g., taxonomy)
  \item \code{tree}: A phylogenetic tree object (class "phylo")
  \item \code{feature.agg.list}: Pre-aggregated feature tables by taxonomy
}
Data objects can be created using converters like
\code{\link{mStat_convert_phyloseq_to_data_obj}} or importers like
\code{\link{mStat_import_qiime2_as_data_obj}}.}
}
\value{
Logical. TRUE if data is rarefied (all samples have equal counts),
  FALSE otherwise.
}
\description{
Checks if the data has been rarefied by inspecting if all samples have
equal total counts.
}
\examples{
# Assuming peerj32.obj is a data object with OTU and metadata tables
data(peerj32.obj)
is_rarefied(peerj32.obj)

}
