% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linda.R
\name{linda}
\alias{linda}
\title{Linear (Lin) Model for Differential Abundance (DA) Analysis of High-dimensional Compositional Data}
\description{
The function implements a simple, robust and highly scalable approach to tackle
the compositional effects in differential abundance analysis of high-dimensional compositional data. 
It fits linear regression models on the centered log2-ratio transformed data, identifies a bias term due to the transformation
and compositional effect, and corrects the bias using the mode of the regression coefficients.
It could fit mixed-effect models for analysis of correlated data.
}
\usage{
linda(
  feature.dat,
  meta.dat,
  formula,
  feature.dat.type = c('count', 'proportion'),
  prev.filter = 0,
  mean.abund.filter = 0, 
  max.abund.filter = 0,
  is.winsor = TRUE,
  outlier.pct = 0.03,
  adaptive = TRUE,
  zero.handling = c('pseudo-count', 'imputation'),
  pseudo.cnt = 0.5,
  corr.cut = 0.1,
  p.adj.method = "BH",
  alpha = 0.05,
  n.cores = 1, 
  verbose = TRUE
)
}
\arguments{

\item{feature.dat}{a matrix of counts/proportions, row - features (OTUs, genes, etc) , column - samples.}

\item{meta.dat}{a data frame containing the sample meta data. If there are NAs, the corresponding samples
 will be removed in the analysis.}

\item{formula}{a character string for the formula. The formula should conform to that used by \code{lm} (independent 
data) or \code{lmer} (correlated data).
 For example: \code{formula = '~x1*x2+x3+(1|id)'}. At least one fixed effect is required.}

\item{feature.dat.type}{the type of the feature data. It could be "count" or "proportion".}

\item{prev.filter}{the prevalence (percentage of non-zeros) cutoff, under which the features will  be filtered. The default is 0. }

\item{mean.abund.filter}{the mean relative abundance cutoff, under which the features will  be filtered. The default is 0.}

\item{max.abund.filter}{the max relative abundance cutoff, under which the features will  be filtered. The default is 0.}

\item{is.winsor}{a logical value indicating whether winsorization should be performed to replace outliers (high values).
 The default is TRUE.}

\item{outlier.pct}{the expected percentage of outliers. These outliers will be winsorized. The default is 0.03.}


\item{adaptive}{a logical value indicating whether the approach to handle zeros (pseudo-count or imputation)
will be determined based on the correlations between the log(sequencing depth) and the explanatory variables
in \code{formula} when \code{feature.dat} is 'count'. If TRUE and the correlation p-value for any explanatory variable 
 is smaller than or equal to \code{corr.cut}, the imputation approach will be used; otherwise, the pseudo-count approach will be used. }


\item{zero.handling}{a character string of 'pseudo-count' or 'imputation' indicating the zero handling method
used when \code{feature.dat} is 'count'.  If 'pseudo-count', a\code{pseudo.cnt} will be added to each value in \code{feature.dat}. 
If 'imputation', then we use the imputation approach using the formula in the referenced paper. Basically,
zeros are imputed with values proportional to the sequencing depth. When \code{feature.dat} is 'proportion',
this parameter will be ignored and zeros will be imputed by half of the minimum for each feature.}

\item{pseudo.cnt}{a positive numeric value for the pseudo-count to be added if \code{zero.handling}
is 'pseudo-count'. Default is 0.5. }

\item{corr.cut}{a numerical value between 0 and 1, indicating the significance level used for determining
the zero-handling approach when \code{adaptive} is TRUE. Default is 0.1.}

\item{p.adj.method}{a character string indicating the p-value adjustment approach for 
addressing multiple testing. See R function \code{p.adjust}. Default is 'BH'.}

\item{alpha}{a numerical value between 0 and 1 indicating the significance level 
for declaring differential features. Default is 0.05.}


\item{n.cores}{a positive integer. If \code{n.cores > 1} and formula is in a form of mixed-effect model,
\code{n.cores} parallels will be conducted. Default is 1.}

\item{verbose}{a logical value indicating whether the trace information should be printed out.}

}
\value{
A list with the elements
\item{variables}{A vector of variable names of all fixed effects in \code{formula}. For example: \code{formula = '~x1*x2+x3+(1|id)'}.
Suppose \code{x1} and \code{x2} are numerical, and \code{x3} is a categorical variable of three levels: a, b and c.
Then the elements of \code{variables} would be \code{('x1', 'x2', 'x3b', 'x3c', 'x1:x2')}.}
\item{bias}{numeric vector; each element corresponds to one variable in \code{variables};
the estimated bias of the regression coefficients due to the compositional effect.}
\item{output}{a list of data frames with columns 'baseMean', 'log2FoldChange', 'lfcSE', 'stat', 'pvalue', 'padj', 'reject',
 'df'; \code{names(output)} is equal to \code{variables}; the rows of the data frame corresponds to taxa.
 Note: if there are taxa being excluded due to \code{prev.cut}, the number of the rows of the output data frame
 will be not equal to the number of the rows of \code{otu.tab}. Taxa are identified by the rownames.
 If the rownames of \code{otu.tab} are NULL, then \code{1 : nrow(otu.tab)} is set as the rownames of \code{otu.tab}.
\describe{
 \item{baseMean:}{ 2 to the power of the intercept coefficients (normalized by one million)}
 \item{log2FoldChange:}{ bias-corrected coefficients}
 \item{lfcSE:}{ standard errors of the coefficients}
 \item{stat:}{ \code{log2FoldChange / lfcSE}}
 \item{pvalue:}{ \code{2 * pt(-abs(stat), df)}}
 \item{padj:}{ \code{p.adjust(pvalue, method = p.adj.method)}}
 \item{reject:}{ \code{padj <= alpha}}
 \item{df:}{ degrees of freedom. The number of samples minus the number of explanatory variables (intercept included) for
 fixed-effect models; estimates from R package \code{lmerTest} with Satterthwaite method of approximation for mixed-effect models.}
 }
 }
 \item{covariance}{a list of data frames; the data frame records the covariances between a regression coefficient with other coefficients;
 \code{names(covariance)} is equal to \code{variables}; the rows of the data frame corresponds to taxa. If the length of \code{variables}
 is equal to 1, then the \code{covariance} is NULL.}
\item{otu.tab.use}{the OTU table used in the abundance analysis (the \code{otu.tab} after the preprocessing:
samples that have NAs in the variables in \code{formula} or have less than \code{lib.cut} read counts are removed;
taxa with prevalence less than \code{prev.cut} are removed and data is winsorized if \code{!is.null(winsor.quan)};
and zeros are treated, i.e., imputed or pseudo-count added).}
\item{meta.use}{the meta data used in the abundance analysis (only variables in \code{formula} are stored; samples that have NAs
or have less than \code{lib.cut} read counts are removed; numerical variables are scaled).}
\item{wald}{a list for use in Wald test. If the fitting model is a linear model, then it includes
\describe{
\item{beta:}{ a matrix of the biased regression coefficients including intercept and all fixed effects; the culumns correspond to taxa}
\item{sig:}{ the standard errors; the elements corresponding to taxa}
\item{X:}{ the design matrix of the fitting model}
\item{bias:}{ the estimated biases of the regression coefficients including intercept and all fixed effects}
}
If the fitting model is a linear mixed-effect model, then it includes
\describe{
\item{beta:}{ a matrix of the biased regression coefficients including intercept and all fixed effects; the culumns correspond to taxa}
\item{beta.cov:}{ a list of covariance matrices of \code{beta}; the elements corresponding to taxa}
\item{rand.cov:}{ a list with covariance matrices of variance parameters of random effects; the elements corresponding to taxa; see more details in the paper of 'lmerTest'}
\item{Joc.beta.cov.rand:}{ a list of a list of Jacobian matrices of \code{beta.cov} with respect to the variance parameters; the elements corresponding to taxa}
\item{bias:}{ the estimated biases of the regression coefficients including intercept and all fixed effects}
}
}
}
\examples{

data(smokers)

ind <- smokers$meta$AIRWAYSITE == 'Throat'
otu.tab <- as.data.frame(smokers$otu[, ind])
depth <- colSums(otu.tab)
meta <- cbind.data.frame(Smoke = factor(smokers$meta$SMOKER[ind]),
                         Sex = factor(smokers$meta$SEX[ind]),
                         Site = factor(smokers$meta$SIDEOFBODY[ind]),
                         SubjectID = factor(smokers$meta$HOST_SUBJECT_ID[ind]))

# Differential abundance analysis using the left throat data                       
ind1 <- meta$Site == 'Left' & depth >= 1000
linda.obj  <- linda(otu.tab[, ind1], meta[ind1, ], formula = '~Smoke+Sex', 
           feature.dat.type = 'count', 
           prev.filter = 0.1, is.winsor = TRUE, outlier.pct = 0.03,
           p.adj.method = "BH", alpha = 0.1)
           
           


linda.plot(linda.obj, c('Smokey', 'Sexmale'),
           titles = c('Smoke: n v.s. y', 'Sex: female v.s. male'), 
           alpha = 0.1, lfc.cut = 1, legend = TRUE, directory = NULL,
            width = 11, height = 8)
            
rownames(linda.obj $output[[1]])[which(linda.obj $output[[1]]$reject)]


# Differential abundance analysis pooling both the left and right throat data 
# Mixed effects model is used 
\donttest{
ind  <- depth >= 1000
linda.obj <- linda(otu.tab[, ind], meta[ind, ], formula = '~Smoke+Sex+(1|SubjectID)',
           feature.dat.type = 'count', 
           prev.filter = 0.1, is.winsor = TRUE, outlier.pct = 0.03,
           p.adj.method = "BH", alpha = 0.1)

}       
    
# For proportion data   
otu.tab.p <- t(t(otu.tab) / colSums(otu.tab))
ind1 <- meta$Site == 'Left' & depth >= 1000
lind.obj  <- linda(otu.tab[, ind1], meta[ind1, ], formula = '~Smoke+Sex', 
           feature.dat.type = 'proportion', 
           prev.filter = 0.1, is.winsor = TRUE, outlier.pct = 0.03,
           p.adj.method = "BH", alpha = 0.1)

}

\references{
Zhou, H., He, K., Chen, J., Zhang, X. (2022). LinDA: linear models for differential abundance analysis of microbiome compositional data. Genome biology, 23(1), 95.
}
\author{
Huijuan Zhou,
Jun Chen,
Xianyang Zhang

}
