% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_calculate_adjusted_distance.R
\name{mStat_calculate_adjusted_distance}
\alias{mStat_calculate_adjusted_distance}
\title{Calculate Adjusted Distances}
\usage{
mStat_calculate_adjusted_distance(data.obj, dist.obj, adj.vars, dist.name)
}
\arguments{
\item{data.obj}{A MicrobiomeStat data object. The heart of the MicrobiomeStat, consisting of several key components:
\itemize{
  \item \strong{Feature.tab (Matrix)}: Meeting point for research objects (OTU/ASV/KEGG/Gene, etc.) and samples.
  \item \strong{Meta.dat (Data frame)}: Rows correspond to the samples, and columns serve as annotations, describing the samples.
  \item \strong{Feature.ann (Matrix)}: Annotations, carrying classification information like Kingdom, Phylum, etc.
  \item \strong{Phylogenetic tree (Optional)}: An evolutionary perspective, illuminating the relationships among various research objects.
  \item \strong{Feature.agg.list (Optional)}: Aggregated results based on the feature.tab and feature.ann.
}}

\item{dist.obj}{A list of distance matrices corresponding to the names in `dist.name`.}

\item{adj.vars}{A character vector of variable names to be used for adjustment.}

\item{dist.name}{A character vector of names corresponding to distance matrices in `dist.obj`.}
}
\value{
A list of adjusted distance matrices.
}
\description{
This function calculates the adjusted distances for a given set of variables,
using a model matrix dynamically created from the specified adjustment variables.
}
\examples{
\dontrun{
data("subset_T2D.obj")
dist.obj <- mStat_calculate_beta_diversity(subset_T2D.obj, c("BC","Jaccard"))
adj.dist.obj <- mStat_calculate_adjusted_distance(
data.obj = subset_T2D.obj,
dist.obj = dist.obj,
adj.vars = c("subject_gender", "subject_race"),
dist.name = c("BC", "Jaccard"))
}
}
