% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_calculate_beta_diversity.R
\name{mStat_calculate_beta_diversity}
\alias{mStat_calculate_beta_diversity}
\title{Calculate various beta diversity indices}
\usage{
mStat_calculate_beta_diversity(
  data.obj,
  dist.name = c("BC", "Jaccard", "UniFrac", "GUniFrac", "WUniFrac", "JS")
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{dist.name}{A character vector specifying which beta diversity indices to calculate. Supported indices are "BC" (Bray-Curtis), "Jaccard", "UniFrac" (unweighted UniFrac), "GUniFrac" (generalized UniFrac), "WUniFrac" (weighted UniFrac), and "JS" (Jensen-Shannon divergence). If a name is provided but the corresponding object does not exist within dist.obj, it will be computed internally. If the specific index is not supported, an error message will be returned.
"BC" (Bray-Curtis), "Jaccard", "UniFrac" (unweighted UniFrac), "GUniFrac" (generalized UniFrac),
"WUniFrac" (weighted UniFrac), and "JS" (Jensen-Shannon divergence).}
}
\value{
A list containing the calculated beta diversity indices. The indices are named with the abbreviation.
}
\description{
This function calculates a variety of beta diversity indices based on the input data, such as
Bray-Curtis (BC), Jaccard, unweighted UniFrac (UniFrac), generalized UniFrac (GUniFrac),
weighted UniFrac (WUniFrac), and Jensen-Shannon divergence (JS).
}
\examples{
\dontrun{
library(vegan)
library(GUniFrac)
library(ape)
library(philentropy)
library(phyloseq)

# Load example data
data(GlobalPatterns)

GlobalPatterns.obj <- mStat_convert_phyloseq_to_data_obj(GlobalPatterns)

# Calculate various beta diversity indices
dist.obj <- mStat_calculate_beta_diversity(GlobalPatterns.obj, dist.name = c('BC'))
dist.obj <- mStat_calculate_beta_diversity(GlobalPatterns.obj, dist.name = c("Jaccard"))
dist.obj <- mStat_calculate_beta_diversity(GlobalPatterns.obj, dist.name = c('UniFrac'))
dist.obj <- mStat_calculate_beta_diversity(GlobalPatterns.obj, dist.name = c('WUniFrac'))
dist.obj <- mStat_calculate_beta_diversity(GlobalPatterns.obj, dist.name = c('JS'))
}

}
