% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_combine_data.R
\name{mStat_combine_data}
\alias{mStat_combine_data}
\title{mStat_combine_data}
\usage{
mStat_combine_data(data.obj1, data.obj2)
}
\arguments{
\item{data.obj1}{The first MicrobiomeStat data object to be combined.}

\item{data.obj2}{The second MicrobiomeStat data object to be combined.}
}
\value{
A list with the combined 'feature.tab', 'meta.dat' and 'feature.ann'.
}
\description{
This function is designed specifically for combining two MicrobiomeStat data objects,
assuming that they are both in 'Raw' format. Each MicrobiomeStat data object should be a list that contains
'norm.status' (indicating the status of normalization, should be 'Raw'), 'feature.tab' (a matrix of features),
'meta.dat' (a data frame of metadata), and 'feature.ann' (a matrix of feature annotations).
}
\examples{
# combined_data <- mStat_combine_data(data_obj1, data_obj2)

}
