% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_generate_report_long.R
\name{mStat_generate_report_long}
\alias{mStat_generate_report_long}
\title{Generate a Comprehensive Report for Microbial Ecology Analysis of Longitudinal Data}
\usage{
mStat_generate_report_long(
  data.obj,
  alpha.obj = NULL,
  dist.obj = NULL,
  pc.obj = NULL,
  group.var,
  strata.var = NULL,
  adj.vars = NULL,
  subject.var,
  time.var,
  alpha.name = c("shannon", "simpson"),
  dist.name = c("BC", "Jaccard"),
  t0.level,
  ts.levels,
  change.func = "relative difference",
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  prev.filter = 0,
  abund.filter = 0,
  feature.number = 15,
  feature.level = NULL,
  feature.dat.type = c("count", "proportion", "other"),
  feature.mt.method = c("fdr", "none"),
  feature.sig.level = 0.1,
  transform = c("identity", "sqrt", "log"),
  output.file,
  ...
)
}
\arguments{
\item{data.obj}{A data object created by mStat_convert_phyloseq_to_data_obj.}

\item{alpha.obj}{An alpha diversity object (optional).}

\item{dist.obj}{A distance object created by mStat_calculate_beta_diversity.}

\item{pc.obj}{Principal coordinate analysis result for beta diversity, generated by mStat_calculate_PC.}

\item{group.var}{Variable name used for grouping samples.}

\item{strata.var}{Variable to stratify the analysis by (optional).}

\item{adj.vars}{Variables to adjust for in the analysis.}

\item{subject.var}{Variable name used for subject identification.}

\item{time.var}{Variable name used for time points in longitudinal data.}

\item{alpha.name}{Names of alpha diversity indices to include in the analysis.}

\item{dist.name}{Names of beta diversity distance metrics to include in the analysis.}

\item{t0.level}{The base level for time points in longitudinal data.}

\item{ts.levels}{The levels for time points in longitudinal data.}

\item{change.func}{A function to calculate change between time points, e.g. log2fold change. Default is NULL.}

\item{base.size}{Base font size for the generated plots.}

\item{theme.choice}{Plot theme choice (default: "prism").}

\item{custom.theme}{Custom ggplot2 theme (optional).}

\item{palette}{Color palette used for the plots.}

\item{pdf}{Logical indicating whether to save plots as PDF files (default: TRUE).}

\item{file.ann}{Annotation text for the PDF file names.}

\item{pdf.wid}{Width of the PDF plots.}

\item{pdf.hei}{Height of the PDF plots.}

\item{prev.filter}{Prevalence filter for feature analysis.}

\item{abund.filter}{Abundance filter for feature analysis.}

\item{feature.number}{The number of features to plot. Default is 15.}

\item{feature.level}{Taxonomic level for feature analysis.}

\item{feature.dat.type}{Data type for feature analysis (count, proportion, or other).}

\item{feature.mt.method}{Multiple testing correction method for taxonomic features, "fdr" or "none".}

\item{feature.sig.level}{Significance level cutoff for highlighting significant taxonomic features.}

\item{transform}{The transformation function applied to the data (default: "log").}

\item{output.file}{Output file name for the report.}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
A PDF report file containing the microbial ecology analysis results for longitudinal data.
}
\description{
This function generates a comprehensive report for microbial ecology analysis,
             including changes in alpha diversity, beta diversity, and taxonomic features in longitudinal data.
}
\examples{
\dontrun{
# Assuming peerj32.obj, dist.obj, alpha.obj are pre-defined objects
data(peerj32.obj)
mStat_generate_report_long(
  data.obj = peerj32.obj,
  dist.obj = NULL,
  alpha.obj = NULL,
  group.var = "group",
  adj.vars = c("sex"),
  subject.var = "subject",
  time.var = "time",
  alpha.name = c("shannon","simpson"),
  dist.name = c("BC",'Jaccard'),
  t0.level = "1",
  ts.levels = "2",
  strata.var = "sex",
  feature.level = c("Phylum"),
  feature.dat.type = "count",
  transform = "log",
  theme.choice = "bw",
  base.size = 12,
  output.file = "path/report.pdf"
)
data(subset_T2D.obj)
mStat_generate_report_long(
  data.obj = subset_T2D.obj,
  dist.obj = NULL,
  alpha.obj = NULL,
  pc.obj = NULL,
  group.var = "subject_gender",
  strata.var = "subject_race",
  adj.vars = NULL,
  subject.var = "subject_id",
  time.var = "visit_number_num",
  alpha.name = c("shannon","simpson"),
  dist.name = c("BC",'Jaccard'),
  t0.level = unique(sort(subset_T2D.obj$meta.dat$visit_number_num))[1],
  ts.levels = unique(sort(subset_T2D.obj$meta.dat$visit_number_num))[-1],
  feature.mt.method = "none",
  feature.sig.level = 0.1,
  feature.level = c("Phylum"),
  change.func = "relative difference",
  feature.dat.type = "count",
  prev.filter = 1e-5,
  abund.filter = 1e-5,
  transform = "sqrt",
  theme.choice = "bw",
  base.size = 12,
  output.file = "/Users/apple/Microbiome/Longitudinal/MicrobiomeStat_Paper/报告/mStat_generate_report_long_example.pdf"
)
}
data(subset_T2D.obj)
mStat_generate_report_long(
  data.obj = subset_T2D.obj,
  dist.obj = NULL,
  alpha.obj = NULL,
  pc.obj = NULL,
  group.var = "subject_gender",
  strata.var = "subject_race",
  adj.vars = NULL,
  subject.var = "subject_id",
  time.var = "visit_number_num",
  alpha.name = c("shannon","simpson"),
  dist.name = c("BC",'Jaccard'),
  t0.level = unique(sort(subset_T2D.obj$meta.dat$visit_number_num))[1],
  ts.levels = unique(sort(subset_T2D.obj$meta.dat$visit_number_num))[-1],
  feature.mt.method = "none",
  feature.sig.level = 0.1,
  feature.level = c("Phylum"),
  change.func = "relative difference",
  feature.dat.type = "count",
  prev.filter = 1e-5,
  abund.filter = 1e-5,
  transform = "sqrt",
  theme.choice = "bw",
  base.size = 12,
  output.file = "/Users/apple/Microbiome/Longitudinal/MicrobiomeStat_Paper/报告/mStat_generate_report_long_example.pdf"
)
}
