% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_generate_report_single.R
\name{mStat_generate_report_single}
\alias{mStat_generate_report_single}
\title{Generate a report for microbial ecology analysis of cross-sectional, longitudinal or paired data at a single time point}
\usage{
mStat_generate_report_single(
  data.obj,
  dist.obj = NULL,
  alpha.obj = NULL,
  depth = NULL,
  group.var,
  adj.vars,
  subject.var,
  time.var,
  alpha.name = c("shannon", "simpson", "observed_species", "chao1", "ace", "pielou"),
  dist.name = c("BC", "Jaccard", "UniFrac", "GUniFrac", "WUniFrac", "JS"),
  t.level = NULL,
  strata.var = NULL,
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  prev.filter = 0,
  abund.filter = 0,
  feature.level = NULL,
  feature.dat.type = c("count", "proportion", "other"),
  output.file,
  ...
)
}
\arguments{
\item{data.obj}{A data object created by mStat_convert_phyloseq_to_data_obj.}

\item{dist.obj}{A distance object created by mStat_calculate_beta_diversity.}

\item{alpha.obj}{An alpha diversity object (optional).}

\item{depth}{Sampling depth (optional).}

\item{group.var}{Variable name used for grouping samples.}

\item{adj.vars}{Variables to adjust for in the longitudinal analysis.}

\item{subject.var}{Variable name used for subject identification.}

\item{time.var}{Variable name used for time points.}

\item{alpha.name}{Names of alpha diversity indices to include in the analysis.}

\item{dist.name}{Names of beta diversity distance metrics to include in the analysis.}

\item{t.level}{Time point(s) to include in the analysis (optional).}

\item{strata.var}{Variable to stratify the analysis by (optional).}

\item{base.size}{Base font size for the generated plots.}

\item{theme.choice}{Plot theme choice (default: "prism").}

\item{custom.theme}{Custom ggplot2 theme (optional).}

\item{palette}{Color palette used for the plots.}

\item{pdf}{Logical indicating whether to save plots as PDF files (default: TRUE).}

\item{file.ann}{Annotation text for the PDF file names.}

\item{pdf.wid}{Width of the PDF plots.}

\item{pdf.hei}{Height of the PDF plots.}

\item{prev.filter}{Prevalence filter for feature analysis.}

\item{abund.filter}{Abundance filter for feature analysis.}

\item{feature.level}{Taxonomic level for feature analysis.}

\item{feature.dat.type}{Data type for feature analysis (count, proportion, or other).}

\item{output.file}{Output file name for the report.}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
A report file containing the microbial ecology analysis results.
}
\description{
This function generates a comprehensive report for microbial ecology analysis,
including alpha diversity, beta diversity, and taxonomic feature analyses. The function is designed
to perform analysis on cross-sectional data, a single time point from longitudinal or paired data.
}
\examples{
library(GUniFrac)
library(pheatmap)
library(vegan)
library(ggh4x)


data(peerj32.obj)

dist.obj <- mStat_calculate_beta_diversity(peerj32.obj, dist.name = c('BC', 'Jaccard'))

# Generate a report for microbial ecology analysis
mStat_generate_report_single(
  data.obj = peerj32.obj,
  dist.obj = NULL,
  group.var = "group",
  adj.vars = c("sex"),
  subject.var = "subject",
  time.var = "time",
  alpha.name = c("shannon","simpson"),
  dist.name = c("BC",'Jaccard'),
  t.level = "1",
  strata.var = "sex",
  feature.level = c("Phylum","Family"),
  feature.dat.type = "count",
  theme.choice = "bw",
  base.size = 12,
  output.file = "/Users/apple/Microbiome/Longitudinal/
  MicrobiomeStat/mStat_generate_report_single_example.pdf"
)
}
