% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_normalize_data.R
\name{mStat_normalize_data}
\alias{mStat_normalize_data}
\title{Normalize a MicrobiomeStat Data Object}
\usage{
mStat_normalize_data(
  data.obj,
  method = c("Rarefy-TSS", "Rarefy", "TSS", "GMPR", "CSS", "DESeq", "TMM"),
  depth = NULL
)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.}

\item{method}{A string. The normalization method to be applied. It must be one of the following: "Rarefy-TSS", "Rarefy", "TSS", "GMPR", "CSS", "DESeq", "TMM". The default is "Rarefy-TSS".
- "Rarefy-TSS": Rarefaction followed by Total Sum Scaling normalization.
- "Rarefy": Rarefaction normalization only.
- "TSS": Total Sum Scaling normalization only.
- "GMPR": Geometric Mean of Pairwise Ratios normalization method.
- "CSS": Cumulative Sum Scaling normalization method.
- "DESeq": Normalization using the DESeq method for RNA-seq data.
- "TMM": Normalization using the Trimmed Mean of M-values (TMM) method from the edgeR package.}

\item{depth}{An integer. The sequencing depth to be used for the "Rarefy" and "Rarefy-TSS" methods. If NULL, the smallest total count across samples is used as the rarefaction depth.}
}
\value{
A list. The normalized data object and the scale factor used for normalization.
}
\description{
This function is part of the MicrobiomeStat package. It normalizes a data object based on the chosen method.
}
\details{
The function first checks if 'data.obj' is a list. It then retrieves the OTU table and estimates the normalization/scale factor based on the chosen method. The data object is then updated with the normalized OTU table and the chosen method is added as 'norm.status'. The function returns the normalized data object and the scale factor.
}
\examples{
\dontrun{
library(vegan)
peerj32.obj <- list()
peerj32.phy <- peerj32$phyloseq
peerj32.obj <- mStat_convert_phyloseq_to_data_obj(peerj32.phy)
peerj32.obj <- mStat_aggregate_by_taxonomy(peerj32.obj, feature.level = c("Phylum","Family"))
}

}
