% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_process_time_variable.R
\name{mStat_process_time_variable}
\alias{mStat_process_time_variable}
\title{Handle Time Variable in Meta Data Table}
\usage{
mStat_process_time_variable(
  data.obj,
  time.var,
  t0.level = NULL,
  ts.levels = NULL
)
}
\arguments{
\item{time.var}{A character string that specifies the name of the time variable in the meta data table.}

\item{t0.level}{A single value that specifies the level to be used as the reference level (optional).}

\item{ts.levels}{A vector of values that specifies the levels to be used as the time series levels (optional).}

\item{meta.dat}{A data frame that contains the meta data.}
}
\value{
A data frame that is the modified version of the input meta data table.
}
\description{
This function handles a time variable in a meta data table. It sets the levels of the time variable based on provided parameters and subsets the data if necessary.
}
\details{
The function first gets the unique values of the time variable. Then, it sets the levels of the time variable based on the provided t0.level and ts.levels. If these parameters are not provided, the levels are set to the unique values of the time variable. The function then modifies the time variable in the meta data table based on its type (factor, numeric, or character). If the time variable is a character, it is converted to numeric. If there are any NA values generated during this conversion, a message is displayed. Finally, if the levels set do not include all the unique values of the time variable, the function subsets the data to exclude the missing levels and displays a message.
}
\examples{
# Assuming 'meta_data' is your meta data table and 'time' is your time variable
# meta_data_handled <- mStat_process_time_variable(meta_data, 'time')

mStat_process_time_variable(data.obj = ecam.obj,time.var = "month",
t0.level = "0",ts.levels = c("1","2"))

}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[rlang]{sym}}
}
\author{
Your Name
}
