% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_subset_data.R
\name{mStat_subset_data}
\alias{mStat_subset_data}
\title{Subset Data Object by Sample IDs in MicrobiomeStat}
\usage{
mStat_subset_data(data.obj, samIDs = NULL, condition = NULL)
}
\arguments{
\item{data.obj}{A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.
table (`feature.tab`), feature names, a full feature name list, and a feature aggregation list
(`feature.agg.list`).}

\item{samIDs}{A vector of sample IDs to subset by. This can be a logical vector, a numeric
vector, or a character vector of sample IDs.}

\item{condition}{A logical expression string describing rows to keep. This allows subsetting samples based on metadata, similarly to dplyr::filter. Only one of 'samIDs' and 'condition' should be provided.}
}
\value{
A list that is the modified version of the input data object. It includes the subsetted
metadata, feature table, feature names, full feature name list, and feature aggregation list.
}
\description{
This function is part of the MicrobiomeStat package. It subsets a multi-omics data object by
a specified set of sample IDs. It checks for the presence of various data components in the
data object and subsets them if they exist.
}
\details{
The function first checks if samIDs is logical or numeric, and if so, converts it to a
character vector of sample IDs. It then subsets the metadata by the sample IDs. If a feature
table exists, it subsets the feature table and the feature names by the sample IDs. If a full
feature name list exists, it subsets the full feature name list by the sample IDs. If a feature
aggregation list exists, it subsets each feature aggregation table in the list by the sample IDs.
The function returns the subsetted data object.
}
\examples{
\dontrun{
# Load the required libraries
library(microbiome)
library(MicrobiomeStat)

# Prepare data for the function
data(peerj32)
peerj32.obj <- list()
peerj32.phy <- peerj32$phyloseq
peerj32.obj <- mStat_convert_phyloseq_to_data_obj(peerj32.phy)

# Subset data object by sample IDs
subset.ids <- rownames(peerj32.obj$meta.dat)[1:10] # for example
subset_peerj32.obj <- mStat_subset_data(data.obj = peerj32.obj, samIDs = subset.ids)

# Subset data object by condition
condition <- "time == '1'" # for example
subset_peerj32.obj2 <- mStat_subset_data(data.obj = peerj32.obj, condition = condition)
}

}
\seealso{
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}
}
