% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStat_summarize_data_obj.R
\name{mStat_summarize_data_obj}
\alias{mStat_summarize_data_obj}
\title{Summarize a MicrobiomeStat Data Object}
\usage{
mStat_summarize_data_obj(
  data.obj,
  time.var = NULL,
  group.var = NULL,
  palette = NULL
)
}
\arguments{
\item{data.obj}{A MicrobiomeStat data object to be summarized.}

\item{time.var}{An optional time variable column name from meta.dat. If provided, the function will additionally dplyr::summarize the temporal distribution of samples. Default is NULL.}

\item{group.var}{An optional grouping variable column name from meta.dat. If provided together with time.var, the histogram of sample counts over time will be grouped by this variable. Default is NULL.}

\item{palette}{A vector of colors to use for grouping in the histogram. Default is a preset color palette with 10 colors.}
}
\value{
A list containing summaries for:
\itemize{
  \item feature.tab: The overall summary includes the number of features, number of samples, sparsity, and description of singleton features.
  \item meta.dat: The overall summary includes the number of samples, number of metadata fields, missing data, and sample distribution if a time variable is provided. If time.var is provided, a distribution of samples over time is also generated.
  \item feature.ann: The overall summary includes number of features, number of annotations, and the proportion of NA values for each annotation.
  \item tree: (optional) Whether a phylogenetic tree exists in the data object.
}
}
\description{
This function takes a MicrobiomeStat data object and provides a summary of the various components such as feature.tab, meta.dat, feature.ann, and optionally time.var and tree.
The function is useful to give an overview of the input data object before any further microbiome statistical analysis. Providing a time variable can offer additional insights into the temporal dynamics of your samples.
}
\details{
The function first checks if each component of the MicrobiomeStat data object is not null. If a component is not null, it is summarized and added to the output list. For the feature.tab, it computes the sparsity and singleton features. For the meta.dat, it computes the number of samples and metadata fields, and the distribution of samples if a time variable is provided. The inclusion of a time variable allows the user to gain insights into how samples are distributed over time. For the feature.ann, it computes the number of features, annotations, and the proportion of NA values for each annotation. It also checks if a phylogenetic tree exists in the data object.
}
\examples{
\dontrun{
  # Assuming 'data.obj' is your MicrobiomeStat data object
  # Summary with time variable
  # summary_list <- mStat_summarize_data_obj(data.obj, time.var = "time")

  # Summary without time variable
  # summary_list <- mStat_summarize_data_obj(data.obj)

  # If you have a microbiome data available as a MicrobiomeStat data object
  # you can dplyr::summarize it using:
  # library(MicrobiomeStat)
  # data(data.obj)
  # Summary with time variable
  # summary_list <- mStat_summarize_data_obj(data.obj, time.var = "time")

  # Summary without time variable
  # summary_list <- mStat_summarize_data_obj(data.obj)
  data(subset_T2D.obj)
  summary <- mStat_summarize_data_obj(subset_T2D.obj, "visit_number", "subject_race")
}

}
