% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_utils.R
\name{resolve_change_type}
\alias{resolve_change_type}
\title{Resolve Change Type Parameters}
\usage{
resolve_change_type(change.type = "none")
}
\arguments{
\item{change.type}{Change type: "none", "relative", "log_fold", "absolute"}
}
\value{
List with feature.change.func and related parameters
}
\description{
Converts unified change.type to function-specific parameters.
}
\keyword{internal}
