% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smokers.R
\name{smokers}
\alias{smokers}
\title{Smokers Microbiome Data from the Human Upper Respiratory Tract}
\format{
A list containing:
\describe{
  \item{otu}{OTU table, 2156 taxa by 290 samples}
  \item{tax}{Taxonomy table, 2156 taxa by 7 taxonomic ranks}
  \item{meta}{Metadata table, 290 samples by 53 variables}
  \item{genus}{304 genera by 290 samples}
  \item{family}{113 families by 290 samples}
}
}
\source{
Charlson ES, Chen J, Custers-Allen R, Bittinger K, Li H, et al. (2010)
  Disordered Microbial Communities in the Upper Respiratory
  Tract of Cigarette Smokers. PLoS ONE 5(12): e15216. \url{https://qiita.ucsd.edu/study/description/524}
}
\description{
The smokers dataset contains microbiome data from left and right throat
samples of 290 subjects, along with taxonomy and metadata tables.
}
\references{
Charlson ES, et al. PLoS ONE 2010
}
